fix(Company) : edit by code review
