+++ .eslintrc
@@ -1,11 +1,8 @@
 {
-    "extends": ["plugin:prettier/recommended"],
-    "parserOptions": {
-      "ecmaFeatures": {
-        "jsx": true
-      },
-      "ecmaVersion": 2018,
-      "sourceType": "module"
+  "extends": ["plugin:prettier/recommended"],
+  "parserOptions": {
+    "ecmaFeatures": {
+      "jsx": true
     },
     "ecmaVersion": 2018,
     "sourceType": "module"
