+++ src/components/Company/Company.jsx
@@ -1,20 +1,51 @@
-import React, { useState } from 'react';
-import CompanyPresenter from './CompanyPresenter';
-import { companyProfile, companyQuote } from './dummyData';
+import React from 'react';
+import { ReactComponent as TitleIcon } from 'assets/icons/company-title-icon.svg';
+import StockChart from './StockChart';
+import './css/Company.scss';
 
-const getCompanyProfile = () => {
-  return companyProfile[0];
+const CompanyPresenter = ({ profile, quote }) => {
+	return (
+		<div className="container">
+			<div className="wrapper">
+				<div className="left-box">
+					<div className="left-title shadow-box">
+						<div className="icon-box shadow-box">
+							<TitleIcon width="3em" height="3em" fill="white" />
+						</div>
+						<div className="company-name">
+							{profile.companyName} ({profile.symbol})
+						</div>
+						<div>
+							<div className="price">
+								{new Intl.NumberFormat('en-US', {
+									style: 'currency',
+									currency: `${profile.currency}`,
+								}).format(`${quote.price}`)}
+							</div>
+							<div className={quote.change > 0 ? 'price-ratio-up' : 'price-ratio-down'}>
+								{quote.change > 0
+									? `▲ ${quote.change.toFixed(2)} ${quote.changesPercentage.toFixed(2)}%`
+									: `▼ ${quote.change.toFixed(2)} ${quote.changesPercentage.toFixed(2)}%`}
+							</div>
+						</div>
+					</div>
+					<StockChart />
+				</div>
+				<div className="vertical-line" />
+				<div className="summary-box">
+					<div className="summary-logo">
+						<div className="logo-box shadow-box">
+							<img alt={profile.companyName} src={profile.image}></img>
+						</div>
+					</div>
+					<div className="summary-detail shadow-box">
+						<div className="summary-title">회사 개요</div>
+						<div className="summary-content">{profile.description}</div>
+					</div>
+				</div>
+			</div>
+		</div>
+	);
 };
 
-const getCompanyQuote = () => {
-  return companyQuote[0];
-};
-
-const Company = () => {
-  const [profile, setProfile] = useState(getCompanyProfile());
-  const [quote, setQuote] = useState(getCompanyQuote());
-
-  return <CompanyPresenter profile={profile} quote={quote} />;
-};
-
-export default Company;
+export default CompanyPresenter;
