+++ src/components/Company/css/Company.scss
@@ -1,145 +1,139 @@
 .container {
-  height: 100vh;
-  padding: 32px;
-  background-color: #ececec;
+	height: 100vh;
+	padding: 32px;
+	background-color: #ececec;
 }
 
 .wrapper {
-  height: 100%;
-  display: flex;
-  font-size: 2vh;
-  margin: auto 80px;
+	height: 100%;
+	display: flex;
+	font-size: 2vh;
+	margin: auto 80px;
 }
 
 .left-box {
-  width: 60%;
-  height: 100%;
+	width: 60%;
+	height: 100%;
 
-  .left-title {
-    display: grid;
-    grid-template-columns: 0.6fr 2fr 1fr;
-    align-items: center;
-    background-color: white;
-    height: 20%;
-    padding: 8px;
+	.left-title {
+		display: grid;
+		grid-template-columns: 0.6fr 2fr 1fr;
+		align-items: center;
+		background-color: white;
+		height: 20%;
+		padding: 8px;
 
-    .icon-box {
-      margin: auto;
-      background-color: #2e9bff;
-      padding: 1em;
-    }
+		.icon-box {
+			margin: auto;
+			background-color: #2e9bff;
+			padding: 1em;
+		}
 
-    .company-name {
-      font-size: 2.5em;
-      font-weight: bold;
-    }
+		.company-name {
+			font-size: 2.5em;
+			font-weight: bold;
+		}
 
-    .price {
-      font-size: 2.5em;
-      font-weight: bold;
-    }
+		.price {
+			font-size: 2.5em;
+			font-weight: bold;
+		}
 
-    .price-ratio-up {
-      color: red;
-      font-size: 1.45em;
-    }
+		.price-ratio-up {
+			color: red;
+			font-size: 1.45em;
+		}
 
-    .price-ratio-down {
-      color: blue;
-      font-size: 1.45em;
-    }
-  }
+		.price-ratio-down {
+			color: blue;
+			font-size: 1.45em;
+		}
+	}
 
-  .graph {
-    height: 80%;
+	.graph {
+		height: 80%;
 
-    ul {
-      width: 100%;
-      height: 15%;
-      display: flex;
-      align-items: center;
-      li {
-        float: left;
+		ul {
+			width: 100%;
+			height: 15%;
+			display: flex;
+			align-items: center;
+			li {
+				float: left;
 
-        input {
-          display: none;
-        }
+				input {
+					display: none;
+				}
 
-        label {
-          background-color: white;
-          font-weight: bold;
-          padding: 8px;
-          margin: 8px;
-          cursor: pointer;
-        }
+				label {
+					background-color: white;
+					font-weight: bold;
+					padding: 8px;
+					margin: 8px;
+					cursor: pointer;
+				}
 
-        input:checked + label {
-          background-color: #60b2ff;
-          color: white;
-        }
-      }
-    }
+				input:checked + label {
+					background-color: #60b2ff;
+					color: white;
+				}
+			}
+		}
 
-    .graph-chart {
-      background-color: white;
-      width: 100%;
-      height: 85%;
-      padding: 20px;
-    }
-  }
+		.graph-chart {
+			background-color: white;
+			width: 100%;
+			height: 85%;
+			padding: 20px;
+		}
+	}
 }
 
 .vertical-line {
-  height: 100%;
-  border-left: 1px solid #b5b5b5;
-  margin: 0 24px;
+	height: 100%;
+	border-left: 1px solid #b5b5b5;
+	margin: 0 24px;
 }
 
 .summary-box {
-  width: 40%;
-  height: 100%;
+	width: 40%;
+	height: 100%;
 
-  .summary-logo {
-    height: 30%;
+	.summary-logo {
+		height: 30%;
 
-    .logo-box {
-      background-color: white;
-      width: 10em;
-      height: 10em;
-      padding: 8px;
-      margin: auto;
+		.logo-box {
+			background-color: white;
+			width: 10em;
+			height: 10em;
+			padding: 8px;
+			margin: auto;
 
-      img {
-        width: 100%;
-        height: 100%;
-        object-fit: contain;
-      }
-    }
-  }
+			img {
+				width: 100%;
+				height: 100%;
+				object-fit: contain;
+			}
+		}
+	}
 
-  .summary-detail {
-    background-color: white;
-    height: 70%;
-    padding: 24px;
-    display: grid;
-    grid-template-rows: 0.1fr 1fr;
+	.summary-detail {
+		background-color: white;
+		height: 70%;
+		padding: 24px;
+		display: grid;
+		grid-template-rows: 0.1fr 1fr;
 
-    .summary-title {
-      font-size: 1.5em;
-      font-weight: bold;
-      padding-bottom: 24px;
-    }
+		.summary-title {
+			font-size: 1.5em;
+			font-weight: bold;
+			padding-bottom: 24px;
+		}
 
-    .summary-content {
-      word-break: break-all;
-      overflow: auto;
-      font-size: 1.2em;
-      // -ms-overflow-style: none; /* IE and Edge */
-      // scrollbar-width: none; /* Firefox */
-    }
-
-    .summary-content::-webkit-scrollbar {
-      // display: none; /* Chrome, Safari, Opera*/
-    }
-  }
+		.summary-content {
+			word-break: break-all;
+			overflow: auto;
+			font-size: 1.2em;
+		}
+	}
 }
