+++ src/components/Company/dummyData.js
@@ -1,7361 +1,7361 @@
 export const companyProfile = [
-  {
-    symbol: 'AAPL', //티커
-    price: 163.4, //실시간 가격
-    beta: 1.202736,
-    volAvg: 93034629,
-    mktCap: 2669040893952,
-    lastDiv: 0.865,
-    range: '116.21-182.94',
-    changes: -1.1100006, //가격 등락 -> 비율 등락 계산
-    companyName: 'Apple Inc.', //회사 이름
-    currency: 'USD', //기준 단위
-    cik: '0000320193',
-    isin: 'US0378331005',
-    cusip: '037833100',
-    exchange: 'Nasdaq Global Select',
-    exchangeShortName: 'NASDAQ',
-    industry: 'Consumer Electronics',
-    website: 'https://www.apple.com',
-    //설명
-    description:
-      "Apple Inc. (Apple) designs, manufactures and markets mobile communication and media devices, personal computers, and portable digital music players, and a variety of related software, services, peripherals, networking solutions, and third-party digital content and applications. The Company's products and services include iPhone, iPad, Mac, iPod, Apple TV, a portfolio of consumer and professional software applications, the iOS and OS X operating systems, iCloud, and a variety of accessory, service and support offerings. The Company also delivers digital content and applications through the iTunes Store, App StoreSM, iBookstoreSM, and Mac App Store. The Company distributes its products worldwide through its retail stores, online stores, and direct sales force, as well as through third-party cellular network carriers, wholesalers, retailers, and value-added resellers. In February 2012, the Company acquired app-search engine Chomp.",
-    ceo: 'Mr. Timothy Cook',
-    sector: 'Technology',
-    country: 'US',
-    fullTimeEmployees: '154000',
-    phone: '14089961010',
-    address: '1 Apple Park Way',
-    city: 'Cupertino',
-    state: 'CALIFORNIA',
-    zip: '95014',
-    dcfDiff: 89.92,
-    dcf: 166.582,
-    image: 'https://financialmodelingprep.com/image-stock/AAPL.png', //로고
-    ipoDate: '1980-12-12',
-    defaultImage: false,
-    isEtf: false,
-    isActivelyTrading: true,
-    isAdr: false,
-    isFund: false,
-  },
+	{
+		symbol: 'AAPL', //티커
+		price: 163.4, //실시간 가격
+		beta: 1.202736,
+		volAvg: 93034629,
+		mktCap: 2669040893952,
+		lastDiv: 0.865,
+		range: '116.21-182.94',
+		changes: -1.1100006, //가격 등락 -> 비율 등락 계산
+		companyName: 'Apple Inc.', //회사 이름
+		currency: 'USD', //기준 단위
+		cik: '0000320193',
+		isin: 'US0378331005',
+		cusip: '037833100',
+		exchange: 'Nasdaq Global Select',
+		exchangeShortName: 'NASDAQ',
+		industry: 'Consumer Electronics',
+		website: 'https://www.apple.com',
+		//설명
+		description:
+			"Apple Inc. (Apple) designs, manufactures and markets mobile communication and media devices, personal computers, and portable digital music players, and a variety of related software, services, peripherals, networking solutions, and third-party digital content and applications. The Company's products and services include iPhone, iPad, Mac, iPod, Apple TV, a portfolio of consumer and professional software applications, the iOS and OS X operating systems, iCloud, and a variety of accessory, service and support offerings. The Company also delivers digital content and applications through the iTunes Store, App StoreSM, iBookstoreSM, and Mac App Store. The Company distributes its products worldwide through its retail stores, online stores, and direct sales force, as well as through third-party cellular network carriers, wholesalers, retailers, and value-added resellers. In February 2012, the Company acquired app-search engine Chomp.",
+		ceo: 'Mr. Timothy Cook',
+		sector: 'Technology',
+		country: 'US',
+		fullTimeEmployees: '154000',
+		phone: '14089961010',
+		address: '1 Apple Park Way',
+		city: 'Cupertino',
+		state: 'CALIFORNIA',
+		zip: '95014',
+		dcfDiff: 89.92,
+		dcf: 166.582,
+		image: 'https://financialmodelingprep.com/image-stock/AAPL.png', //로고
+		ipoDate: '1980-12-12',
+		defaultImage: false,
+		isEtf: false,
+		isActivelyTrading: true,
+		isAdr: false,
+		isFund: false,
+	},
 ];
 
 export const companyQuote = [
-  {
-    symbol: 'AAPL', //티커
-    name: 'Apple Inc.', //이름
-    price: 164.105, //실시간 가격
-    changesPercentage: -0.2461849, //등락 비율
-    change: -0.40499878, //등락 가격
-    dayLow: 162.97,
-    dayHigh: 166.32,
-    yearHigh: 182.94,
-    yearLow: 116.21,
-    marketCap: 2680556617728.0,
-    priceAvg50: 168.0756,
-    priceAvg200: 147.22935,
-    volume: 77200885,
-    avgVolume: 93034629,
-    exchange: 'NASDAQ',
-    open: 164.415,
-    previousClose: 164.51, //전일 종가
-    eps: 5.61,
-    pe: 29.252226,
-    earningsAnnouncement: '2022-01-27T21:00:00.000+0000',
-    sharesOutstanding: 16334399426,
-    timestamp: 1642791863, //받아올때 시간
-  },
+	{
+		symbol: 'AAPL', //티커
+		name: 'Apple Inc.', //이름
+		price: 164.105, //실시간 가격
+		changesPercentage: -0.2461849, //등락 비율
+		change: -0.40499878, //등락 가격
+		dayLow: 162.97,
+		dayHigh: 166.32,
+		yearHigh: 182.94,
+		yearLow: 116.21,
+		marketCap: 2680556617728.0,
+		priceAvg50: 168.0756,
+		priceAvg200: 147.22935,
+		volume: 77200885,
+		avgVolume: 93034629,
+		exchange: 'NASDAQ',
+		open: 164.415,
+		previousClose: 164.51, //전일 종가
+		eps: 5.61,
+		pe: 29.252226,
+		earningsAnnouncement: '2022-01-27T21:00:00.000+0000',
+		sharesOutstanding: 16334399426,
+		timestamp: 1642791863, //받아올때 시간
+	},
 ];
 
 // 1Day / 1min
 export const stockHistory1Day = [
-  {
-    date: '2022-01-21 16:00:00',
-    open: 162.4,
-    low: 162.39,
-    high: 162.69,
-    close: 162.55,
-    volume: 1753842,
-  },
-  {
-    date: '2022-01-21 15:55:00',
-    open: 162.65,
-    low: 162.3,
-    high: 163.18,
-    close: 162.41,
-    volume: 4583337,
-  },
-  {
-    date: '2022-01-21 15:50:00',
-    open: 163.18,
-    low: 162.52,
-    high: 163.29,
-    close: 162.6399,
-    volume: 2425008,
-  },
-  {
-    date: '2022-01-21 15:45:00',
-    open: 163.19,
-    low: 162.92,
-    high: 163.4083,
-    close: 163.18,
-    volume: 1648816,
-  },
-  {
-    date: '2022-01-21 15:40:00',
-    open: 163.15,
-    low: 163.14,
-    high: 163.885,
-    close: 163.18,
-    volume: 1896376,
-  },
-  {
-    date: '2022-01-21 15:35:00',
-    open: 163.04,
-    low: 162.62,
-    high: 163.24,
-    close: 163.16,
-    volume: 1873912,
-  },
-  {
-    date: '2022-01-21 15:30:00',
-    open: 162.86,
-    low: 162.57,
-    high: 163.1699,
-    close: 163.06,
-    volume: 1896565,
-  },
-  {
-    date: '2022-01-21 15:25:00',
-    open: 163.0,
-    low: 162.581,
-    high: 163.1,
-    close: 162.86,
-    volume: 1477045,
-  },
-  {
-    date: '2022-01-21 15:20:00',
-    open: 163.09,
-    low: 162.7,
-    high: 163.15,
-    close: 162.99,
-    volume: 1348166,
-  },
-  {
-    date: '2022-01-21 15:15:00',
-    open: 162.88,
-    low: 162.7,
-    high: 163.45,
-    close: 163.12,
-    volume: 1490197,
-  },
-  {
-    date: '2022-01-21 15:10:00',
-    open: 162.53,
-    low: 162.405,
-    high: 162.9399,
-    close: 162.89,
-    volume: 1494587,
-  },
-  {
-    date: '2022-01-21 15:05:00',
-    open: 162.88,
-    low: 162.43,
-    high: 162.89,
-    close: 162.53,
-    volume: 1736393,
-  },
-  {
-    date: '2022-01-21 15:00:00',
-    open: 163.1195,
-    low: 162.8,
-    high: 163.38,
-    close: 162.88,
-    volume: 1480466,
-  },
-  {
-    date: '2022-01-21 14:55:00',
-    open: 163.48,
-    low: 163.06,
-    high: 163.64,
-    close: 163.11,
-    volume: 1255614,
-  },
-  {
-    date: '2022-01-21 14:50:00',
-    open: 163.73,
-    low: 163.31,
-    high: 163.73,
-    close: 163.4727,
-    volume: 692963,
-  },
-  {
-    date: '2022-01-21 14:45:00',
-    open: 163.53,
-    low: 163.49,
-    high: 163.765,
-    close: 163.7283,
-    volume: 691663,
-  },
-  {
-    date: '2022-01-21 14:40:00',
-    open: 163.7,
-    low: 163.44,
-    high: 163.92,
-    close: 163.51,
-    volume: 868701,
-  },
-  {
-    date: '2022-01-21 14:35:00',
-    open: 163.39,
-    low: 163.2,
-    high: 163.79,
-    close: 163.7,
-    volume: 1044987,
-  },
-  {
-    date: '2022-01-21 14:30:00',
-    open: 163.365,
-    low: 163.25,
-    high: 163.62,
-    close: 163.38,
-    volume: 855139,
-  },
-  {
-    date: '2022-01-21 14:25:00',
-    open: 164.22,
-    low: 163.3,
-    high: 164.24,
-    close: 163.37,
-    volume: 1089930,
-  },
-  {
-    date: '2022-01-21 14:20:00',
-    open: 163.91,
-    low: 163.69,
-    high: 164.32,
-    close: 164.23,
-    volume: 931845,
-  },
-  {
-    date: '2022-01-21 14:15:00',
-    open: 164.0,
-    low: 163.78,
-    high: 164.29,
-    close: 163.91,
-    volume: 771609,
-  },
-  {
-    date: '2022-01-21 14:10:00',
-    open: 164.015,
-    low: 163.9,
-    high: 164.3299,
-    close: 163.996,
-    volume: 802700,
-  },
-  {
-    date: '2022-01-21 14:05:00',
-    open: 164.07,
-    low: 163.84,
-    high: 164.3299,
-    close: 164.0,
-    volume: 854633,
-  },
-  {
-    date: '2022-01-21 14:00:00',
-    open: 163.765,
-    low: 163.7,
-    high: 164.14,
-    close: 164.08,
-    volume: 771297,
-  },
-  {
-    date: '2022-01-21 13:55:00',
-    open: 164.2819,
-    low: 163.62,
-    high: 164.34,
-    close: 163.765,
-    volume: 794903,
-  },
-  {
-    date: '2022-01-21 13:50:00',
-    open: 164.22,
-    low: 164.05,
-    high: 164.53,
-    close: 164.28,
-    volume: 677393,
-  },
-  {
-    date: '2022-01-21 13:45:00',
-    open: 164.17,
-    low: 163.91,
-    high: 164.32,
-    close: 164.24,
-    volume: 769442,
-  },
-  {
-    date: '2022-01-21 13:40:00',
-    open: 164.26,
-    low: 164.08,
-    high: 164.5,
-    close: 164.16,
-    volume: 706178,
-  },
-  {
-    date: '2022-01-21 13:35:00',
-    open: 163.729,
-    low: 163.57,
-    high: 164.42,
-    close: 164.25,
-    volume: 811891,
-  },
-  {
-    date: '2022-01-21 13:30:00',
-    open: 163.75,
-    low: 163.66,
-    high: 164.07,
-    close: 163.73,
-    volume: 844913,
-  },
-  {
-    date: '2022-01-21 13:25:00',
-    open: 163.45,
-    low: 163.39,
-    high: 163.93,
-    close: 163.75,
-    volume: 868020,
-  },
-  {
-    date: '2022-01-21 13:20:00',
-    open: 164.15,
-    low: 163.35,
-    high: 164.2,
-    close: 163.44,
-    volume: 907445,
-  },
-  {
-    date: '2022-01-21 13:15:00',
-    open: 164.62,
-    low: 163.9725,
-    high: 164.69,
-    close: 164.16,
-    volume: 736636,
-  },
-  {
-    date: '2022-01-21 13:10:00',
-    open: 164.02,
-    low: 164.02,
-    high: 164.78,
-    close: 164.625,
-    volume: 987059,
-  },
-  {
-    date: '2022-01-21 13:05:00',
-    open: 164.0185,
-    low: 163.86,
-    high: 164.245,
-    close: 164.025,
-    volume: 827993,
-  },
-  {
-    date: '2022-01-21 13:00:00',
-    open: 163.95,
-    low: 163.9,
-    high: 164.28,
-    close: 164.01,
-    volume: 794247,
-  },
-  {
-    date: '2022-01-21 12:55:00',
-    open: 163.61,
-    low: 163.46,
-    high: 164.03,
-    close: 163.96,
-    volume: 888389,
-  },
-  {
-    date: '2022-01-21 12:50:00',
-    open: 163.38,
-    low: 163.26,
-    high: 163.64,
-    close: 163.61,
-    volume: 648479,
-  },
-  {
-    date: '2022-01-21 12:45:00',
-    open: 163.54,
-    low: 163.25,
-    high: 163.82,
-    close: 163.37,
-    volume: 802509,
-  },
-  {
-    date: '2022-01-21 12:40:00',
-    open: 163.64,
-    low: 163.46,
-    high: 163.8914,
-    close: 163.5304,
-    volume: 709103,
-  },
-  {
-    date: '2022-01-21 12:35:00',
-    open: 163.63,
-    low: 163.402,
-    high: 163.97,
-    close: 163.65,
-    volume: 832765,
-  },
-  {
-    date: '2022-01-21 12:30:00',
-    open: 163.73,
-    low: 163.48,
-    high: 163.84,
-    close: 163.6201,
-    volume: 758662,
-  },
-  {
-    date: '2022-01-21 12:25:00',
-    open: 164.02,
-    low: 163.6946,
-    high: 164.23,
-    close: 163.73,
-    volume: 828334,
-  },
-  {
-    date: '2022-01-21 12:20:00',
-    open: 163.79,
-    low: 163.52,
-    high: 164.04,
-    close: 164.04,
-    volume: 1163135,
-  },
-  {
-    date: '2022-01-21 12:15:00',
-    open: 163.91,
-    low: 163.68,
-    high: 164.15,
-    close: 163.8,
-    volume: 922942,
-  },
-  {
-    date: '2022-01-21 12:10:00',
-    open: 164.13,
-    low: 163.8,
-    high: 164.31,
-    close: 163.91,
-    volume: 1051596,
-  },
-  {
-    date: '2022-01-21 12:05:00',
-    open: 164.4,
-    low: 164.0901,
-    high: 164.63,
-    close: 164.16,
-    volume: 802163,
-  },
-  {
-    date: '2022-01-21 12:00:00',
-    open: 164.42,
-    low: 164.17,
-    high: 164.67,
-    close: 164.38,
-    volume: 826696,
-  },
-  {
-    date: '2022-01-21 11:55:00',
-    open: 164.22,
-    low: 164.15,
-    high: 164.54,
-    close: 164.44,
-    volume: 909647,
-  },
-  {
-    date: '2022-01-21 11:50:00',
-    open: 164.6,
-    low: 164.02,
-    high: 164.7,
-    close: 164.2088,
-    volume: 894350,
-  },
-  {
-    date: '2022-01-21 11:45:00',
-    open: 165.08,
-    low: 164.45,
-    high: 165.11,
-    close: 164.6,
-    volume: 1276147,
-  },
-  {
-    date: '2022-01-21 11:40:00',
-    open: 165.06,
-    low: 164.87,
-    high: 165.2182,
-    close: 165.05,
-    volume: 750005,
-  },
-  {
-    date: '2022-01-21 11:35:00',
-    open: 165.78,
-    low: 164.86,
-    high: 165.94,
-    close: 165.06,
-    volume: 1308744,
-  },
-  {
-    date: '2022-01-21 11:30:00',
-    open: 165.94,
-    low: 165.1428,
-    high: 165.95,
-    close: 165.78,
-    volume: 1371342,
-  },
-  {
-    date: '2022-01-21 11:25:00',
-    open: 166.07,
-    low: 165.78,
-    high: 166.1898,
-    close: 165.95,
-    volume: 963225,
-  },
-  {
-    date: '2022-01-21 11:20:00',
-    open: 165.7261,
-    low: 165.44,
-    high: 166.33,
-    close: 166.07,
-    volume: 1372225,
-  },
-  {
-    date: '2022-01-21 11:15:00',
-    open: 165.07,
-    low: 165.07,
-    high: 165.88,
-    close: 165.715,
-    volume: 1500335,
-  },
-  {
-    date: '2022-01-21 11:10:00',
-    open: 165.39,
-    low: 164.93,
-    high: 165.51,
-    close: 165.06,
-    volume: 2200959,
-  },
-  {
-    date: '2022-01-21 11:05:00',
-    open: 164.97,
-    low: 164.93,
-    high: 165.41,
-    close: 165.4,
-    volume: 1213898,
-  },
-  {
-    date: '2022-01-21 11:00:00',
-    open: 165.17,
-    low: 164.65,
-    high: 165.25,
-    close: 164.95,
-    volume: 1179103,
-  },
-  {
-    date: '2022-01-21 10:55:00',
-    open: 165.52,
-    low: 164.82,
-    high: 165.67,
-    close: 165.16,
-    volume: 1263572,
-  },
-  {
-    date: '2022-01-21 10:50:00',
-    open: 165.22,
-    low: 165.13,
-    high: 165.84,
-    close: 165.53,
-    volume: 1795511,
-  },
-  {
-    date: '2022-01-21 10:45:00',
-    open: 165.03,
-    low: 164.67,
-    high: 165.2897,
-    close: 165.2014,
-    volume: 1113790,
-  },
-  {
-    date: '2022-01-21 10:40:00',
-    open: 164.94,
-    low: 164.62,
-    high: 165.29,
-    close: 165.03,
-    volume: 1515370,
-  },
-  {
-    date: '2022-01-21 10:35:00',
-    open: 165.45,
-    low: 164.9,
-    high: 165.78,
-    close: 164.9399,
-    volume: 1762388,
-  },
-  {
-    date: '2022-01-21 10:30:00',
-    open: 164.31,
-    low: 164.13,
-    high: 165.54,
-    close: 165.46,
-    volume: 2130613,
-  },
-  {
-    date: '2022-01-21 10:25:00',
-    open: 163.67,
-    low: 163.54,
-    high: 164.6,
-    close: 164.31,
-    volume: 1752836,
-  },
-  {
-    date: '2022-01-21 10:20:00',
-    open: 163.3,
-    low: 162.96,
-    high: 163.8376,
-    close: 163.68,
-    volume: 2401416,
-  },
-  {
-    date: '2022-01-21 10:15:00',
-    open: 163.6165,
-    low: 163.11,
-    high: 163.82,
-    close: 163.28,
-    volume: 1613638,
-  },
-  {
-    date: '2022-01-21 10:10:00',
-    open: 163.9,
-    low: 163.3299,
-    high: 164.0,
-    close: 163.63,
-    volume: 1542727,
-  },
-  {
-    date: '2022-01-21 10:05:00',
-    open: 163.32,
-    low: 163.1604,
-    high: 164.06,
-    close: 163.9,
-    volume: 1951077,
-  },
-  {
-    date: '2022-01-21 10:00:00',
-    open: 163.54,
-    low: 163.22,
-    high: 163.95,
-    close: 163.3201,
-    volume: 1647151,
-  },
-  {
-    date: '2022-01-21 09:55:00',
-    open: 163.35,
-    low: 163.18,
-    high: 163.78,
-    close: 163.53,
-    volume: 1408586,
-  },
-  {
-    date: '2022-01-21 09:50:00',
-    open: 164.77,
-    low: 163.23,
-    high: 164.9,
-    close: 163.365,
-    volume: 2033414,
-  },
-  {
-    date: '2022-01-21 09:45:00',
-    open: 163.71,
-    low: 163.67,
-    high: 164.89,
-    close: 164.77,
-    volume: 1909257,
-  },
-  {
-    date: '2022-01-21 09:40:00',
-    open: 164.52,
-    low: 163.7,
-    high: 164.71,
-    close: 163.72,
-    volume: 3802988,
-  },
-  {
-    date: '2022-01-21 09:35:00',
-    open: 165.2501,
-    low: 164.42,
-    high: 165.6,
-    close: 164.51,
-    volume: 2253073,
-  },
-  {
-    date: '2022-01-21 09:30:00',
-    open: 164.415,
-    low: 164.36,
-    high: 166.0,
-    close: 165.26,
-    volume: 7362686,
-  },
+	{
+		date: '2022-01-21 16:00:00',
+		open: 162.4,
+		low: 162.39,
+		high: 162.69,
+		close: 162.55,
+		volume: 1753842,
+	},
+	{
+		date: '2022-01-21 15:55:00',
+		open: 162.65,
+		low: 162.3,
+		high: 163.18,
+		close: 162.41,
+		volume: 4583337,
+	},
+	{
+		date: '2022-01-21 15:50:00',
+		open: 163.18,
+		low: 162.52,
+		high: 163.29,
+		close: 162.6399,
+		volume: 2425008,
+	},
+	{
+		date: '2022-01-21 15:45:00',
+		open: 163.19,
+		low: 162.92,
+		high: 163.4083,
+		close: 163.18,
+		volume: 1648816,
+	},
+	{
+		date: '2022-01-21 15:40:00',
+		open: 163.15,
+		low: 163.14,
+		high: 163.885,
+		close: 163.18,
+		volume: 1896376,
+	},
+	{
+		date: '2022-01-21 15:35:00',
+		open: 163.04,
+		low: 162.62,
+		high: 163.24,
+		close: 163.16,
+		volume: 1873912,
+	},
+	{
+		date: '2022-01-21 15:30:00',
+		open: 162.86,
+		low: 162.57,
+		high: 163.1699,
+		close: 163.06,
+		volume: 1896565,
+	},
+	{
+		date: '2022-01-21 15:25:00',
+		open: 163.0,
+		low: 162.581,
+		high: 163.1,
+		close: 162.86,
+		volume: 1477045,
+	},
+	{
+		date: '2022-01-21 15:20:00',
+		open: 163.09,
+		low: 162.7,
+		high: 163.15,
+		close: 162.99,
+		volume: 1348166,
+	},
+	{
+		date: '2022-01-21 15:15:00',
+		open: 162.88,
+		low: 162.7,
+		high: 163.45,
+		close: 163.12,
+		volume: 1490197,
+	},
+	{
+		date: '2022-01-21 15:10:00',
+		open: 162.53,
+		low: 162.405,
+		high: 162.9399,
+		close: 162.89,
+		volume: 1494587,
+	},
+	{
+		date: '2022-01-21 15:05:00',
+		open: 162.88,
+		low: 162.43,
+		high: 162.89,
+		close: 162.53,
+		volume: 1736393,
+	},
+	{
+		date: '2022-01-21 15:00:00',
+		open: 163.1195,
+		low: 162.8,
+		high: 163.38,
+		close: 162.88,
+		volume: 1480466,
+	},
+	{
+		date: '2022-01-21 14:55:00',
+		open: 163.48,
+		low: 163.06,
+		high: 163.64,
+		close: 163.11,
+		volume: 1255614,
+	},
+	{
+		date: '2022-01-21 14:50:00',
+		open: 163.73,
+		low: 163.31,
+		high: 163.73,
+		close: 163.4727,
+		volume: 692963,
+	},
+	{
+		date: '2022-01-21 14:45:00',
+		open: 163.53,
+		low: 163.49,
+		high: 163.765,
+		close: 163.7283,
+		volume: 691663,
+	},
+	{
+		date: '2022-01-21 14:40:00',
+		open: 163.7,
+		low: 163.44,
+		high: 163.92,
+		close: 163.51,
+		volume: 868701,
+	},
+	{
+		date: '2022-01-21 14:35:00',
+		open: 163.39,
+		low: 163.2,
+		high: 163.79,
+		close: 163.7,
+		volume: 1044987,
+	},
+	{
+		date: '2022-01-21 14:30:00',
+		open: 163.365,
+		low: 163.25,
+		high: 163.62,
+		close: 163.38,
+		volume: 855139,
+	},
+	{
+		date: '2022-01-21 14:25:00',
+		open: 164.22,
+		low: 163.3,
+		high: 164.24,
+		close: 163.37,
+		volume: 1089930,
+	},
+	{
+		date: '2022-01-21 14:20:00',
+		open: 163.91,
+		low: 163.69,
+		high: 164.32,
+		close: 164.23,
+		volume: 931845,
+	},
+	{
+		date: '2022-01-21 14:15:00',
+		open: 164.0,
+		low: 163.78,
+		high: 164.29,
+		close: 163.91,
+		volume: 771609,
+	},
+	{
+		date: '2022-01-21 14:10:00',
+		open: 164.015,
+		low: 163.9,
+		high: 164.3299,
+		close: 163.996,
+		volume: 802700,
+	},
+	{
+		date: '2022-01-21 14:05:00',
+		open: 164.07,
+		low: 163.84,
+		high: 164.3299,
+		close: 164.0,
+		volume: 854633,
+	},
+	{
+		date: '2022-01-21 14:00:00',
+		open: 163.765,
+		low: 163.7,
+		high: 164.14,
+		close: 164.08,
+		volume: 771297,
+	},
+	{
+		date: '2022-01-21 13:55:00',
+		open: 164.2819,
+		low: 163.62,
+		high: 164.34,
+		close: 163.765,
+		volume: 794903,
+	},
+	{
+		date: '2022-01-21 13:50:00',
+		open: 164.22,
+		low: 164.05,
+		high: 164.53,
+		close: 164.28,
+		volume: 677393,
+	},
+	{
+		date: '2022-01-21 13:45:00',
+		open: 164.17,
+		low: 163.91,
+		high: 164.32,
+		close: 164.24,
+		volume: 769442,
+	},
+	{
+		date: '2022-01-21 13:40:00',
+		open: 164.26,
+		low: 164.08,
+		high: 164.5,
+		close: 164.16,
+		volume: 706178,
+	},
+	{
+		date: '2022-01-21 13:35:00',
+		open: 163.729,
+		low: 163.57,
+		high: 164.42,
+		close: 164.25,
+		volume: 811891,
+	},
+	{
+		date: '2022-01-21 13:30:00',
+		open: 163.75,
+		low: 163.66,
+		high: 164.07,
+		close: 163.73,
+		volume: 844913,
+	},
+	{
+		date: '2022-01-21 13:25:00',
+		open: 163.45,
+		low: 163.39,
+		high: 163.93,
+		close: 163.75,
+		volume: 868020,
+	},
+	{
+		date: '2022-01-21 13:20:00',
+		open: 164.15,
+		low: 163.35,
+		high: 164.2,
+		close: 163.44,
+		volume: 907445,
+	},
+	{
+		date: '2022-01-21 13:15:00',
+		open: 164.62,
+		low: 163.9725,
+		high: 164.69,
+		close: 164.16,
+		volume: 736636,
+	},
+	{
+		date: '2022-01-21 13:10:00',
+		open: 164.02,
+		low: 164.02,
+		high: 164.78,
+		close: 164.625,
+		volume: 987059,
+	},
+	{
+		date: '2022-01-21 13:05:00',
+		open: 164.0185,
+		low: 163.86,
+		high: 164.245,
+		close: 164.025,
+		volume: 827993,
+	},
+	{
+		date: '2022-01-21 13:00:00',
+		open: 163.95,
+		low: 163.9,
+		high: 164.28,
+		close: 164.01,
+		volume: 794247,
+	},
+	{
+		date: '2022-01-21 12:55:00',
+		open: 163.61,
+		low: 163.46,
+		high: 164.03,
+		close: 163.96,
+		volume: 888389,
+	},
+	{
+		date: '2022-01-21 12:50:00',
+		open: 163.38,
+		low: 163.26,
+		high: 163.64,
+		close: 163.61,
+		volume: 648479,
+	},
+	{
+		date: '2022-01-21 12:45:00',
+		open: 163.54,
+		low: 163.25,
+		high: 163.82,
+		close: 163.37,
+		volume: 802509,
+	},
+	{
+		date: '2022-01-21 12:40:00',
+		open: 163.64,
+		low: 163.46,
+		high: 163.8914,
+		close: 163.5304,
+		volume: 709103,
+	},
+	{
+		date: '2022-01-21 12:35:00',
+		open: 163.63,
+		low: 163.402,
+		high: 163.97,
+		close: 163.65,
+		volume: 832765,
+	},
+	{
+		date: '2022-01-21 12:30:00',
+		open: 163.73,
+		low: 163.48,
+		high: 163.84,
+		close: 163.6201,
+		volume: 758662,
+	},
+	{
+		date: '2022-01-21 12:25:00',
+		open: 164.02,
+		low: 163.6946,
+		high: 164.23,
+		close: 163.73,
+		volume: 828334,
+	},
+	{
+		date: '2022-01-21 12:20:00',
+		open: 163.79,
+		low: 163.52,
+		high: 164.04,
+		close: 164.04,
+		volume: 1163135,
+	},
+	{
+		date: '2022-01-21 12:15:00',
+		open: 163.91,
+		low: 163.68,
+		high: 164.15,
+		close: 163.8,
+		volume: 922942,
+	},
+	{
+		date: '2022-01-21 12:10:00',
+		open: 164.13,
+		low: 163.8,
+		high: 164.31,
+		close: 163.91,
+		volume: 1051596,
+	},
+	{
+		date: '2022-01-21 12:05:00',
+		open: 164.4,
+		low: 164.0901,
+		high: 164.63,
+		close: 164.16,
+		volume: 802163,
+	},
+	{
+		date: '2022-01-21 12:00:00',
+		open: 164.42,
+		low: 164.17,
+		high: 164.67,
+		close: 164.38,
+		volume: 826696,
+	},
+	{
+		date: '2022-01-21 11:55:00',
+		open: 164.22,
+		low: 164.15,
+		high: 164.54,
+		close: 164.44,
+		volume: 909647,
+	},
+	{
+		date: '2022-01-21 11:50:00',
+		open: 164.6,
+		low: 164.02,
+		high: 164.7,
+		close: 164.2088,
+		volume: 894350,
+	},
+	{
+		date: '2022-01-21 11:45:00',
+		open: 165.08,
+		low: 164.45,
+		high: 165.11,
+		close: 164.6,
+		volume: 1276147,
+	},
+	{
+		date: '2022-01-21 11:40:00',
+		open: 165.06,
+		low: 164.87,
+		high: 165.2182,
+		close: 165.05,
+		volume: 750005,
+	},
+	{
+		date: '2022-01-21 11:35:00',
+		open: 165.78,
+		low: 164.86,
+		high: 165.94,
+		close: 165.06,
+		volume: 1308744,
+	},
+	{
+		date: '2022-01-21 11:30:00',
+		open: 165.94,
+		low: 165.1428,
+		high: 165.95,
+		close: 165.78,
+		volume: 1371342,
+	},
+	{
+		date: '2022-01-21 11:25:00',
+		open: 166.07,
+		low: 165.78,
+		high: 166.1898,
+		close: 165.95,
+		volume: 963225,
+	},
+	{
+		date: '2022-01-21 11:20:00',
+		open: 165.7261,
+		low: 165.44,
+		high: 166.33,
+		close: 166.07,
+		volume: 1372225,
+	},
+	{
+		date: '2022-01-21 11:15:00',
+		open: 165.07,
+		low: 165.07,
+		high: 165.88,
+		close: 165.715,
+		volume: 1500335,
+	},
+	{
+		date: '2022-01-21 11:10:00',
+		open: 165.39,
+		low: 164.93,
+		high: 165.51,
+		close: 165.06,
+		volume: 2200959,
+	},
+	{
+		date: '2022-01-21 11:05:00',
+		open: 164.97,
+		low: 164.93,
+		high: 165.41,
+		close: 165.4,
+		volume: 1213898,
+	},
+	{
+		date: '2022-01-21 11:00:00',
+		open: 165.17,
+		low: 164.65,
+		high: 165.25,
+		close: 164.95,
+		volume: 1179103,
+	},
+	{
+		date: '2022-01-21 10:55:00',
+		open: 165.52,
+		low: 164.82,
+		high: 165.67,
+		close: 165.16,
+		volume: 1263572,
+	},
+	{
+		date: '2022-01-21 10:50:00',
+		open: 165.22,
+		low: 165.13,
+		high: 165.84,
+		close: 165.53,
+		volume: 1795511,
+	},
+	{
+		date: '2022-01-21 10:45:00',
+		open: 165.03,
+		low: 164.67,
+		high: 165.2897,
+		close: 165.2014,
+		volume: 1113790,
+	},
+	{
+		date: '2022-01-21 10:40:00',
+		open: 164.94,
+		low: 164.62,
+		high: 165.29,
+		close: 165.03,
+		volume: 1515370,
+	},
+	{
+		date: '2022-01-21 10:35:00',
+		open: 165.45,
+		low: 164.9,
+		high: 165.78,
+		close: 164.9399,
+		volume: 1762388,
+	},
+	{
+		date: '2022-01-21 10:30:00',
+		open: 164.31,
+		low: 164.13,
+		high: 165.54,
+		close: 165.46,
+		volume: 2130613,
+	},
+	{
+		date: '2022-01-21 10:25:00',
+		open: 163.67,
+		low: 163.54,
+		high: 164.6,
+		close: 164.31,
+		volume: 1752836,
+	},
+	{
+		date: '2022-01-21 10:20:00',
+		open: 163.3,
+		low: 162.96,
+		high: 163.8376,
+		close: 163.68,
+		volume: 2401416,
+	},
+	{
+		date: '2022-01-21 10:15:00',
+		open: 163.6165,
+		low: 163.11,
+		high: 163.82,
+		close: 163.28,
+		volume: 1613638,
+	},
+	{
+		date: '2022-01-21 10:10:00',
+		open: 163.9,
+		low: 163.3299,
+		high: 164.0,
+		close: 163.63,
+		volume: 1542727,
+	},
+	{
+		date: '2022-01-21 10:05:00',
+		open: 163.32,
+		low: 163.1604,
+		high: 164.06,
+		close: 163.9,
+		volume: 1951077,
+	},
+	{
+		date: '2022-01-21 10:00:00',
+		open: 163.54,
+		low: 163.22,
+		high: 163.95,
+		close: 163.3201,
+		volume: 1647151,
+	},
+	{
+		date: '2022-01-21 09:55:00',
+		open: 163.35,
+		low: 163.18,
+		high: 163.78,
+		close: 163.53,
+		volume: 1408586,
+	},
+	{
+		date: '2022-01-21 09:50:00',
+		open: 164.77,
+		low: 163.23,
+		high: 164.9,
+		close: 163.365,
+		volume: 2033414,
+	},
+	{
+		date: '2022-01-21 09:45:00',
+		open: 163.71,
+		low: 163.67,
+		high: 164.89,
+		close: 164.77,
+		volume: 1909257,
+	},
+	{
+		date: '2022-01-21 09:40:00',
+		open: 164.52,
+		low: 163.7,
+		high: 164.71,
+		close: 163.72,
+		volume: 3802988,
+	},
+	{
+		date: '2022-01-21 09:35:00',
+		open: 165.2501,
+		low: 164.42,
+		high: 165.6,
+		close: 164.51,
+		volume: 2253073,
+	},
+	{
+		date: '2022-01-21 09:30:00',
+		open: 164.415,
+		low: 164.36,
+		high: 166.0,
+		close: 165.26,
+		volume: 7362686,
+	},
 ];
 
 //week / 5min
 export const stockHistory5Day = [
-  {
-    date: '2022-01-21 16:00:00',
-    open: 162.4,
-    low: 162.39,
-    high: 162.69,
-    close: 162.55,
-    volume: 1753842,
-  },
-  {
-    date: '2022-01-21 15:55:00',
-    open: 162.65,
-    low: 162.3,
-    high: 163.18,
-    close: 162.41,
-    volume: 4583337,
-  },
-  {
-    date: '2022-01-21 15:50:00',
-    open: 163.18,
-    low: 162.52,
-    high: 163.29,
-    close: 162.6399,
-    volume: 2425008,
-  },
-  {
-    date: '2022-01-21 15:45:00',
-    open: 163.19,
-    low: 162.92,
-    high: 163.4083,
-    close: 163.18,
-    volume: 1648816,
-  },
-  {
-    date: '2022-01-21 15:40:00',
-    open: 163.15,
-    low: 163.14,
-    high: 163.885,
-    close: 163.18,
-    volume: 1896376,
-  },
-  {
-    date: '2022-01-21 15:35:00',
-    open: 163.04,
-    low: 162.62,
-    high: 163.24,
-    close: 163.16,
-    volume: 1873912,
-  },
-  {
-    date: '2022-01-21 15:30:00',
-    open: 162.86,
-    low: 162.57,
-    high: 163.1699,
-    close: 163.06,
-    volume: 1896565,
-  },
-  {
-    date: '2022-01-21 15:25:00',
-    open: 163.0,
-    low: 162.581,
-    high: 163.1,
-    close: 162.86,
-    volume: 1477045,
-  },
-  {
-    date: '2022-01-21 15:20:00',
-    open: 163.09,
-    low: 162.7,
-    high: 163.15,
-    close: 162.99,
-    volume: 1348166,
-  },
-  {
-    date: '2022-01-21 15:15:00',
-    open: 162.88,
-    low: 162.7,
-    high: 163.45,
-    close: 163.12,
-    volume: 1490197,
-  },
-  {
-    date: '2022-01-21 15:10:00',
-    open: 162.53,
-    low: 162.405,
-    high: 162.9399,
-    close: 162.89,
-    volume: 1494587,
-  },
-  {
-    date: '2022-01-21 15:05:00',
-    open: 162.88,
-    low: 162.43,
-    high: 162.89,
-    close: 162.53,
-    volume: 1736393,
-  },
-  {
-    date: '2022-01-21 15:00:00',
-    open: 163.1195,
-    low: 162.8,
-    high: 163.38,
-    close: 162.88,
-    volume: 1480466,
-  },
-  {
-    date: '2022-01-21 14:55:00',
-    open: 163.48,
-    low: 163.06,
-    high: 163.64,
-    close: 163.11,
-    volume: 1255614,
-  },
-  {
-    date: '2022-01-21 14:50:00',
-    open: 163.73,
-    low: 163.31,
-    high: 163.73,
-    close: 163.4727,
-    volume: 692963,
-  },
-  {
-    date: '2022-01-21 14:45:00',
-    open: 163.53,
-    low: 163.49,
-    high: 163.765,
-    close: 163.7283,
-    volume: 691663,
-  },
-  {
-    date: '2022-01-21 14:40:00',
-    open: 163.7,
-    low: 163.44,
-    high: 163.92,
-    close: 163.51,
-    volume: 868701,
-  },
-  {
-    date: '2022-01-21 14:35:00',
-    open: 163.39,
-    low: 163.2,
-    high: 163.79,
-    close: 163.7,
-    volume: 1044987,
-  },
-  {
-    date: '2022-01-21 14:30:00',
-    open: 163.365,
-    low: 163.25,
-    high: 163.62,
-    close: 163.38,
-    volume: 855139,
-  },
-  {
-    date: '2022-01-21 14:25:00',
-    open: 164.22,
-    low: 163.3,
-    high: 164.24,
-    close: 163.37,
-    volume: 1089930,
-  },
-  {
-    date: '2022-01-21 14:20:00',
-    open: 163.91,
-    low: 163.69,
-    high: 164.32,
-    close: 164.23,
-    volume: 931845,
-  },
-  {
-    date: '2022-01-21 14:15:00',
-    open: 164.0,
-    low: 163.78,
-    high: 164.29,
-    close: 163.91,
-    volume: 771609,
-  },
-  {
-    date: '2022-01-21 14:10:00',
-    open: 164.015,
-    low: 163.9,
-    high: 164.3299,
-    close: 163.996,
-    volume: 802700,
-  },
-  {
-    date: '2022-01-21 14:05:00',
-    open: 164.07,
-    low: 163.84,
-    high: 164.3299,
-    close: 164.0,
-    volume: 854633,
-  },
-  {
-    date: '2022-01-21 14:00:00',
-    open: 163.765,
-    low: 163.7,
-    high: 164.14,
-    close: 164.08,
-    volume: 771297,
-  },
-  {
-    date: '2022-01-21 13:55:00',
-    open: 164.2819,
-    low: 163.62,
-    high: 164.34,
-    close: 163.765,
-    volume: 794903,
-  },
-  {
-    date: '2022-01-21 13:50:00',
-    open: 164.22,
-    low: 164.05,
-    high: 164.53,
-    close: 164.28,
-    volume: 677393,
-  },
-  {
-    date: '2022-01-21 13:45:00',
-    open: 164.17,
-    low: 163.91,
-    high: 164.32,
-    close: 164.24,
-    volume: 769442,
-  },
-  {
-    date: '2022-01-21 13:40:00',
-    open: 164.26,
-    low: 164.08,
-    high: 164.5,
-    close: 164.16,
-    volume: 706178,
-  },
-  {
-    date: '2022-01-21 13:35:00',
-    open: 163.729,
-    low: 163.57,
-    high: 164.42,
-    close: 164.25,
-    volume: 811891,
-  },
-  {
-    date: '2022-01-21 13:30:00',
-    open: 163.75,
-    low: 163.66,
-    high: 164.07,
-    close: 163.73,
-    volume: 844913,
-  },
-  {
-    date: '2022-01-21 13:25:00',
-    open: 163.45,
-    low: 163.39,
-    high: 163.93,
-    close: 163.75,
-    volume: 868020,
-  },
-  {
-    date: '2022-01-21 13:20:00',
-    open: 164.15,
-    low: 163.35,
-    high: 164.2,
-    close: 163.44,
-    volume: 907445,
-  },
-  {
-    date: '2022-01-21 13:15:00',
-    open: 164.62,
-    low: 163.9725,
-    high: 164.69,
-    close: 164.16,
-    volume: 736636,
-  },
-  {
-    date: '2022-01-21 13:10:00',
-    open: 164.02,
-    low: 164.02,
-    high: 164.78,
-    close: 164.625,
-    volume: 987059,
-  },
-  {
-    date: '2022-01-21 13:05:00',
-    open: 164.0185,
-    low: 163.86,
-    high: 164.245,
-    close: 164.025,
-    volume: 827993,
-  },
-  {
-    date: '2022-01-21 13:00:00',
-    open: 163.95,
-    low: 163.9,
-    high: 164.28,
-    close: 164.01,
-    volume: 794247,
-  },
-  {
-    date: '2022-01-21 12:55:00',
-    open: 163.61,
-    low: 163.46,
-    high: 164.03,
-    close: 163.96,
-    volume: 888389,
-  },
-  {
-    date: '2022-01-21 12:50:00',
-    open: 163.38,
-    low: 163.26,
-    high: 163.64,
-    close: 163.61,
-    volume: 648479,
-  },
-  {
-    date: '2022-01-21 12:45:00',
-    open: 163.54,
-    low: 163.25,
-    high: 163.82,
-    close: 163.37,
-    volume: 802509,
-  },
-  {
-    date: '2022-01-21 12:40:00',
-    open: 163.64,
-    low: 163.46,
-    high: 163.8914,
-    close: 163.5304,
-    volume: 709103,
-  },
-  {
-    date: '2022-01-21 12:35:00',
-    open: 163.63,
-    low: 163.402,
-    high: 163.97,
-    close: 163.65,
-    volume: 832765,
-  },
-  {
-    date: '2022-01-21 12:30:00',
-    open: 163.73,
-    low: 163.48,
-    high: 163.84,
-    close: 163.6201,
-    volume: 758662,
-  },
-  {
-    date: '2022-01-21 12:25:00',
-    open: 164.02,
-    low: 163.6946,
-    high: 164.23,
-    close: 163.73,
-    volume: 828334,
-  },
-  {
-    date: '2022-01-21 12:20:00',
-    open: 163.79,
-    low: 163.52,
-    high: 164.04,
-    close: 164.04,
-    volume: 1163135,
-  },
-  {
-    date: '2022-01-21 12:15:00',
-    open: 163.91,
-    low: 163.68,
-    high: 164.15,
-    close: 163.8,
-    volume: 922942,
-  },
-  {
-    date: '2022-01-21 12:10:00',
-    open: 164.13,
-    low: 163.8,
-    high: 164.31,
-    close: 163.91,
-    volume: 1051596,
-  },
-  {
-    date: '2022-01-21 12:05:00',
-    open: 164.4,
-    low: 164.0901,
-    high: 164.63,
-    close: 164.16,
-    volume: 802163,
-  },
-  {
-    date: '2022-01-21 12:00:00',
-    open: 164.42,
-    low: 164.17,
-    high: 164.67,
-    close: 164.38,
-    volume: 826696,
-  },
-  {
-    date: '2022-01-21 11:55:00',
-    open: 164.22,
-    low: 164.15,
-    high: 164.54,
-    close: 164.44,
-    volume: 909647,
-  },
-  {
-    date: '2022-01-21 11:50:00',
-    open: 164.6,
-    low: 164.02,
-    high: 164.7,
-    close: 164.2088,
-    volume: 894350,
-  },
-  {
-    date: '2022-01-21 11:45:00',
-    open: 165.08,
-    low: 164.45,
-    high: 165.11,
-    close: 164.6,
-    volume: 1276147,
-  },
-  {
-    date: '2022-01-21 11:40:00',
-    open: 165.06,
-    low: 164.87,
-    high: 165.2182,
-    close: 165.05,
-    volume: 750005,
-  },
-  {
-    date: '2022-01-21 11:35:00',
-    open: 165.78,
-    low: 164.86,
-    high: 165.94,
-    close: 165.06,
-    volume: 1308744,
-  },
-  {
-    date: '2022-01-21 11:30:00',
-    open: 165.94,
-    low: 165.1428,
-    high: 165.95,
-    close: 165.78,
-    volume: 1371342,
-  },
-  {
-    date: '2022-01-21 11:25:00',
-    open: 166.07,
-    low: 165.78,
-    high: 166.1898,
-    close: 165.95,
-    volume: 963225,
-  },
-  {
-    date: '2022-01-21 11:20:00',
-    open: 165.7261,
-    low: 165.44,
-    high: 166.33,
-    close: 166.07,
-    volume: 1372225,
-  },
-  {
-    date: '2022-01-21 11:15:00',
-    open: 165.07,
-    low: 165.07,
-    high: 165.88,
-    close: 165.715,
-    volume: 1500335,
-  },
-  {
-    date: '2022-01-21 11:10:00',
-    open: 165.39,
-    low: 164.93,
-    high: 165.51,
-    close: 165.06,
-    volume: 2200959,
-  },
-  {
-    date: '2022-01-21 11:05:00',
-    open: 164.97,
-    low: 164.93,
-    high: 165.41,
-    close: 165.4,
-    volume: 1213898,
-  },
-  {
-    date: '2022-01-21 11:00:00',
-    open: 165.17,
-    low: 164.65,
-    high: 165.25,
-    close: 164.95,
-    volume: 1179103,
-  },
-  {
-    date: '2022-01-21 10:55:00',
-    open: 165.52,
-    low: 164.82,
-    high: 165.67,
-    close: 165.16,
-    volume: 1263572,
-  },
-  {
-    date: '2022-01-21 10:50:00',
-    open: 165.22,
-    low: 165.13,
-    high: 165.84,
-    close: 165.53,
-    volume: 1795511,
-  },
-  {
-    date: '2022-01-21 10:45:00',
-    open: 165.03,
-    low: 164.67,
-    high: 165.2897,
-    close: 165.2014,
-    volume: 1113790,
-  },
-  {
-    date: '2022-01-21 10:40:00',
-    open: 164.94,
-    low: 164.62,
-    high: 165.29,
-    close: 165.03,
-    volume: 1515370,
-  },
-  {
-    date: '2022-01-21 10:35:00',
-    open: 165.45,
-    low: 164.9,
-    high: 165.78,
-    close: 164.9399,
-    volume: 1762388,
-  },
-  {
-    date: '2022-01-21 10:30:00',
-    open: 164.31,
-    low: 164.13,
-    high: 165.54,
-    close: 165.46,
-    volume: 2130613,
-  },
-  {
-    date: '2022-01-21 10:25:00',
-    open: 163.67,
-    low: 163.54,
-    high: 164.6,
-    close: 164.31,
-    volume: 1752836,
-  },
-  {
-    date: '2022-01-21 10:20:00',
-    open: 163.3,
-    low: 162.96,
-    high: 163.8376,
-    close: 163.68,
-    volume: 2401416,
-  },
-  {
-    date: '2022-01-21 10:15:00',
-    open: 163.6165,
-    low: 163.11,
-    high: 163.82,
-    close: 163.28,
-    volume: 1613638,
-  },
-  {
-    date: '2022-01-21 10:10:00',
-    open: 163.9,
-    low: 163.3299,
-    high: 164.0,
-    close: 163.63,
-    volume: 1542727,
-  },
-  {
-    date: '2022-01-21 10:05:00',
-    open: 163.32,
-    low: 163.1604,
-    high: 164.06,
-    close: 163.9,
-    volume: 1951077,
-  },
-  {
-    date: '2022-01-21 10:00:00',
-    open: 163.54,
-    low: 163.22,
-    high: 163.95,
-    close: 163.3201,
-    volume: 1647151,
-  },
-  {
-    date: '2022-01-21 09:55:00',
-    open: 163.35,
-    low: 163.18,
-    high: 163.78,
-    close: 163.53,
-    volume: 1408586,
-  },
-  {
-    date: '2022-01-21 09:50:00',
-    open: 164.77,
-    low: 163.23,
-    high: 164.9,
-    close: 163.365,
-    volume: 2033414,
-  },
-  {
-    date: '2022-01-21 09:45:00',
-    open: 163.71,
-    low: 163.67,
-    high: 164.89,
-    close: 164.77,
-    volume: 1909257,
-  },
-  {
-    date: '2022-01-21 09:40:00',
-    open: 164.52,
-    low: 163.7,
-    high: 164.71,
-    close: 163.72,
-    volume: 3802988,
-  },
-  {
-    date: '2022-01-21 09:35:00',
-    open: 165.2501,
-    low: 164.42,
-    high: 165.6,
-    close: 164.51,
-    volume: 2253073,
-  },
-  {
-    date: '2022-01-21 09:30:00',
-    open: 164.415,
-    low: 164.36,
-    high: 166.0,
-    close: 165.26,
-    volume: 7362686,
-  },
-  {
-    date: '2022-01-20 16:00:00',
-    open: 164.51,
-    low: 163.93,
-    high: 167.4329,
-    close: 163.99,
-    volume: 1693026,
-  },
-  {
-    date: '2022-01-20 15:55:00',
-    open: 164.58,
-    low: 164.18,
-    high: 164.96,
-    close: 164.52,
-    volume: 4242438,
-  },
-  {
-    date: '2022-01-20 15:50:00',
-    open: 164.95,
-    low: 164.5,
-    high: 165.24,
-    close: 164.5918,
-    volume: 2780585,
-  },
-  {
-    date: '2022-01-20 15:45:00',
-    open: 165.11,
-    low: 164.9,
-    high: 165.48,
-    close: 164.9499,
-    volume: 1804633,
-  },
-  {
-    date: '2022-01-20 15:40:00',
-    open: 165.205,
-    low: 165.06,
-    high: 165.56,
-    close: 165.1,
-    volume: 1547931,
-  },
-  {
-    date: '2022-01-20 15:35:00',
-    open: 165.71,
-    low: 165.0,
-    high: 165.71,
-    close: 165.22,
-    volume: 2269633,
-  },
-  {
-    date: '2022-01-20 15:30:00',
-    open: 165.84,
-    low: 165.495,
-    high: 165.97,
-    close: 165.705,
-    volume: 1207993,
-  },
-  {
-    date: '2022-01-20 15:25:00',
-    open: 165.98,
-    low: 165.65,
-    high: 166.07,
-    close: 165.845,
-    volume: 1232471,
-  },
-  {
-    date: '2022-01-20 15:20:00',
-    open: 166.25,
-    low: 165.95,
-    high: 166.2599,
-    close: 165.97,
-    volume: 910737,
-  },
-  {
-    date: '2022-01-20 15:15:00',
-    open: 166.04,
-    low: 165.92,
-    high: 166.32,
-    close: 166.26,
-    volume: 1001245,
-  },
-  {
-    date: '2022-01-20 15:10:00',
-    open: 166.2,
-    low: 165.96,
-    high: 166.29,
-    close: 166.04,
-    volume: 1276014,
-  },
-  {
-    date: '2022-01-20 15:05:00',
-    open: 166.4,
-    low: 166.09,
-    high: 166.53,
-    close: 166.1999,
-    volume: 933259,
-  },
-  {
-    date: '2022-01-20 15:00:00',
-    open: 166.66,
-    low: 166.27,
-    high: 166.97,
-    close: 166.408,
-    volume: 1068984,
-  },
-  {
-    date: '2022-01-20 14:55:00',
-    open: 166.73,
-    low: 166.6,
-    high: 166.88,
-    close: 166.66,
-    volume: 749847,
-  },
-  {
-    date: '2022-01-20 14:50:00',
-    open: 166.63,
-    low: 166.55,
-    high: 166.78,
-    close: 166.73,
-    volume: 613845,
-  },
-  {
-    date: '2022-01-20 14:45:00',
-    open: 166.63,
-    low: 166.32,
-    high: 166.8,
-    close: 166.64,
-    volume: 904658,
-  },
-  {
-    date: '2022-01-20 14:40:00',
-    open: 166.86,
-    low: 166.59,
-    high: 166.86,
-    close: 166.61,
-    volume: 571688,
-  },
-  {
-    date: '2022-01-20 14:35:00',
-    open: 166.72,
-    low: 166.61,
-    high: 166.93,
-    close: 166.865,
-    volume: 907051,
-  },
-  {
-    date: '2022-01-20 14:30:00',
-    open: 166.6474,
-    low: 166.42,
-    high: 166.7599,
-    close: 166.719,
-    volume: 853708,
-  },
-  {
-    date: '2022-01-20 14:25:00',
-    open: 166.8921,
-    low: 166.6,
-    high: 166.9735,
-    close: 166.6401,
-    volume: 727879,
-  },
-  {
-    date: '2022-01-20 14:20:00',
-    open: 166.98,
-    low: 166.71,
-    high: 167.0999,
-    close: 166.9099,
-    volume: 816666,
-  },
-  {
-    date: '2022-01-20 14:15:00',
-    open: 166.775,
-    low: 166.46,
-    high: 167.13,
-    close: 166.9701,
-    volume: 1293989,
-  },
-  {
-    date: '2022-01-20 14:10:00',
-    open: 167.64,
-    low: 166.64,
-    high: 167.74,
-    close: 166.77,
-    volume: 1153934,
-  },
-  {
-    date: '2022-01-20 14:05:00',
-    open: 167.8599,
-    low: 167.6,
-    high: 167.88,
-    close: 167.6399,
-    volume: 435383,
-  },
-  {
-    date: '2022-01-20 14:00:00',
-    open: 167.4388,
-    low: 167.39,
-    high: 167.91,
-    close: 167.8429,
-    volume: 682593,
-  },
-  {
-    date: '2022-01-20 13:55:00',
-    open: 167.7181,
-    low: 167.31,
-    high: 167.7287,
-    close: 167.4167,
-    volume: 649184,
-  },
-  {
-    date: '2022-01-20 13:50:00',
-    open: 167.59,
-    low: 167.53,
-    high: 167.77,
-    close: 167.7101,
-    volume: 440474,
-  },
-  {
-    date: '2022-01-20 13:45:00',
-    open: 167.6075,
-    low: 167.31,
-    high: 167.75,
-    close: 167.6,
-    volume: 534014,
-  },
-  {
-    date: '2022-01-20 13:40:00',
-    open: 167.5899,
-    low: 167.51,
-    high: 167.8,
-    close: 167.6,
-    volume: 602757,
-  },
-  {
-    date: '2022-01-20 13:35:00',
-    open: 167.2938,
-    low: 167.25,
-    high: 167.66,
-    close: 167.58,
-    volume: 717176,
-  },
-  {
-    date: '2022-01-20 13:30:00',
-    open: 167.61,
-    low: 167.15,
-    high: 167.7,
-    close: 167.3,
-    volume: 1115739,
-  },
-  {
-    date: '2022-01-20 13:25:00',
-    open: 167.73,
-    low: 167.6,
-    high: 167.94,
-    close: 167.6,
-    volume: 556454,
-  },
-  {
-    date: '2022-01-20 13:20:00',
-    open: 167.88,
-    low: 167.68,
-    high: 168.02,
-    close: 167.7204,
-    volume: 509774,
-  },
-  {
-    date: '2022-01-20 13:15:00',
-    open: 168.07,
-    low: 167.69,
-    high: 168.07,
-    close: 167.89,
-    volume: 531857,
-  },
-  {
-    date: '2022-01-20 13:10:00',
-    open: 168.025,
-    low: 167.86,
-    high: 168.15,
-    close: 168.07,
-    volume: 862052,
-  },
-  {
-    date: '2022-01-20 13:05:00',
-    open: 168.13,
-    low: 167.76,
-    high: 168.13,
-    close: 168.02,
-    volume: 962708,
-  },
-  {
-    date: '2022-01-20 13:00:00',
-    open: 168.64,
-    low: 168.06,
-    high: 168.64,
-    close: 168.14,
-    volume: 718480,
-  },
-  {
-    date: '2022-01-20 12:55:00',
-    open: 168.6397,
-    low: 168.495,
-    high: 168.73,
-    close: 168.645,
-    volume: 593316,
-  },
-  {
-    date: '2022-01-20 12:50:00',
-    open: 169.03,
-    low: 168.59,
-    high: 169.04,
-    close: 168.63,
-    volume: 665207,
-  },
-  {
-    date: '2022-01-20 12:45:00',
-    open: 169.185,
-    low: 168.9501,
-    high: 169.22,
-    close: 169.02,
-    volume: 357964,
-  },
-  {
-    date: '2022-01-20 12:40:00',
-    open: 168.895,
-    low: 168.85,
-    high: 169.22,
-    close: 169.1801,
-    volume: 1159387,
-  },
-  {
-    date: '2022-01-20 12:35:00',
-    open: 168.94,
-    low: 168.82,
-    high: 169.0299,
-    close: 168.895,
-    volume: 359459,
-  },
-  {
-    date: '2022-01-20 12:30:00',
-    open: 169.09,
-    low: 168.895,
-    high: 169.2,
-    close: 168.93,
-    volume: 481016,
-  },
-  {
-    date: '2022-01-20 12:25:00',
-    open: 169.02,
-    low: 168.81,
-    high: 169.1,
-    close: 169.075,
-    volume: 1502074,
-  },
-  {
-    date: '2022-01-20 12:20:00',
-    open: 168.94,
-    low: 166.85,
-    high: 169.06,
-    close: 169.02,
-    volume: 425870,
-  },
-  {
-    date: '2022-01-20 12:15:00',
-    open: 169.11,
-    low: 168.94,
-    high: 169.13,
-    close: 168.95,
-    volume: 373478,
-  },
-  {
-    date: '2022-01-20 12:10:00',
-    open: 169.0,
-    low: 168.935,
-    high: 169.21,
-    close: 169.1099,
-    volume: 483782,
-  },
-  {
-    date: '2022-01-20 12:05:00',
-    open: 169.149,
-    low: 168.92,
-    high: 169.15,
-    close: 169.01,
-    volume: 737036,
-  },
-  {
-    date: '2022-01-20 12:00:00',
-    open: 169.11,
-    low: 169.1,
-    high: 169.25,
-    close: 169.1399,
-    volume: 363455,
-  },
-  {
-    date: '2022-01-20 11:55:00',
-    open: 169.11,
-    low: 169.07,
-    high: 169.32,
-    close: 169.11,
-    volume: 1690310,
-  },
-  {
-    date: '2022-01-20 11:50:00',
-    open: 168.98,
-    low: 168.97,
-    high: 169.29,
-    close: 169.1,
-    volume: 559553,
-  },
-  {
-    date: '2022-01-20 11:45:00',
-    open: 168.78,
-    low: 168.78,
-    high: 169.05,
-    close: 168.99,
-    volume: 672407,
-  },
-  {
-    date: '2022-01-20 11:40:00',
-    open: 169.325,
-    low: 168.7893,
-    high: 169.3788,
-    close: 168.79,
-    volume: 915476,
-  },
-  {
-    date: '2022-01-20 11:35:00',
-    open: 169.42,
-    low: 169.1218,
-    high: 169.48,
-    close: 169.33,
-    volume: 585187,
-  },
-  {
-    date: '2022-01-20 11:30:00',
-    open: 169.36,
-    low: 169.28,
-    high: 169.46,
-    close: 169.42,
-    volume: 576425,
-  },
-  {
-    date: '2022-01-20 11:25:00',
-    open: 169.07,
-    low: 169.02,
-    high: 169.43,
-    close: 169.35,
-    volume: 591248,
-  },
-  {
-    date: '2022-01-20 11:20:00',
-    open: 169.215,
-    low: 168.98,
-    high: 169.215,
-    close: 169.075,
-    volume: 667630,
-  },
-  {
-    date: '2022-01-20 11:15:00',
-    open: 169.13,
-    low: 168.96,
-    high: 169.31,
-    close: 169.21,
-    volume: 743612,
-  },
-  {
-    date: '2022-01-20 11:10:00',
-    open: 169.2,
-    low: 168.96,
-    high: 169.22,
-    close: 169.13,
-    volume: 693010,
-  },
-  {
-    date: '2022-01-20 11:05:00',
-    open: 169.312,
-    low: 169.15,
-    high: 169.43,
-    close: 169.18,
-    volume: 727346,
-  },
-  {
-    date: '2022-01-20 11:00:00',
-    open: 169.23,
-    low: 169.18,
-    high: 169.5498,
-    close: 169.32,
-    volume: 767871,
-  },
-  {
-    date: '2022-01-20 10:55:00',
-    open: 169.33,
-    low: 169.04,
-    high: 169.35,
-    close: 169.229,
-    volume: 630325,
-  },
-  {
-    date: '2022-01-20 10:50:00',
-    open: 169.2,
-    low: 169.0281,
-    high: 169.45,
-    close: 169.339,
-    volume: 782358,
-  },
-  {
-    date: '2022-01-20 10:45:00',
-    open: 168.9961,
-    low: 168.94,
-    high: 169.33,
-    close: 169.195,
-    volume: 908414,
-  },
-  {
-    date: '2022-01-20 10:40:00',
-    open: 169.17,
-    low: 168.98,
-    high: 169.305,
-    close: 169.01,
-    volume: 726188,
-  },
-  {
-    date: '2022-01-20 10:35:00',
-    open: 169.38,
-    low: 169.12,
-    high: 169.54,
-    close: 169.165,
-    volume: 852422,
-  },
-  {
-    date: '2022-01-20 10:30:00',
-    open: 169.2399,
-    low: 169.239,
-    high: 169.68,
-    close: 169.3655,
-    volume: 1161920,
-  },
-  {
-    date: '2022-01-20 10:25:00',
-    open: 168.975,
-    low: 168.73,
-    high: 169.34,
-    close: 169.23,
-    volume: 1124190,
-  },
-  {
-    date: '2022-01-20 10:20:00',
-    open: 168.6511,
-    low: 168.61,
-    high: 168.99,
-    close: 168.98,
-    volume: 889660,
-  },
-  {
-    date: '2022-01-20 10:15:00',
-    open: 168.75,
-    low: 168.56,
-    high: 168.91,
-    close: 168.64,
-    volume: 1071189,
-  },
-  {
-    date: '2022-01-20 10:10:00',
-    open: 168.75,
-    low: 168.5522,
-    high: 168.9,
-    close: 168.75,
-    volume: 993577,
-  },
-  {
-    date: '2022-01-20 10:05:00',
-    open: 168.66,
-    low: 168.47,
-    high: 168.84,
-    close: 168.75,
-    volume: 1164082,
-  },
-  {
-    date: '2022-01-20 10:00:00',
-    open: 168.87,
-    low: 168.61,
-    high: 169.06,
-    close: 168.656,
-    volume: 1319335,
-  },
-  {
-    date: '2022-01-20 09:55:00',
-    open: 168.41,
-    low: 168.26,
-    high: 168.92,
-    close: 168.88,
-    volume: 1373398,
-  },
-  {
-    date: '2022-01-20 09:50:00',
-    open: 168.41,
-    low: 168.225,
-    high: 168.595,
-    close: 168.39,
-    volume: 1519216,
-  },
-  {
-    date: '2022-01-20 09:45:00',
-    open: 168.48,
-    low: 168.185,
-    high: 168.67,
-    close: 168.42,
-    volume: 1656337,
-  },
-  {
-    date: '2022-01-20 09:40:00',
-    open: 167.35,
-    low: 167.2899,
-    high: 168.55,
-    close: 168.5,
-    volume: 1884411,
-  },
-  {
-    date: '2022-01-20 09:35:00',
-    open: 167.48,
-    low: 166.9,
-    high: 167.6298,
-    close: 167.345,
-    volume: 1543085,
-  },
-  {
-    date: '2022-01-20 09:30:00',
-    open: 166.98,
-    low: 166.76,
-    high: 167.74,
-    close: 167.5017,
-    volume: 3034473,
-  },
-  {
-    date: '2022-01-19 16:00:00',
-    open: 166.23,
-    low: 165.9,
-    high: 166.23,
-    close: 165.93,
-    volume: 2650254,
-  },
-  {
-    date: '2022-01-19 15:55:00',
-    open: 166.59,
-    low: 165.94,
-    high: 166.72,
-    close: 166.21,
-    volume: 3900789,
-  },
-  {
-    date: '2022-01-19 15:50:00',
-    open: 166.86,
-    low: 166.48,
-    high: 166.87,
-    close: 166.59,
-    volume: 2683419,
-  },
-  {
-    date: '2022-01-19 15:45:00',
-    open: 167.2,
-    low: 166.86,
-    high: 167.2742,
-    close: 166.87,
-    volume: 2035440,
-  },
-  {
-    date: '2022-01-19 15:40:00',
-    open: 167.2,
-    low: 167.18,
-    high: 167.39,
-    close: 167.21,
-    volume: 1139956,
-  },
-  {
-    date: '2022-01-19 15:35:00',
-    open: 167.34,
-    low: 167.18,
-    high: 167.36,
-    close: 167.1928,
-    volume: 1083892,
-  },
-  {
-    date: '2022-01-19 15:30:00',
-    open: 167.51,
-    low: 167.28,
-    high: 167.675,
-    close: 167.35,
-    volume: 1111815,
-  },
-  {
-    date: '2022-01-19 15:25:00',
-    open: 167.7345,
-    low: 167.49,
-    high: 167.89,
-    close: 167.51,
-    volume: 1165017,
-  },
-  {
-    date: '2022-01-19 15:20:00',
-    open: 167.84,
-    low: 167.69,
-    high: 168.05,
-    close: 167.73,
-    volume: 694522,
-  },
-  {
-    date: '2022-01-19 15:15:00',
-    open: 168.155,
-    low: 167.75,
-    high: 168.225,
-    close: 167.85,
-    volume: 690920,
-  },
-  {
-    date: '2022-01-19 15:10:00',
-    open: 167.86,
-    low: 167.79,
-    high: 168.2699,
-    close: 168.1599,
-    volume: 800848,
-  },
-  {
-    date: '2022-01-19 15:05:00',
-    open: 167.79,
-    low: 167.66,
-    high: 167.97,
-    close: 167.86,
-    volume: 789264,
-  },
-  {
-    date: '2022-01-19 15:00:00',
-    open: 167.95,
-    low: 167.67,
-    high: 168.065,
-    close: 167.79,
-    volume: 850513,
-  },
-  {
-    date: '2022-01-19 14:55:00',
-    open: 168.05,
-    low: 167.92,
-    high: 168.24,
-    close: 167.95,
-    volume: 724762,
-  },
-  {
-    date: '2022-01-19 14:50:00',
-    open: 167.7551,
-    low: 167.7551,
-    high: 168.07,
-    close: 168.05,
-    volume: 651995,
-  },
-  {
-    date: '2022-01-19 14:45:00',
-    open: 167.98,
-    low: 167.69,
-    high: 168.0,
-    close: 167.76,
-    volume: 770116,
-  },
-  {
-    date: '2022-01-19 14:40:00',
-    open: 167.695,
-    low: 167.61,
-    high: 168.04,
-    close: 167.98,
-    volume: 853176,
-  },
-  {
-    date: '2022-01-19 14:35:00',
-    open: 167.76,
-    low: 167.65,
-    high: 167.96,
-    close: 167.6955,
-    volume: 769848,
-  },
-  {
-    date: '2022-01-19 14:30:00',
-    open: 167.645,
-    low: 167.54,
-    high: 167.84,
-    close: 167.76,
-    volume: 847311,
-  },
-  {
-    date: '2022-01-19 14:25:00',
-    open: 168.08,
-    low: 167.62,
-    high: 168.0951,
-    close: 167.65,
-    volume: 912682,
-  },
-  {
-    date: '2022-01-19 14:20:00',
-    open: 168.06,
-    low: 167.815,
-    high: 168.1499,
-    close: 168.094,
-    volume: 686981,
-  },
-  {
-    date: '2022-01-19 14:15:00',
-    open: 167.99,
-    low: 167.86,
-    high: 168.105,
-    close: 168.0638,
-    volume: 629779,
-  },
-  {
-    date: '2022-01-19 14:10:00',
-    open: 168.13,
-    low: 167.94,
-    high: 168.25,
-    close: 167.995,
-    volume: 781192,
-  },
-  {
-    date: '2022-01-19 14:05:00',
-    open: 168.43,
-    low: 167.94,
-    high: 168.44,
-    close: 168.13,
-    volume: 1009872,
-  },
-  {
-    date: '2022-01-19 14:00:00',
-    open: 168.44,
-    low: 168.36,
-    high: 168.66,
-    close: 168.43,
-    volume: 691926,
-  },
-  {
-    date: '2022-01-19 13:55:00',
-    open: 168.62,
-    low: 168.4333,
-    high: 168.78,
-    close: 168.45,
-    volume: 531511,
-  },
-  {
-    date: '2022-01-19 13:50:00',
-    open: 168.53,
-    low: 168.37,
-    high: 168.6699,
-    close: 168.6199,
-    volume: 679346,
-  },
-  {
-    date: '2022-01-19 13:45:00',
-    open: 168.66,
-    low: 168.5,
-    high: 168.7,
-    close: 168.53,
-    volume: 524912,
-  },
-  {
-    date: '2022-01-19 13:40:00',
-    open: 168.6985,
-    low: 168.5,
-    high: 168.72,
-    close: 168.6511,
-    volume: 545286,
-  },
-  {
-    date: '2022-01-19 13:35:00',
-    open: 168.87,
-    low: 168.59,
-    high: 168.87,
-    close: 168.7,
-    volume: 570100,
-  },
-  {
-    date: '2022-01-19 13:30:00',
-    open: 168.81,
-    low: 168.67,
-    high: 169.03,
-    close: 168.87,
-    volume: 607161,
-  },
-  {
-    date: '2022-01-19 13:25:00',
-    open: 169.19,
-    low: 168.79,
-    high: 169.26,
-    close: 168.8103,
-    volume: 487033,
-  },
-  {
-    date: '2022-01-19 13:20:00',
-    open: 169.22,
-    low: 169.01,
-    high: 169.35,
-    close: 169.185,
-    volume: 543957,
-  },
-  {
-    date: '2022-01-19 13:15:00',
-    open: 169.23,
-    low: 169.14,
-    high: 169.45,
-    close: 169.22,
-    volume: 1539638,
-  },
-  {
-    date: '2022-01-19 13:10:00',
-    open: 169.01,
-    low: 169.01,
-    high: 169.29,
-    close: 169.23,
-    volume: 647487,
-  },
-  {
-    date: '2022-01-19 13:05:00',
-    open: 169.1,
-    low: 168.91,
-    high: 169.1668,
-    close: 169.0,
-    volume: 606960,
-  },
-  {
-    date: '2022-01-19 13:00:00',
-    open: 169.08,
-    low: 168.76,
-    high: 169.24,
-    close: 169.1,
-    volume: 704566,
-  },
-  {
-    date: '2022-01-19 12:55:00',
-    open: 168.66,
-    low: 168.66,
-    high: 169.13,
-    close: 169.08,
-    volume: 560451,
-  },
-  {
-    date: '2022-01-19 12:50:00',
-    open: 168.97,
-    low: 168.63,
-    high: 169.1,
-    close: 168.658,
-    volume: 1013004,
-  },
-  {
-    date: '2022-01-19 12:45:00',
-    open: 168.94,
-    low: 168.85,
-    high: 169.2892,
-    close: 168.9689,
-    volume: 642158,
-  },
-  {
-    date: '2022-01-19 12:40:00',
-    open: 168.95,
-    low: 168.855,
-    high: 169.07,
-    close: 168.96,
-    volume: 482438,
-  },
-  {
-    date: '2022-01-19 12:35:00',
-    open: 168.9599,
-    low: 168.9,
-    high: 169.17,
-    close: 168.9401,
-    volume: 550259,
-  },
-  {
-    date: '2022-01-19 12:30:00',
-    open: 169.065,
-    low: 168.9,
-    high: 169.2,
-    close: 168.96,
-    volume: 604682,
-  },
-  {
-    date: '2022-01-19 12:25:00',
-    open: 169.03,
-    low: 168.91,
-    high: 169.1245,
-    close: 169.04,
-    volume: 527622,
-  },
-  {
-    date: '2022-01-19 12:20:00',
-    open: 168.85,
-    low: 168.7901,
-    high: 169.11,
-    close: 169.035,
-    volume: 753031,
-  },
-  {
-    date: '2022-01-19 12:15:00',
-    open: 169.07,
-    low: 168.7236,
-    high: 169.09,
-    close: 168.855,
-    volume: 760001,
-  },
-  {
-    date: '2022-01-19 12:10:00',
-    open: 168.825,
-    low: 168.799,
-    high: 169.62,
-    close: 169.08,
-    volume: 1232031,
-  },
-  {
-    date: '2022-01-19 12:05:00',
-    open: 168.565,
-    low: 168.515,
-    high: 168.87,
-    close: 168.8161,
-    volume: 653248,
-  },
-  {
-    date: '2022-01-19 12:00:00',
-    open: 168.33,
-    low: 168.33,
-    high: 168.67,
-    close: 168.5609,
-    volume: 863196,
-  },
-  {
-    date: '2022-01-19 11:55:00',
-    open: 168.2183,
-    low: 168.14,
-    high: 168.39,
-    close: 168.33,
-    volume: 711466,
-  },
-  {
-    date: '2022-01-19 11:50:00',
-    open: 167.77,
-    low: 167.69,
-    high: 168.29,
-    close: 168.215,
-    volume: 1041154,
-  },
-  {
-    date: '2022-01-19 11:45:00',
-    open: 167.44,
-    low: 167.4,
-    high: 167.8897,
-    close: 167.7649,
-    volume: 1059934,
-  },
-  {
-    date: '2022-01-19 11:40:00',
-    open: 167.75,
-    low: 167.32,
-    high: 167.76,
-    close: 167.43,
-    volume: 1201586,
-  },
-  {
-    date: '2022-01-19 11:35:00',
-    open: 167.69,
-    low: 167.52,
-    high: 167.87,
-    close: 167.75,
-    volume: 918295,
-  },
-  {
-    date: '2022-01-19 11:30:00',
-    open: 167.7917,
-    low: 167.62,
-    high: 168.18,
-    close: 167.69,
-    volume: 1243798,
-  },
-  {
-    date: '2022-01-19 11:25:00',
-    open: 167.76,
-    low: 167.58,
-    high: 168.0299,
-    close: 167.78,
-    volume: 1742482,
-  },
-  {
-    date: '2022-01-19 11:20:00',
-    open: 168.11,
-    low: 167.54,
-    high: 168.16,
-    close: 167.77,
-    volume: 2055875,
-  },
-  {
-    date: '2022-01-19 11:15:00',
-    open: 168.28,
-    low: 168.05,
-    high: 168.5,
-    close: 168.1022,
-    volume: 1208945,
-  },
-  {
-    date: '2022-01-19 11:10:00',
-    open: 168.86,
-    low: 168.25,
-    high: 168.86,
-    close: 168.2701,
-    volume: 1278690,
-  },
-  {
-    date: '2022-01-19 11:05:00',
-    open: 168.93,
-    low: 168.69,
-    high: 168.985,
-    close: 168.865,
-    volume: 828211,
-  },
-  {
-    date: '2022-01-19 11:00:00',
-    open: 168.9,
-    low: 168.87,
-    high: 169.2485,
-    close: 168.9389,
-    volume: 1078375,
-  },
-  {
-    date: '2022-01-19 10:55:00',
-    open: 168.81,
-    low: 168.661,
-    high: 169.0687,
-    close: 168.895,
-    volume: 1084972,
-  },
-  {
-    date: '2022-01-19 10:50:00',
-    open: 168.71,
-    low: 168.66,
-    high: 168.93,
-    close: 168.815,
-    volume: 1157775,
-  },
-  {
-    date: '2022-01-19 10:45:00',
-    open: 169.13,
-    low: 168.64,
-    high: 169.26,
-    close: 168.7,
-    volume: 1190838,
-  },
-  {
-    date: '2022-01-19 10:40:00',
-    open: 169.289,
-    low: 168.92,
-    high: 169.29,
-    close: 169.13,
-    volume: 1342890,
-  },
-  {
-    date: '2022-01-19 10:35:00',
-    open: 169.82,
-    low: 169.23,
-    high: 169.83,
-    close: 169.28,
-    volume: 1009177,
-  },
-  {
-    date: '2022-01-19 10:30:00',
-    open: 169.89,
-    low: 169.63,
-    high: 170.04,
-    close: 169.8279,
-    volume: 882216,
-  },
-  {
-    date: '2022-01-19 10:25:00',
-    open: 170.05,
-    low: 169.79,
-    high: 170.1,
-    close: 169.87,
-    volume: 756984,
-  },
-  {
-    date: '2022-01-19 10:20:00',
-    open: 170.31,
-    low: 169.8919,
-    high: 170.35,
-    close: 170.0501,
-    volume: 983106,
-  },
-  {
-    date: '2022-01-19 10:15:00',
-    open: 170.98,
-    low: 170.14,
-    high: 171.01,
-    close: 170.31,
-    volume: 1169892,
-  },
-  {
-    date: '2022-01-19 10:10:00',
-    open: 170.72,
-    low: 170.42,
-    high: 171.08,
-    close: 170.9799,
-    volume: 1421334,
-  },
-  {
-    date: '2022-01-19 10:05:00',
-    open: 170.78,
-    low: 170.48,
-    high: 170.92,
-    close: 170.7118,
-    volume: 1124974,
-  },
-  {
-    date: '2022-01-19 10:00:00',
-    open: 170.95,
-    low: 170.52,
-    high: 170.95,
-    close: 170.77,
-    volume: 1236235,
-  },
-  {
-    date: '2022-01-19 09:55:00',
-    open: 170.37,
-    low: 170.31,
-    high: 170.95,
-    close: 170.9315,
-    volume: 1215564,
-  },
-  {
-    date: '2022-01-19 09:50:00',
-    open: 170.3,
-    low: 170.14,
-    high: 170.7399,
-    close: 170.36,
-    volume: 1103963,
-  },
-  {
-    date: '2022-01-19 09:45:00',
-    open: 170.0235,
-    low: 169.95,
-    high: 170.69,
-    close: 170.29,
-    volume: 1524365,
-  },
-  {
-    date: '2022-01-19 09:40:00',
-    open: 170.25,
-    low: 169.56,
-    high: 170.34,
-    close: 170.02,
-    volume: 1693100,
-  },
-  {
-    date: '2022-01-19 09:35:00',
-    open: 169.46,
-    low: 169.28,
-    high: 170.56,
-    close: 170.2301,
-    volume: 1783219,
-  },
-  {
-    date: '2022-01-19 09:30:00',
-    open: 170.0,
-    low: 169.29,
-    high: 170.98,
-    close: 169.4694,
-    volume: 2965328,
-  },
-  {
-    date: '2022-01-18 16:00:00',
-    open: 169.8,
-    low: 169.73,
-    high: 170.0737,
-    close: 169.8,
-    volume: 4636399,
-  },
-  {
-    date: '2022-01-18 15:55:00',
-    open: 169.5,
-    low: 169.405,
-    high: 170.09,
-    close: 169.83,
-    volume: 4447449,
-  },
-  {
-    date: '2022-01-18 15:50:00',
-    open: 170.25,
-    low: 169.45,
-    high: 170.45,
-    close: 169.49,
-    volume: 2210533,
-  },
-  {
-    date: '2022-01-18 15:45:00',
-    open: 169.86,
-    low: 169.83,
-    high: 170.2483,
-    close: 170.2483,
-    volume: 1090935,
-  },
-  {
-    date: '2022-01-18 15:40:00',
-    open: 170.175,
-    low: 169.82,
-    high: 170.25,
-    close: 169.85,
-    volume: 913605,
-  },
-  {
-    date: '2022-01-18 15:35:00',
-    open: 170.24,
-    low: 170.08,
-    high: 170.3,
-    close: 170.175,
-    volume: 888633,
-  },
-  {
-    date: '2022-01-18 15:30:00',
-    open: 170.34,
-    low: 170.16,
-    high: 170.52,
-    close: 170.24,
-    volume: 923270,
-  },
-  {
-    date: '2022-01-18 15:25:00',
-    open: 170.22,
-    low: 170.05,
-    high: 170.38,
-    close: 170.34,
-    volume: 670850,
-  },
-  {
-    date: '2022-01-18 15:20:00',
-    open: 170.0341,
-    low: 169.98,
-    high: 170.31,
-    close: 170.2213,
-    volume: 715710,
-  },
-  {
-    date: '2022-01-18 15:15:00',
-    open: 170.4599,
-    low: 169.84,
-    high: 170.4599,
-    close: 170.03,
-    volume: 1150826,
-  },
-  {
-    date: '2022-01-18 15:10:00',
-    open: 170.5557,
-    low: 170.42,
-    high: 170.66,
-    close: 170.4599,
-    volume: 523257,
-  },
-  {
-    date: '2022-01-18 15:05:00',
-    open: 170.57,
-    low: 170.35,
-    high: 170.68,
-    close: 170.55,
-    volume: 577398,
-  },
-  {
-    date: '2022-01-18 15:00:00',
-    open: 170.5701,
-    low: 170.41,
-    high: 170.77,
-    close: 170.5666,
-    volume: 799279,
-  },
-  {
-    date: '2022-01-18 14:55:00',
-    open: 170.655,
-    low: 170.52,
-    high: 170.79,
-    close: 170.57,
-    volume: 699584,
-  },
-  {
-    date: '2022-01-18 14:50:00',
-    open: 170.52,
-    low: 170.5,
-    high: 170.72,
-    close: 170.66,
-    volume: 621478,
-  },
-  {
-    date: '2022-01-18 14:45:00',
-    open: 170.27,
-    low: 170.21,
-    high: 170.56,
-    close: 170.52,
-    volume: 566150,
-  },
-  {
-    date: '2022-01-18 14:40:00',
-    open: 170.4,
-    low: 170.23,
-    high: 170.43,
-    close: 170.27,
-    volume: 489615,
-  },
-  {
-    date: '2022-01-18 14:35:00',
-    open: 170.32,
-    low: 170.255,
-    high: 170.52,
-    close: 170.395,
-    volume: 485763,
-  },
-  {
-    date: '2022-01-18 14:30:00',
-    open: 170.15,
-    low: 170.14,
-    high: 170.59,
-    close: 170.32,
-    volume: 758337,
-  },
-  {
-    date: '2022-01-18 14:25:00',
-    open: 170.17,
-    low: 170.01,
-    high: 170.24,
-    close: 170.16,
-    volume: 436856,
-  },
-  {
-    date: '2022-01-18 14:20:00',
-    open: 170.525,
-    low: 170.07,
-    high: 170.525,
-    close: 170.16,
-    volume: 759554,
-  },
-  {
-    date: '2022-01-18 14:15:00',
-    open: 170.152,
-    low: 170.1248,
-    high: 170.531,
-    close: 170.52,
-    volume: 532493,
-  },
-  {
-    date: '2022-01-18 14:10:00',
-    open: 170.24,
-    low: 170.14,
-    high: 170.37,
-    close: 170.14,
-    volume: 442158,
-  },
-  {
-    date: '2022-01-18 14:05:00',
-    open: 170.35,
-    low: 170.14,
-    high: 170.45,
-    close: 170.24,
-    volume: 594741,
-  },
-  {
-    date: '2022-01-18 14:00:00',
-    open: 170.33,
-    low: 170.3,
-    high: 170.53,
-    close: 170.35,
-    volume: 669650,
-  },
-  {
-    date: '2022-01-18 13:55:00',
-    open: 170.0701,
-    low: 170.0526,
-    high: 170.44,
-    close: 170.345,
-    volume: 486767,
-  },
-  {
-    date: '2022-01-18 13:50:00',
-    open: 170.08,
-    low: 170.015,
-    high: 170.28,
-    close: 170.09,
-    volume: 517767,
-  },
-  {
-    date: '2022-01-18 13:45:00',
-    open: 170.22,
-    low: 170.0,
-    high: 170.36,
-    close: 170.08,
-    volume: 654971,
-  },
-  {
-    date: '2022-01-18 13:40:00',
-    open: 170.18,
-    low: 170.1,
-    high: 170.33,
-    close: 170.21,
-    volume: 607258,
-  },
-  {
-    date: '2022-01-18 13:35:00',
-    open: 169.91,
-    low: 169.88,
-    high: 170.2,
-    close: 170.1799,
-    volume: 746979,
-  },
-  {
-    date: '2022-01-18 13:30:00',
-    open: 169.75,
-    low: 169.685,
-    high: 169.94,
-    close: 169.91,
-    volume: 2155250,
-  },
-  {
-    date: '2022-01-18 13:25:00',
-    open: 169.8,
-    low: 169.63,
-    high: 169.82,
-    close: 169.755,
-    volume: 870094,
-  },
-  {
-    date: '2022-01-18 13:20:00',
-    open: 170.08,
-    low: 169.69,
-    high: 170.09,
-    close: 169.79,
-    volume: 584759,
-  },
-  {
-    date: '2022-01-18 13:15:00',
-    open: 170.065,
-    low: 169.9306,
-    high: 170.15,
-    close: 170.08,
-    volume: 659204,
-  },
-  {
-    date: '2022-01-18 13:10:00',
-    open: 169.8039,
-    low: 169.74,
-    high: 170.13,
-    close: 170.07,
-    volume: 741896,
-  },
-  {
-    date: '2022-01-18 13:05:00',
-    open: 169.73,
-    low: 169.6,
-    high: 169.85,
-    close: 169.815,
-    volume: 541942,
-  },
-  {
-    date: '2022-01-18 13:00:00',
-    open: 169.75,
-    low: 169.57,
-    high: 169.815,
-    close: 169.7312,
-    volume: 784417,
-  },
-  {
-    date: '2022-01-18 12:55:00',
-    open: 169.77,
-    low: 169.6,
-    high: 169.87,
-    close: 169.76,
-    volume: 634450,
-  },
-  {
-    date: '2022-01-18 12:50:00',
-    open: 169.87,
-    low: 169.7,
-    high: 169.96,
-    close: 169.765,
-    volume: 540053,
-  },
-  {
-    date: '2022-01-18 12:45:00',
-    open: 169.88,
-    low: 169.71,
-    high: 169.9999,
-    close: 169.86,
-    volume: 637964,
-  },
-  {
-    date: '2022-01-18 12:40:00',
-    open: 169.85,
-    low: 169.84,
-    high: 170.11,
-    close: 169.89,
-    volume: 663153,
-  },
-  {
-    date: '2022-01-18 12:35:00',
-    open: 170.0,
-    low: 169.78,
-    high: 170.0,
-    close: 169.855,
-    volume: 456211,
-  },
-  {
-    date: '2022-01-18 12:30:00',
-    open: 169.995,
-    low: 169.74,
-    high: 170.0406,
-    close: 170.0,
-    volume: 717228,
-  },
-  {
-    date: '2022-01-18 12:25:00',
-    open: 169.9101,
-    low: 169.78,
-    high: 170.01,
-    close: 170.0,
-    volume: 600045,
-  },
-  {
-    date: '2022-01-18 12:20:00',
-    open: 170.16,
-    low: 169.81,
-    high: 170.21,
-    close: 169.91,
-    volume: 1234460,
-  },
-  {
-    date: '2022-01-18 12:15:00',
-    open: 170.33,
-    low: 170.12,
-    high: 170.34,
-    close: 170.155,
-    volume: 690857,
-  },
-  {
-    date: '2022-01-18 12:10:00',
-    open: 170.405,
-    low: 170.14,
-    high: 170.47,
-    close: 170.34,
-    volume: 707865,
-  },
-  {
-    date: '2022-01-18 12:05:00',
-    open: 170.67,
-    low: 170.3803,
-    high: 170.67,
-    close: 170.41,
-    volume: 739727,
-  },
-  {
-    date: '2022-01-18 12:00:00',
-    open: 171.14,
-    low: 170.63,
-    high: 171.28,
-    close: 170.674,
-    volume: 743704,
-  },
-  {
-    date: '2022-01-18 11:55:00',
-    open: 171.12,
-    low: 171.0,
-    high: 171.215,
-    close: 171.1244,
-    volume: 521134,
-  },
-  {
-    date: '2022-01-18 11:50:00',
-    open: 170.97,
-    low: 170.94,
-    high: 171.22,
-    close: 171.13,
-    volume: 732806,
-  },
-  {
-    date: '2022-01-18 11:45:00',
-    open: 170.59,
-    low: 170.58,
-    high: 170.9899,
-    close: 170.97,
-    volume: 664562,
-  },
-  {
-    date: '2022-01-18 11:40:00',
-    open: 170.6286,
-    low: 170.56,
-    high: 170.78,
-    close: 170.6,
-    volume: 576399,
-  },
-  {
-    date: '2022-01-18 11:35:00',
-    open: 170.33,
-    low: 170.27,
-    high: 170.69,
-    close: 170.63,
-    volume: 1102364,
-  },
-  {
-    date: '2022-01-18 11:30:00',
-    open: 170.56,
-    low: 170.2,
-    high: 170.58,
-    close: 170.305,
-    volume: 932364,
-  },
-  {
-    date: '2022-01-18 11:25:00',
-    open: 170.66,
-    low: 170.5,
-    high: 170.77,
-    close: 170.5699,
-    volume: 749958,
-  },
-  {
-    date: '2022-01-18 11:20:00',
-    open: 170.69,
-    low: 170.56,
-    high: 170.86,
-    close: 170.645,
-    volume: 683936,
-  },
-  {
-    date: '2022-01-18 11:15:00',
-    open: 170.79,
-    low: 170.625,
-    high: 170.925,
-    close: 170.6901,
-    volume: 898231,
-  },
-  {
-    date: '2022-01-18 11:10:00',
-    open: 170.8164,
-    low: 170.645,
-    high: 170.96,
-    close: 170.79,
-    volume: 717271,
-  },
-  {
-    date: '2022-01-18 11:05:00',
-    open: 170.95,
-    low: 170.67,
-    high: 171.0872,
-    close: 170.8099,
-    volume: 803145,
-  },
-  {
-    date: '2022-01-18 11:00:00',
-    open: 171.1,
-    low: 170.86,
-    high: 171.1477,
-    close: 170.95,
-    volume: 734247,
-  },
-  {
-    date: '2022-01-18 10:55:00',
-    open: 171.11,
-    low: 170.99,
-    high: 171.34,
-    close: 171.0801,
-    volume: 842535,
-  },
-  {
-    date: '2022-01-18 10:50:00',
-    open: 171.46,
-    low: 171.1,
-    high: 171.46,
-    close: 171.1001,
-    volume: 749605,
-  },
-  {
-    date: '2022-01-18 10:45:00',
-    open: 171.995,
-    low: 171.43,
-    high: 172.11,
-    close: 171.44,
-    volume: 708931,
-  },
-  {
-    date: '2022-01-18 10:40:00',
-    open: 171.45,
-    low: 171.43,
-    high: 172.05,
-    close: 172.01,
-    volume: 791641,
-  },
-  {
-    date: '2022-01-18 10:35:00',
-    open: 171.66,
-    low: 171.4201,
-    high: 171.83,
-    close: 171.4516,
-    volume: 742502,
-  },
-  {
-    date: '2022-01-18 10:30:00',
-    open: 172.32,
-    low: 171.59,
-    high: 172.32,
-    close: 171.65,
-    volume: 816497,
-  },
-  {
-    date: '2022-01-18 10:25:00',
-    open: 172.07,
-    low: 171.94,
-    high: 172.32,
-    close: 172.31,
-    volume: 1084175,
-  },
-  {
-    date: '2022-01-18 10:20:00',
-    open: 171.82,
-    low: 171.65,
-    high: 172.14,
-    close: 172.07,
-    volume: 927999,
-  },
-  {
-    date: '2022-01-18 10:15:00',
-    open: 171.8,
-    low: 171.54,
-    high: 172.05,
-    close: 171.815,
-    volume: 1209835,
-  },
-  {
-    date: '2022-01-18 10:10:00',
-    open: 171.86,
-    low: 171.65,
-    high: 171.97,
-    close: 171.796,
-    volume: 1158991,
-  },
-  {
-    date: '2022-01-18 10:05:00',
-    open: 171.26,
-    low: 171.09,
-    high: 171.95,
-    close: 171.87,
-    volume: 1260297,
-  },
-  {
-    date: '2022-01-18 10:00:00',
-    open: 171.11,
-    low: 170.79,
-    high: 171.3225,
-    close: 171.24,
-    volume: 1215983,
-  },
-  {
-    date: '2022-01-18 09:55:00',
-    open: 170.87,
-    low: 170.59,
-    high: 171.22,
-    close: 171.11,
-    volume: 1163090,
-  },
-  {
-    date: '2022-01-18 09:50:00',
-    open: 170.86,
-    low: 170.64,
-    high: 171.24,
-    close: 170.879,
-    volume: 1396774,
-  },
-  {
-    date: '2022-01-18 09:45:00',
-    open: 172.18,
-    low: 170.75,
-    high: 172.24,
-    close: 170.88,
-    volume: 1942992,
-  },
-  {
-    date: '2022-01-18 09:40:00',
-    open: 171.26,
-    low: 171.12,
-    high: 172.4162,
-    close: 172.19,
-    volume: 2070080,
-  },
-  {
-    date: '2022-01-18 09:35:00',
-    open: 171.665,
-    low: 171.23,
-    high: 171.9,
-    close: 171.25,
-    volume: 1917290,
-  },
-  {
-    date: '2022-01-18 09:30:00',
-    open: 171.51,
-    low: 171.44,
-    high: 172.54,
-    close: 171.67,
-    volume: 4031243,
-  },
+	{
+		date: '2022-01-21 16:00:00',
+		open: 162.4,
+		low: 162.39,
+		high: 162.69,
+		close: 162.55,
+		volume: 1753842,
+	},
+	{
+		date: '2022-01-21 15:55:00',
+		open: 162.65,
+		low: 162.3,
+		high: 163.18,
+		close: 162.41,
+		volume: 4583337,
+	},
+	{
+		date: '2022-01-21 15:50:00',
+		open: 163.18,
+		low: 162.52,
+		high: 163.29,
+		close: 162.6399,
+		volume: 2425008,
+	},
+	{
+		date: '2022-01-21 15:45:00',
+		open: 163.19,
+		low: 162.92,
+		high: 163.4083,
+		close: 163.18,
+		volume: 1648816,
+	},
+	{
+		date: '2022-01-21 15:40:00',
+		open: 163.15,
+		low: 163.14,
+		high: 163.885,
+		close: 163.18,
+		volume: 1896376,
+	},
+	{
+		date: '2022-01-21 15:35:00',
+		open: 163.04,
+		low: 162.62,
+		high: 163.24,
+		close: 163.16,
+		volume: 1873912,
+	},
+	{
+		date: '2022-01-21 15:30:00',
+		open: 162.86,
+		low: 162.57,
+		high: 163.1699,
+		close: 163.06,
+		volume: 1896565,
+	},
+	{
+		date: '2022-01-21 15:25:00',
+		open: 163.0,
+		low: 162.581,
+		high: 163.1,
+		close: 162.86,
+		volume: 1477045,
+	},
+	{
+		date: '2022-01-21 15:20:00',
+		open: 163.09,
+		low: 162.7,
+		high: 163.15,
+		close: 162.99,
+		volume: 1348166,
+	},
+	{
+		date: '2022-01-21 15:15:00',
+		open: 162.88,
+		low: 162.7,
+		high: 163.45,
+		close: 163.12,
+		volume: 1490197,
+	},
+	{
+		date: '2022-01-21 15:10:00',
+		open: 162.53,
+		low: 162.405,
+		high: 162.9399,
+		close: 162.89,
+		volume: 1494587,
+	},
+	{
+		date: '2022-01-21 15:05:00',
+		open: 162.88,
+		low: 162.43,
+		high: 162.89,
+		close: 162.53,
+		volume: 1736393,
+	},
+	{
+		date: '2022-01-21 15:00:00',
+		open: 163.1195,
+		low: 162.8,
+		high: 163.38,
+		close: 162.88,
+		volume: 1480466,
+	},
+	{
+		date: '2022-01-21 14:55:00',
+		open: 163.48,
+		low: 163.06,
+		high: 163.64,
+		close: 163.11,
+		volume: 1255614,
+	},
+	{
+		date: '2022-01-21 14:50:00',
+		open: 163.73,
+		low: 163.31,
+		high: 163.73,
+		close: 163.4727,
+		volume: 692963,
+	},
+	{
+		date: '2022-01-21 14:45:00',
+		open: 163.53,
+		low: 163.49,
+		high: 163.765,
+		close: 163.7283,
+		volume: 691663,
+	},
+	{
+		date: '2022-01-21 14:40:00',
+		open: 163.7,
+		low: 163.44,
+		high: 163.92,
+		close: 163.51,
+		volume: 868701,
+	},
+	{
+		date: '2022-01-21 14:35:00',
+		open: 163.39,
+		low: 163.2,
+		high: 163.79,
+		close: 163.7,
+		volume: 1044987,
+	},
+	{
+		date: '2022-01-21 14:30:00',
+		open: 163.365,
+		low: 163.25,
+		high: 163.62,
+		close: 163.38,
+		volume: 855139,
+	},
+	{
+		date: '2022-01-21 14:25:00',
+		open: 164.22,
+		low: 163.3,
+		high: 164.24,
+		close: 163.37,
+		volume: 1089930,
+	},
+	{
+		date: '2022-01-21 14:20:00',
+		open: 163.91,
+		low: 163.69,
+		high: 164.32,
+		close: 164.23,
+		volume: 931845,
+	},
+	{
+		date: '2022-01-21 14:15:00',
+		open: 164.0,
+		low: 163.78,
+		high: 164.29,
+		close: 163.91,
+		volume: 771609,
+	},
+	{
+		date: '2022-01-21 14:10:00',
+		open: 164.015,
+		low: 163.9,
+		high: 164.3299,
+		close: 163.996,
+		volume: 802700,
+	},
+	{
+		date: '2022-01-21 14:05:00',
+		open: 164.07,
+		low: 163.84,
+		high: 164.3299,
+		close: 164.0,
+		volume: 854633,
+	},
+	{
+		date: '2022-01-21 14:00:00',
+		open: 163.765,
+		low: 163.7,
+		high: 164.14,
+		close: 164.08,
+		volume: 771297,
+	},
+	{
+		date: '2022-01-21 13:55:00',
+		open: 164.2819,
+		low: 163.62,
+		high: 164.34,
+		close: 163.765,
+		volume: 794903,
+	},
+	{
+		date: '2022-01-21 13:50:00',
+		open: 164.22,
+		low: 164.05,
+		high: 164.53,
+		close: 164.28,
+		volume: 677393,
+	},
+	{
+		date: '2022-01-21 13:45:00',
+		open: 164.17,
+		low: 163.91,
+		high: 164.32,
+		close: 164.24,
+		volume: 769442,
+	},
+	{
+		date: '2022-01-21 13:40:00',
+		open: 164.26,
+		low: 164.08,
+		high: 164.5,
+		close: 164.16,
+		volume: 706178,
+	},
+	{
+		date: '2022-01-21 13:35:00',
+		open: 163.729,
+		low: 163.57,
+		high: 164.42,
+		close: 164.25,
+		volume: 811891,
+	},
+	{
+		date: '2022-01-21 13:30:00',
+		open: 163.75,
+		low: 163.66,
+		high: 164.07,
+		close: 163.73,
+		volume: 844913,
+	},
+	{
+		date: '2022-01-21 13:25:00',
+		open: 163.45,
+		low: 163.39,
+		high: 163.93,
+		close: 163.75,
+		volume: 868020,
+	},
+	{
+		date: '2022-01-21 13:20:00',
+		open: 164.15,
+		low: 163.35,
+		high: 164.2,
+		close: 163.44,
+		volume: 907445,
+	},
+	{
+		date: '2022-01-21 13:15:00',
+		open: 164.62,
+		low: 163.9725,
+		high: 164.69,
+		close: 164.16,
+		volume: 736636,
+	},
+	{
+		date: '2022-01-21 13:10:00',
+		open: 164.02,
+		low: 164.02,
+		high: 164.78,
+		close: 164.625,
+		volume: 987059,
+	},
+	{
+		date: '2022-01-21 13:05:00',
+		open: 164.0185,
+		low: 163.86,
+		high: 164.245,
+		close: 164.025,
+		volume: 827993,
+	},
+	{
+		date: '2022-01-21 13:00:00',
+		open: 163.95,
+		low: 163.9,
+		high: 164.28,
+		close: 164.01,
+		volume: 794247,
+	},
+	{
+		date: '2022-01-21 12:55:00',
+		open: 163.61,
+		low: 163.46,
+		high: 164.03,
+		close: 163.96,
+		volume: 888389,
+	},
+	{
+		date: '2022-01-21 12:50:00',
+		open: 163.38,
+		low: 163.26,
+		high: 163.64,
+		close: 163.61,
+		volume: 648479,
+	},
+	{
+		date: '2022-01-21 12:45:00',
+		open: 163.54,
+		low: 163.25,
+		high: 163.82,
+		close: 163.37,
+		volume: 802509,
+	},
+	{
+		date: '2022-01-21 12:40:00',
+		open: 163.64,
+		low: 163.46,
+		high: 163.8914,
+		close: 163.5304,
+		volume: 709103,
+	},
+	{
+		date: '2022-01-21 12:35:00',
+		open: 163.63,
+		low: 163.402,
+		high: 163.97,
+		close: 163.65,
+		volume: 832765,
+	},
+	{
+		date: '2022-01-21 12:30:00',
+		open: 163.73,
+		low: 163.48,
+		high: 163.84,
+		close: 163.6201,
+		volume: 758662,
+	},
+	{
+		date: '2022-01-21 12:25:00',
+		open: 164.02,
+		low: 163.6946,
+		high: 164.23,
+		close: 163.73,
+		volume: 828334,
+	},
+	{
+		date: '2022-01-21 12:20:00',
+		open: 163.79,
+		low: 163.52,
+		high: 164.04,
+		close: 164.04,
+		volume: 1163135,
+	},
+	{
+		date: '2022-01-21 12:15:00',
+		open: 163.91,
+		low: 163.68,
+		high: 164.15,
+		close: 163.8,
+		volume: 922942,
+	},
+	{
+		date: '2022-01-21 12:10:00',
+		open: 164.13,
+		low: 163.8,
+		high: 164.31,
+		close: 163.91,
+		volume: 1051596,
+	},
+	{
+		date: '2022-01-21 12:05:00',
+		open: 164.4,
+		low: 164.0901,
+		high: 164.63,
+		close: 164.16,
+		volume: 802163,
+	},
+	{
+		date: '2022-01-21 12:00:00',
+		open: 164.42,
+		low: 164.17,
+		high: 164.67,
+		close: 164.38,
+		volume: 826696,
+	},
+	{
+		date: '2022-01-21 11:55:00',
+		open: 164.22,
+		low: 164.15,
+		high: 164.54,
+		close: 164.44,
+		volume: 909647,
+	},
+	{
+		date: '2022-01-21 11:50:00',
+		open: 164.6,
+		low: 164.02,
+		high: 164.7,
+		close: 164.2088,
+		volume: 894350,
+	},
+	{
+		date: '2022-01-21 11:45:00',
+		open: 165.08,
+		low: 164.45,
+		high: 165.11,
+		close: 164.6,
+		volume: 1276147,
+	},
+	{
+		date: '2022-01-21 11:40:00',
+		open: 165.06,
+		low: 164.87,
+		high: 165.2182,
+		close: 165.05,
+		volume: 750005,
+	},
+	{
+		date: '2022-01-21 11:35:00',
+		open: 165.78,
+		low: 164.86,
+		high: 165.94,
+		close: 165.06,
+		volume: 1308744,
+	},
+	{
+		date: '2022-01-21 11:30:00',
+		open: 165.94,
+		low: 165.1428,
+		high: 165.95,
+		close: 165.78,
+		volume: 1371342,
+	},
+	{
+		date: '2022-01-21 11:25:00',
+		open: 166.07,
+		low: 165.78,
+		high: 166.1898,
+		close: 165.95,
+		volume: 963225,
+	},
+	{
+		date: '2022-01-21 11:20:00',
+		open: 165.7261,
+		low: 165.44,
+		high: 166.33,
+		close: 166.07,
+		volume: 1372225,
+	},
+	{
+		date: '2022-01-21 11:15:00',
+		open: 165.07,
+		low: 165.07,
+		high: 165.88,
+		close: 165.715,
+		volume: 1500335,
+	},
+	{
+		date: '2022-01-21 11:10:00',
+		open: 165.39,
+		low: 164.93,
+		high: 165.51,
+		close: 165.06,
+		volume: 2200959,
+	},
+	{
+		date: '2022-01-21 11:05:00',
+		open: 164.97,
+		low: 164.93,
+		high: 165.41,
+		close: 165.4,
+		volume: 1213898,
+	},
+	{
+		date: '2022-01-21 11:00:00',
+		open: 165.17,
+		low: 164.65,
+		high: 165.25,
+		close: 164.95,
+		volume: 1179103,
+	},
+	{
+		date: '2022-01-21 10:55:00',
+		open: 165.52,
+		low: 164.82,
+		high: 165.67,
+		close: 165.16,
+		volume: 1263572,
+	},
+	{
+		date: '2022-01-21 10:50:00',
+		open: 165.22,
+		low: 165.13,
+		high: 165.84,
+		close: 165.53,
+		volume: 1795511,
+	},
+	{
+		date: '2022-01-21 10:45:00',
+		open: 165.03,
+		low: 164.67,
+		high: 165.2897,
+		close: 165.2014,
+		volume: 1113790,
+	},
+	{
+		date: '2022-01-21 10:40:00',
+		open: 164.94,
+		low: 164.62,
+		high: 165.29,
+		close: 165.03,
+		volume: 1515370,
+	},
+	{
+		date: '2022-01-21 10:35:00',
+		open: 165.45,
+		low: 164.9,
+		high: 165.78,
+		close: 164.9399,
+		volume: 1762388,
+	},
+	{
+		date: '2022-01-21 10:30:00',
+		open: 164.31,
+		low: 164.13,
+		high: 165.54,
+		close: 165.46,
+		volume: 2130613,
+	},
+	{
+		date: '2022-01-21 10:25:00',
+		open: 163.67,
+		low: 163.54,
+		high: 164.6,
+		close: 164.31,
+		volume: 1752836,
+	},
+	{
+		date: '2022-01-21 10:20:00',
+		open: 163.3,
+		low: 162.96,
+		high: 163.8376,
+		close: 163.68,
+		volume: 2401416,
+	},
+	{
+		date: '2022-01-21 10:15:00',
+		open: 163.6165,
+		low: 163.11,
+		high: 163.82,
+		close: 163.28,
+		volume: 1613638,
+	},
+	{
+		date: '2022-01-21 10:10:00',
+		open: 163.9,
+		low: 163.3299,
+		high: 164.0,
+		close: 163.63,
+		volume: 1542727,
+	},
+	{
+		date: '2022-01-21 10:05:00',
+		open: 163.32,
+		low: 163.1604,
+		high: 164.06,
+		close: 163.9,
+		volume: 1951077,
+	},
+	{
+		date: '2022-01-21 10:00:00',
+		open: 163.54,
+		low: 163.22,
+		high: 163.95,
+		close: 163.3201,
+		volume: 1647151,
+	},
+	{
+		date: '2022-01-21 09:55:00',
+		open: 163.35,
+		low: 163.18,
+		high: 163.78,
+		close: 163.53,
+		volume: 1408586,
+	},
+	{
+		date: '2022-01-21 09:50:00',
+		open: 164.77,
+		low: 163.23,
+		high: 164.9,
+		close: 163.365,
+		volume: 2033414,
+	},
+	{
+		date: '2022-01-21 09:45:00',
+		open: 163.71,
+		low: 163.67,
+		high: 164.89,
+		close: 164.77,
+		volume: 1909257,
+	},
+	{
+		date: '2022-01-21 09:40:00',
+		open: 164.52,
+		low: 163.7,
+		high: 164.71,
+		close: 163.72,
+		volume: 3802988,
+	},
+	{
+		date: '2022-01-21 09:35:00',
+		open: 165.2501,
+		low: 164.42,
+		high: 165.6,
+		close: 164.51,
+		volume: 2253073,
+	},
+	{
+		date: '2022-01-21 09:30:00',
+		open: 164.415,
+		low: 164.36,
+		high: 166.0,
+		close: 165.26,
+		volume: 7362686,
+	},
+	{
+		date: '2022-01-20 16:00:00',
+		open: 164.51,
+		low: 163.93,
+		high: 167.4329,
+		close: 163.99,
+		volume: 1693026,
+	},
+	{
+		date: '2022-01-20 15:55:00',
+		open: 164.58,
+		low: 164.18,
+		high: 164.96,
+		close: 164.52,
+		volume: 4242438,
+	},
+	{
+		date: '2022-01-20 15:50:00',
+		open: 164.95,
+		low: 164.5,
+		high: 165.24,
+		close: 164.5918,
+		volume: 2780585,
+	},
+	{
+		date: '2022-01-20 15:45:00',
+		open: 165.11,
+		low: 164.9,
+		high: 165.48,
+		close: 164.9499,
+		volume: 1804633,
+	},
+	{
+		date: '2022-01-20 15:40:00',
+		open: 165.205,
+		low: 165.06,
+		high: 165.56,
+		close: 165.1,
+		volume: 1547931,
+	},
+	{
+		date: '2022-01-20 15:35:00',
+		open: 165.71,
+		low: 165.0,
+		high: 165.71,
+		close: 165.22,
+		volume: 2269633,
+	},
+	{
+		date: '2022-01-20 15:30:00',
+		open: 165.84,
+		low: 165.495,
+		high: 165.97,
+		close: 165.705,
+		volume: 1207993,
+	},
+	{
+		date: '2022-01-20 15:25:00',
+		open: 165.98,
+		low: 165.65,
+		high: 166.07,
+		close: 165.845,
+		volume: 1232471,
+	},
+	{
+		date: '2022-01-20 15:20:00',
+		open: 166.25,
+		low: 165.95,
+		high: 166.2599,
+		close: 165.97,
+		volume: 910737,
+	},
+	{
+		date: '2022-01-20 15:15:00',
+		open: 166.04,
+		low: 165.92,
+		high: 166.32,
+		close: 166.26,
+		volume: 1001245,
+	},
+	{
+		date: '2022-01-20 15:10:00',
+		open: 166.2,
+		low: 165.96,
+		high: 166.29,
+		close: 166.04,
+		volume: 1276014,
+	},
+	{
+		date: '2022-01-20 15:05:00',
+		open: 166.4,
+		low: 166.09,
+		high: 166.53,
+		close: 166.1999,
+		volume: 933259,
+	},
+	{
+		date: '2022-01-20 15:00:00',
+		open: 166.66,
+		low: 166.27,
+		high: 166.97,
+		close: 166.408,
+		volume: 1068984,
+	},
+	{
+		date: '2022-01-20 14:55:00',
+		open: 166.73,
+		low: 166.6,
+		high: 166.88,
+		close: 166.66,
+		volume: 749847,
+	},
+	{
+		date: '2022-01-20 14:50:00',
+		open: 166.63,
+		low: 166.55,
+		high: 166.78,
+		close: 166.73,
+		volume: 613845,
+	},
+	{
+		date: '2022-01-20 14:45:00',
+		open: 166.63,
+		low: 166.32,
+		high: 166.8,
+		close: 166.64,
+		volume: 904658,
+	},
+	{
+		date: '2022-01-20 14:40:00',
+		open: 166.86,
+		low: 166.59,
+		high: 166.86,
+		close: 166.61,
+		volume: 571688,
+	},
+	{
+		date: '2022-01-20 14:35:00',
+		open: 166.72,
+		low: 166.61,
+		high: 166.93,
+		close: 166.865,
+		volume: 907051,
+	},
+	{
+		date: '2022-01-20 14:30:00',
+		open: 166.6474,
+		low: 166.42,
+		high: 166.7599,
+		close: 166.719,
+		volume: 853708,
+	},
+	{
+		date: '2022-01-20 14:25:00',
+		open: 166.8921,
+		low: 166.6,
+		high: 166.9735,
+		close: 166.6401,
+		volume: 727879,
+	},
+	{
+		date: '2022-01-20 14:20:00',
+		open: 166.98,
+		low: 166.71,
+		high: 167.0999,
+		close: 166.9099,
+		volume: 816666,
+	},
+	{
+		date: '2022-01-20 14:15:00',
+		open: 166.775,
+		low: 166.46,
+		high: 167.13,
+		close: 166.9701,
+		volume: 1293989,
+	},
+	{
+		date: '2022-01-20 14:10:00',
+		open: 167.64,
+		low: 166.64,
+		high: 167.74,
+		close: 166.77,
+		volume: 1153934,
+	},
+	{
+		date: '2022-01-20 14:05:00',
+		open: 167.8599,
+		low: 167.6,
+		high: 167.88,
+		close: 167.6399,
+		volume: 435383,
+	},
+	{
+		date: '2022-01-20 14:00:00',
+		open: 167.4388,
+		low: 167.39,
+		high: 167.91,
+		close: 167.8429,
+		volume: 682593,
+	},
+	{
+		date: '2022-01-20 13:55:00',
+		open: 167.7181,
+		low: 167.31,
+		high: 167.7287,
+		close: 167.4167,
+		volume: 649184,
+	},
+	{
+		date: '2022-01-20 13:50:00',
+		open: 167.59,
+		low: 167.53,
+		high: 167.77,
+		close: 167.7101,
+		volume: 440474,
+	},
+	{
+		date: '2022-01-20 13:45:00',
+		open: 167.6075,
+		low: 167.31,
+		high: 167.75,
+		close: 167.6,
+		volume: 534014,
+	},
+	{
+		date: '2022-01-20 13:40:00',
+		open: 167.5899,
+		low: 167.51,
+		high: 167.8,
+		close: 167.6,
+		volume: 602757,
+	},
+	{
+		date: '2022-01-20 13:35:00',
+		open: 167.2938,
+		low: 167.25,
+		high: 167.66,
+		close: 167.58,
+		volume: 717176,
+	},
+	{
+		date: '2022-01-20 13:30:00',
+		open: 167.61,
+		low: 167.15,
+		high: 167.7,
+		close: 167.3,
+		volume: 1115739,
+	},
+	{
+		date: '2022-01-20 13:25:00',
+		open: 167.73,
+		low: 167.6,
+		high: 167.94,
+		close: 167.6,
+		volume: 556454,
+	},
+	{
+		date: '2022-01-20 13:20:00',
+		open: 167.88,
+		low: 167.68,
+		high: 168.02,
+		close: 167.7204,
+		volume: 509774,
+	},
+	{
+		date: '2022-01-20 13:15:00',
+		open: 168.07,
+		low: 167.69,
+		high: 168.07,
+		close: 167.89,
+		volume: 531857,
+	},
+	{
+		date: '2022-01-20 13:10:00',
+		open: 168.025,
+		low: 167.86,
+		high: 168.15,
+		close: 168.07,
+		volume: 862052,
+	},
+	{
+		date: '2022-01-20 13:05:00',
+		open: 168.13,
+		low: 167.76,
+		high: 168.13,
+		close: 168.02,
+		volume: 962708,
+	},
+	{
+		date: '2022-01-20 13:00:00',
+		open: 168.64,
+		low: 168.06,
+		high: 168.64,
+		close: 168.14,
+		volume: 718480,
+	},
+	{
+		date: '2022-01-20 12:55:00',
+		open: 168.6397,
+		low: 168.495,
+		high: 168.73,
+		close: 168.645,
+		volume: 593316,
+	},
+	{
+		date: '2022-01-20 12:50:00',
+		open: 169.03,
+		low: 168.59,
+		high: 169.04,
+		close: 168.63,
+		volume: 665207,
+	},
+	{
+		date: '2022-01-20 12:45:00',
+		open: 169.185,
+		low: 168.9501,
+		high: 169.22,
+		close: 169.02,
+		volume: 357964,
+	},
+	{
+		date: '2022-01-20 12:40:00',
+		open: 168.895,
+		low: 168.85,
+		high: 169.22,
+		close: 169.1801,
+		volume: 1159387,
+	},
+	{
+		date: '2022-01-20 12:35:00',
+		open: 168.94,
+		low: 168.82,
+		high: 169.0299,
+		close: 168.895,
+		volume: 359459,
+	},
+	{
+		date: '2022-01-20 12:30:00',
+		open: 169.09,
+		low: 168.895,
+		high: 169.2,
+		close: 168.93,
+		volume: 481016,
+	},
+	{
+		date: '2022-01-20 12:25:00',
+		open: 169.02,
+		low: 168.81,
+		high: 169.1,
+		close: 169.075,
+		volume: 1502074,
+	},
+	{
+		date: '2022-01-20 12:20:00',
+		open: 168.94,
+		low: 166.85,
+		high: 169.06,
+		close: 169.02,
+		volume: 425870,
+	},
+	{
+		date: '2022-01-20 12:15:00',
+		open: 169.11,
+		low: 168.94,
+		high: 169.13,
+		close: 168.95,
+		volume: 373478,
+	},
+	{
+		date: '2022-01-20 12:10:00',
+		open: 169.0,
+		low: 168.935,
+		high: 169.21,
+		close: 169.1099,
+		volume: 483782,
+	},
+	{
+		date: '2022-01-20 12:05:00',
+		open: 169.149,
+		low: 168.92,
+		high: 169.15,
+		close: 169.01,
+		volume: 737036,
+	},
+	{
+		date: '2022-01-20 12:00:00',
+		open: 169.11,
+		low: 169.1,
+		high: 169.25,
+		close: 169.1399,
+		volume: 363455,
+	},
+	{
+		date: '2022-01-20 11:55:00',
+		open: 169.11,
+		low: 169.07,
+		high: 169.32,
+		close: 169.11,
+		volume: 1690310,
+	},
+	{
+		date: '2022-01-20 11:50:00',
+		open: 168.98,
+		low: 168.97,
+		high: 169.29,
+		close: 169.1,
+		volume: 559553,
+	},
+	{
+		date: '2022-01-20 11:45:00',
+		open: 168.78,
+		low: 168.78,
+		high: 169.05,
+		close: 168.99,
+		volume: 672407,
+	},
+	{
+		date: '2022-01-20 11:40:00',
+		open: 169.325,
+		low: 168.7893,
+		high: 169.3788,
+		close: 168.79,
+		volume: 915476,
+	},
+	{
+		date: '2022-01-20 11:35:00',
+		open: 169.42,
+		low: 169.1218,
+		high: 169.48,
+		close: 169.33,
+		volume: 585187,
+	},
+	{
+		date: '2022-01-20 11:30:00',
+		open: 169.36,
+		low: 169.28,
+		high: 169.46,
+		close: 169.42,
+		volume: 576425,
+	},
+	{
+		date: '2022-01-20 11:25:00',
+		open: 169.07,
+		low: 169.02,
+		high: 169.43,
+		close: 169.35,
+		volume: 591248,
+	},
+	{
+		date: '2022-01-20 11:20:00',
+		open: 169.215,
+		low: 168.98,
+		high: 169.215,
+		close: 169.075,
+		volume: 667630,
+	},
+	{
+		date: '2022-01-20 11:15:00',
+		open: 169.13,
+		low: 168.96,
+		high: 169.31,
+		close: 169.21,
+		volume: 743612,
+	},
+	{
+		date: '2022-01-20 11:10:00',
+		open: 169.2,
+		low: 168.96,
+		high: 169.22,
+		close: 169.13,
+		volume: 693010,
+	},
+	{
+		date: '2022-01-20 11:05:00',
+		open: 169.312,
+		low: 169.15,
+		high: 169.43,
+		close: 169.18,
+		volume: 727346,
+	},
+	{
+		date: '2022-01-20 11:00:00',
+		open: 169.23,
+		low: 169.18,
+		high: 169.5498,
+		close: 169.32,
+		volume: 767871,
+	},
+	{
+		date: '2022-01-20 10:55:00',
+		open: 169.33,
+		low: 169.04,
+		high: 169.35,
+		close: 169.229,
+		volume: 630325,
+	},
+	{
+		date: '2022-01-20 10:50:00',
+		open: 169.2,
+		low: 169.0281,
+		high: 169.45,
+		close: 169.339,
+		volume: 782358,
+	},
+	{
+		date: '2022-01-20 10:45:00',
+		open: 168.9961,
+		low: 168.94,
+		high: 169.33,
+		close: 169.195,
+		volume: 908414,
+	},
+	{
+		date: '2022-01-20 10:40:00',
+		open: 169.17,
+		low: 168.98,
+		high: 169.305,
+		close: 169.01,
+		volume: 726188,
+	},
+	{
+		date: '2022-01-20 10:35:00',
+		open: 169.38,
+		low: 169.12,
+		high: 169.54,
+		close: 169.165,
+		volume: 852422,
+	},
+	{
+		date: '2022-01-20 10:30:00',
+		open: 169.2399,
+		low: 169.239,
+		high: 169.68,
+		close: 169.3655,
+		volume: 1161920,
+	},
+	{
+		date: '2022-01-20 10:25:00',
+		open: 168.975,
+		low: 168.73,
+		high: 169.34,
+		close: 169.23,
+		volume: 1124190,
+	},
+	{
+		date: '2022-01-20 10:20:00',
+		open: 168.6511,
+		low: 168.61,
+		high: 168.99,
+		close: 168.98,
+		volume: 889660,
+	},
+	{
+		date: '2022-01-20 10:15:00',
+		open: 168.75,
+		low: 168.56,
+		high: 168.91,
+		close: 168.64,
+		volume: 1071189,
+	},
+	{
+		date: '2022-01-20 10:10:00',
+		open: 168.75,
+		low: 168.5522,
+		high: 168.9,
+		close: 168.75,
+		volume: 993577,
+	},
+	{
+		date: '2022-01-20 10:05:00',
+		open: 168.66,
+		low: 168.47,
+		high: 168.84,
+		close: 168.75,
+		volume: 1164082,
+	},
+	{
+		date: '2022-01-20 10:00:00',
+		open: 168.87,
+		low: 168.61,
+		high: 169.06,
+		close: 168.656,
+		volume: 1319335,
+	},
+	{
+		date: '2022-01-20 09:55:00',
+		open: 168.41,
+		low: 168.26,
+		high: 168.92,
+		close: 168.88,
+		volume: 1373398,
+	},
+	{
+		date: '2022-01-20 09:50:00',
+		open: 168.41,
+		low: 168.225,
+		high: 168.595,
+		close: 168.39,
+		volume: 1519216,
+	},
+	{
+		date: '2022-01-20 09:45:00',
+		open: 168.48,
+		low: 168.185,
+		high: 168.67,
+		close: 168.42,
+		volume: 1656337,
+	},
+	{
+		date: '2022-01-20 09:40:00',
+		open: 167.35,
+		low: 167.2899,
+		high: 168.55,
+		close: 168.5,
+		volume: 1884411,
+	},
+	{
+		date: '2022-01-20 09:35:00',
+		open: 167.48,
+		low: 166.9,
+		high: 167.6298,
+		close: 167.345,
+		volume: 1543085,
+	},
+	{
+		date: '2022-01-20 09:30:00',
+		open: 166.98,
+		low: 166.76,
+		high: 167.74,
+		close: 167.5017,
+		volume: 3034473,
+	},
+	{
+		date: '2022-01-19 16:00:00',
+		open: 166.23,
+		low: 165.9,
+		high: 166.23,
+		close: 165.93,
+		volume: 2650254,
+	},
+	{
+		date: '2022-01-19 15:55:00',
+		open: 166.59,
+		low: 165.94,
+		high: 166.72,
+		close: 166.21,
+		volume: 3900789,
+	},
+	{
+		date: '2022-01-19 15:50:00',
+		open: 166.86,
+		low: 166.48,
+		high: 166.87,
+		close: 166.59,
+		volume: 2683419,
+	},
+	{
+		date: '2022-01-19 15:45:00',
+		open: 167.2,
+		low: 166.86,
+		high: 167.2742,
+		close: 166.87,
+		volume: 2035440,
+	},
+	{
+		date: '2022-01-19 15:40:00',
+		open: 167.2,
+		low: 167.18,
+		high: 167.39,
+		close: 167.21,
+		volume: 1139956,
+	},
+	{
+		date: '2022-01-19 15:35:00',
+		open: 167.34,
+		low: 167.18,
+		high: 167.36,
+		close: 167.1928,
+		volume: 1083892,
+	},
+	{
+		date: '2022-01-19 15:30:00',
+		open: 167.51,
+		low: 167.28,
+		high: 167.675,
+		close: 167.35,
+		volume: 1111815,
+	},
+	{
+		date: '2022-01-19 15:25:00',
+		open: 167.7345,
+		low: 167.49,
+		high: 167.89,
+		close: 167.51,
+		volume: 1165017,
+	},
+	{
+		date: '2022-01-19 15:20:00',
+		open: 167.84,
+		low: 167.69,
+		high: 168.05,
+		close: 167.73,
+		volume: 694522,
+	},
+	{
+		date: '2022-01-19 15:15:00',
+		open: 168.155,
+		low: 167.75,
+		high: 168.225,
+		close: 167.85,
+		volume: 690920,
+	},
+	{
+		date: '2022-01-19 15:10:00',
+		open: 167.86,
+		low: 167.79,
+		high: 168.2699,
+		close: 168.1599,
+		volume: 800848,
+	},
+	{
+		date: '2022-01-19 15:05:00',
+		open: 167.79,
+		low: 167.66,
+		high: 167.97,
+		close: 167.86,
+		volume: 789264,
+	},
+	{
+		date: '2022-01-19 15:00:00',
+		open: 167.95,
+		low: 167.67,
+		high: 168.065,
+		close: 167.79,
+		volume: 850513,
+	},
+	{
+		date: '2022-01-19 14:55:00',
+		open: 168.05,
+		low: 167.92,
+		high: 168.24,
+		close: 167.95,
+		volume: 724762,
+	},
+	{
+		date: '2022-01-19 14:50:00',
+		open: 167.7551,
+		low: 167.7551,
+		high: 168.07,
+		close: 168.05,
+		volume: 651995,
+	},
+	{
+		date: '2022-01-19 14:45:00',
+		open: 167.98,
+		low: 167.69,
+		high: 168.0,
+		close: 167.76,
+		volume: 770116,
+	},
+	{
+		date: '2022-01-19 14:40:00',
+		open: 167.695,
+		low: 167.61,
+		high: 168.04,
+		close: 167.98,
+		volume: 853176,
+	},
+	{
+		date: '2022-01-19 14:35:00',
+		open: 167.76,
+		low: 167.65,
+		high: 167.96,
+		close: 167.6955,
+		volume: 769848,
+	},
+	{
+		date: '2022-01-19 14:30:00',
+		open: 167.645,
+		low: 167.54,
+		high: 167.84,
+		close: 167.76,
+		volume: 847311,
+	},
+	{
+		date: '2022-01-19 14:25:00',
+		open: 168.08,
+		low: 167.62,
+		high: 168.0951,
+		close: 167.65,
+		volume: 912682,
+	},
+	{
+		date: '2022-01-19 14:20:00',
+		open: 168.06,
+		low: 167.815,
+		high: 168.1499,
+		close: 168.094,
+		volume: 686981,
+	},
+	{
+		date: '2022-01-19 14:15:00',
+		open: 167.99,
+		low: 167.86,
+		high: 168.105,
+		close: 168.0638,
+		volume: 629779,
+	},
+	{
+		date: '2022-01-19 14:10:00',
+		open: 168.13,
+		low: 167.94,
+		high: 168.25,
+		close: 167.995,
+		volume: 781192,
+	},
+	{
+		date: '2022-01-19 14:05:00',
+		open: 168.43,
+		low: 167.94,
+		high: 168.44,
+		close: 168.13,
+		volume: 1009872,
+	},
+	{
+		date: '2022-01-19 14:00:00',
+		open: 168.44,
+		low: 168.36,
+		high: 168.66,
+		close: 168.43,
+		volume: 691926,
+	},
+	{
+		date: '2022-01-19 13:55:00',
+		open: 168.62,
+		low: 168.4333,
+		high: 168.78,
+		close: 168.45,
+		volume: 531511,
+	},
+	{
+		date: '2022-01-19 13:50:00',
+		open: 168.53,
+		low: 168.37,
+		high: 168.6699,
+		close: 168.6199,
+		volume: 679346,
+	},
+	{
+		date: '2022-01-19 13:45:00',
+		open: 168.66,
+		low: 168.5,
+		high: 168.7,
+		close: 168.53,
+		volume: 524912,
+	},
+	{
+		date: '2022-01-19 13:40:00',
+		open: 168.6985,
+		low: 168.5,
+		high: 168.72,
+		close: 168.6511,
+		volume: 545286,
+	},
+	{
+		date: '2022-01-19 13:35:00',
+		open: 168.87,
+		low: 168.59,
+		high: 168.87,
+		close: 168.7,
+		volume: 570100,
+	},
+	{
+		date: '2022-01-19 13:30:00',
+		open: 168.81,
+		low: 168.67,
+		high: 169.03,
+		close: 168.87,
+		volume: 607161,
+	},
+	{
+		date: '2022-01-19 13:25:00',
+		open: 169.19,
+		low: 168.79,
+		high: 169.26,
+		close: 168.8103,
+		volume: 487033,
+	},
+	{
+		date: '2022-01-19 13:20:00',
+		open: 169.22,
+		low: 169.01,
+		high: 169.35,
+		close: 169.185,
+		volume: 543957,
+	},
+	{
+		date: '2022-01-19 13:15:00',
+		open: 169.23,
+		low: 169.14,
+		high: 169.45,
+		close: 169.22,
+		volume: 1539638,
+	},
+	{
+		date: '2022-01-19 13:10:00',
+		open: 169.01,
+		low: 169.01,
+		high: 169.29,
+		close: 169.23,
+		volume: 647487,
+	},
+	{
+		date: '2022-01-19 13:05:00',
+		open: 169.1,
+		low: 168.91,
+		high: 169.1668,
+		close: 169.0,
+		volume: 606960,
+	},
+	{
+		date: '2022-01-19 13:00:00',
+		open: 169.08,
+		low: 168.76,
+		high: 169.24,
+		close: 169.1,
+		volume: 704566,
+	},
+	{
+		date: '2022-01-19 12:55:00',
+		open: 168.66,
+		low: 168.66,
+		high: 169.13,
+		close: 169.08,
+		volume: 560451,
+	},
+	{
+		date: '2022-01-19 12:50:00',
+		open: 168.97,
+		low: 168.63,
+		high: 169.1,
+		close: 168.658,
+		volume: 1013004,
+	},
+	{
+		date: '2022-01-19 12:45:00',
+		open: 168.94,
+		low: 168.85,
+		high: 169.2892,
+		close: 168.9689,
+		volume: 642158,
+	},
+	{
+		date: '2022-01-19 12:40:00',
+		open: 168.95,
+		low: 168.855,
+		high: 169.07,
+		close: 168.96,
+		volume: 482438,
+	},
+	{
+		date: '2022-01-19 12:35:00',
+		open: 168.9599,
+		low: 168.9,
+		high: 169.17,
+		close: 168.9401,
+		volume: 550259,
+	},
+	{
+		date: '2022-01-19 12:30:00',
+		open: 169.065,
+		low: 168.9,
+		high: 169.2,
+		close: 168.96,
+		volume: 604682,
+	},
+	{
+		date: '2022-01-19 12:25:00',
+		open: 169.03,
+		low: 168.91,
+		high: 169.1245,
+		close: 169.04,
+		volume: 527622,
+	},
+	{
+		date: '2022-01-19 12:20:00',
+		open: 168.85,
+		low: 168.7901,
+		high: 169.11,
+		close: 169.035,
+		volume: 753031,
+	},
+	{
+		date: '2022-01-19 12:15:00',
+		open: 169.07,
+		low: 168.7236,
+		high: 169.09,
+		close: 168.855,
+		volume: 760001,
+	},
+	{
+		date: '2022-01-19 12:10:00',
+		open: 168.825,
+		low: 168.799,
+		high: 169.62,
+		close: 169.08,
+		volume: 1232031,
+	},
+	{
+		date: '2022-01-19 12:05:00',
+		open: 168.565,
+		low: 168.515,
+		high: 168.87,
+		close: 168.8161,
+		volume: 653248,
+	},
+	{
+		date: '2022-01-19 12:00:00',
+		open: 168.33,
+		low: 168.33,
+		high: 168.67,
+		close: 168.5609,
+		volume: 863196,
+	},
+	{
+		date: '2022-01-19 11:55:00',
+		open: 168.2183,
+		low: 168.14,
+		high: 168.39,
+		close: 168.33,
+		volume: 711466,
+	},
+	{
+		date: '2022-01-19 11:50:00',
+		open: 167.77,
+		low: 167.69,
+		high: 168.29,
+		close: 168.215,
+		volume: 1041154,
+	},
+	{
+		date: '2022-01-19 11:45:00',
+		open: 167.44,
+		low: 167.4,
+		high: 167.8897,
+		close: 167.7649,
+		volume: 1059934,
+	},
+	{
+		date: '2022-01-19 11:40:00',
+		open: 167.75,
+		low: 167.32,
+		high: 167.76,
+		close: 167.43,
+		volume: 1201586,
+	},
+	{
+		date: '2022-01-19 11:35:00',
+		open: 167.69,
+		low: 167.52,
+		high: 167.87,
+		close: 167.75,
+		volume: 918295,
+	},
+	{
+		date: '2022-01-19 11:30:00',
+		open: 167.7917,
+		low: 167.62,
+		high: 168.18,
+		close: 167.69,
+		volume: 1243798,
+	},
+	{
+		date: '2022-01-19 11:25:00',
+		open: 167.76,
+		low: 167.58,
+		high: 168.0299,
+		close: 167.78,
+		volume: 1742482,
+	},
+	{
+		date: '2022-01-19 11:20:00',
+		open: 168.11,
+		low: 167.54,
+		high: 168.16,
+		close: 167.77,
+		volume: 2055875,
+	},
+	{
+		date: '2022-01-19 11:15:00',
+		open: 168.28,
+		low: 168.05,
+		high: 168.5,
+		close: 168.1022,
+		volume: 1208945,
+	},
+	{
+		date: '2022-01-19 11:10:00',
+		open: 168.86,
+		low: 168.25,
+		high: 168.86,
+		close: 168.2701,
+		volume: 1278690,
+	},
+	{
+		date: '2022-01-19 11:05:00',
+		open: 168.93,
+		low: 168.69,
+		high: 168.985,
+		close: 168.865,
+		volume: 828211,
+	},
+	{
+		date: '2022-01-19 11:00:00',
+		open: 168.9,
+		low: 168.87,
+		high: 169.2485,
+		close: 168.9389,
+		volume: 1078375,
+	},
+	{
+		date: '2022-01-19 10:55:00',
+		open: 168.81,
+		low: 168.661,
+		high: 169.0687,
+		close: 168.895,
+		volume: 1084972,
+	},
+	{
+		date: '2022-01-19 10:50:00',
+		open: 168.71,
+		low: 168.66,
+		high: 168.93,
+		close: 168.815,
+		volume: 1157775,
+	},
+	{
+		date: '2022-01-19 10:45:00',
+		open: 169.13,
+		low: 168.64,
+		high: 169.26,
+		close: 168.7,
+		volume: 1190838,
+	},
+	{
+		date: '2022-01-19 10:40:00',
+		open: 169.289,
+		low: 168.92,
+		high: 169.29,
+		close: 169.13,
+		volume: 1342890,
+	},
+	{
+		date: '2022-01-19 10:35:00',
+		open: 169.82,
+		low: 169.23,
+		high: 169.83,
+		close: 169.28,
+		volume: 1009177,
+	},
+	{
+		date: '2022-01-19 10:30:00',
+		open: 169.89,
+		low: 169.63,
+		high: 170.04,
+		close: 169.8279,
+		volume: 882216,
+	},
+	{
+		date: '2022-01-19 10:25:00',
+		open: 170.05,
+		low: 169.79,
+		high: 170.1,
+		close: 169.87,
+		volume: 756984,
+	},
+	{
+		date: '2022-01-19 10:20:00',
+		open: 170.31,
+		low: 169.8919,
+		high: 170.35,
+		close: 170.0501,
+		volume: 983106,
+	},
+	{
+		date: '2022-01-19 10:15:00',
+		open: 170.98,
+		low: 170.14,
+		high: 171.01,
+		close: 170.31,
+		volume: 1169892,
+	},
+	{
+		date: '2022-01-19 10:10:00',
+		open: 170.72,
+		low: 170.42,
+		high: 171.08,
+		close: 170.9799,
+		volume: 1421334,
+	},
+	{
+		date: '2022-01-19 10:05:00',
+		open: 170.78,
+		low: 170.48,
+		high: 170.92,
+		close: 170.7118,
+		volume: 1124974,
+	},
+	{
+		date: '2022-01-19 10:00:00',
+		open: 170.95,
+		low: 170.52,
+		high: 170.95,
+		close: 170.77,
+		volume: 1236235,
+	},
+	{
+		date: '2022-01-19 09:55:00',
+		open: 170.37,
+		low: 170.31,
+		high: 170.95,
+		close: 170.9315,
+		volume: 1215564,
+	},
+	{
+		date: '2022-01-19 09:50:00',
+		open: 170.3,
+		low: 170.14,
+		high: 170.7399,
+		close: 170.36,
+		volume: 1103963,
+	},
+	{
+		date: '2022-01-19 09:45:00',
+		open: 170.0235,
+		low: 169.95,
+		high: 170.69,
+		close: 170.29,
+		volume: 1524365,
+	},
+	{
+		date: '2022-01-19 09:40:00',
+		open: 170.25,
+		low: 169.56,
+		high: 170.34,
+		close: 170.02,
+		volume: 1693100,
+	},
+	{
+		date: '2022-01-19 09:35:00',
+		open: 169.46,
+		low: 169.28,
+		high: 170.56,
+		close: 170.2301,
+		volume: 1783219,
+	},
+	{
+		date: '2022-01-19 09:30:00',
+		open: 170.0,
+		low: 169.29,
+		high: 170.98,
+		close: 169.4694,
+		volume: 2965328,
+	},
+	{
+		date: '2022-01-18 16:00:00',
+		open: 169.8,
+		low: 169.73,
+		high: 170.0737,
+		close: 169.8,
+		volume: 4636399,
+	},
+	{
+		date: '2022-01-18 15:55:00',
+		open: 169.5,
+		low: 169.405,
+		high: 170.09,
+		close: 169.83,
+		volume: 4447449,
+	},
+	{
+		date: '2022-01-18 15:50:00',
+		open: 170.25,
+		low: 169.45,
+		high: 170.45,
+		close: 169.49,
+		volume: 2210533,
+	},
+	{
+		date: '2022-01-18 15:45:00',
+		open: 169.86,
+		low: 169.83,
+		high: 170.2483,
+		close: 170.2483,
+		volume: 1090935,
+	},
+	{
+		date: '2022-01-18 15:40:00',
+		open: 170.175,
+		low: 169.82,
+		high: 170.25,
+		close: 169.85,
+		volume: 913605,
+	},
+	{
+		date: '2022-01-18 15:35:00',
+		open: 170.24,
+		low: 170.08,
+		high: 170.3,
+		close: 170.175,
+		volume: 888633,
+	},
+	{
+		date: '2022-01-18 15:30:00',
+		open: 170.34,
+		low: 170.16,
+		high: 170.52,
+		close: 170.24,
+		volume: 923270,
+	},
+	{
+		date: '2022-01-18 15:25:00',
+		open: 170.22,
+		low: 170.05,
+		high: 170.38,
+		close: 170.34,
+		volume: 670850,
+	},
+	{
+		date: '2022-01-18 15:20:00',
+		open: 170.0341,
+		low: 169.98,
+		high: 170.31,
+		close: 170.2213,
+		volume: 715710,
+	},
+	{
+		date: '2022-01-18 15:15:00',
+		open: 170.4599,
+		low: 169.84,
+		high: 170.4599,
+		close: 170.03,
+		volume: 1150826,
+	},
+	{
+		date: '2022-01-18 15:10:00',
+		open: 170.5557,
+		low: 170.42,
+		high: 170.66,
+		close: 170.4599,
+		volume: 523257,
+	},
+	{
+		date: '2022-01-18 15:05:00',
+		open: 170.57,
+		low: 170.35,
+		high: 170.68,
+		close: 170.55,
+		volume: 577398,
+	},
+	{
+		date: '2022-01-18 15:00:00',
+		open: 170.5701,
+		low: 170.41,
+		high: 170.77,
+		close: 170.5666,
+		volume: 799279,
+	},
+	{
+		date: '2022-01-18 14:55:00',
+		open: 170.655,
+		low: 170.52,
+		high: 170.79,
+		close: 170.57,
+		volume: 699584,
+	},
+	{
+		date: '2022-01-18 14:50:00',
+		open: 170.52,
+		low: 170.5,
+		high: 170.72,
+		close: 170.66,
+		volume: 621478,
+	},
+	{
+		date: '2022-01-18 14:45:00',
+		open: 170.27,
+		low: 170.21,
+		high: 170.56,
+		close: 170.52,
+		volume: 566150,
+	},
+	{
+		date: '2022-01-18 14:40:00',
+		open: 170.4,
+		low: 170.23,
+		high: 170.43,
+		close: 170.27,
+		volume: 489615,
+	},
+	{
+		date: '2022-01-18 14:35:00',
+		open: 170.32,
+		low: 170.255,
+		high: 170.52,
+		close: 170.395,
+		volume: 485763,
+	},
+	{
+		date: '2022-01-18 14:30:00',
+		open: 170.15,
+		low: 170.14,
+		high: 170.59,
+		close: 170.32,
+		volume: 758337,
+	},
+	{
+		date: '2022-01-18 14:25:00',
+		open: 170.17,
+		low: 170.01,
+		high: 170.24,
+		close: 170.16,
+		volume: 436856,
+	},
+	{
+		date: '2022-01-18 14:20:00',
+		open: 170.525,
+		low: 170.07,
+		high: 170.525,
+		close: 170.16,
+		volume: 759554,
+	},
+	{
+		date: '2022-01-18 14:15:00',
+		open: 170.152,
+		low: 170.1248,
+		high: 170.531,
+		close: 170.52,
+		volume: 532493,
+	},
+	{
+		date: '2022-01-18 14:10:00',
+		open: 170.24,
+		low: 170.14,
+		high: 170.37,
+		close: 170.14,
+		volume: 442158,
+	},
+	{
+		date: '2022-01-18 14:05:00',
+		open: 170.35,
+		low: 170.14,
+		high: 170.45,
+		close: 170.24,
+		volume: 594741,
+	},
+	{
+		date: '2022-01-18 14:00:00',
+		open: 170.33,
+		low: 170.3,
+		high: 170.53,
+		close: 170.35,
+		volume: 669650,
+	},
+	{
+		date: '2022-01-18 13:55:00',
+		open: 170.0701,
+		low: 170.0526,
+		high: 170.44,
+		close: 170.345,
+		volume: 486767,
+	},
+	{
+		date: '2022-01-18 13:50:00',
+		open: 170.08,
+		low: 170.015,
+		high: 170.28,
+		close: 170.09,
+		volume: 517767,
+	},
+	{
+		date: '2022-01-18 13:45:00',
+		open: 170.22,
+		low: 170.0,
+		high: 170.36,
+		close: 170.08,
+		volume: 654971,
+	},
+	{
+		date: '2022-01-18 13:40:00',
+		open: 170.18,
+		low: 170.1,
+		high: 170.33,
+		close: 170.21,
+		volume: 607258,
+	},
+	{
+		date: '2022-01-18 13:35:00',
+		open: 169.91,
+		low: 169.88,
+		high: 170.2,
+		close: 170.1799,
+		volume: 746979,
+	},
+	{
+		date: '2022-01-18 13:30:00',
+		open: 169.75,
+		low: 169.685,
+		high: 169.94,
+		close: 169.91,
+		volume: 2155250,
+	},
+	{
+		date: '2022-01-18 13:25:00',
+		open: 169.8,
+		low: 169.63,
+		high: 169.82,
+		close: 169.755,
+		volume: 870094,
+	},
+	{
+		date: '2022-01-18 13:20:00',
+		open: 170.08,
+		low: 169.69,
+		high: 170.09,
+		close: 169.79,
+		volume: 584759,
+	},
+	{
+		date: '2022-01-18 13:15:00',
+		open: 170.065,
+		low: 169.9306,
+		high: 170.15,
+		close: 170.08,
+		volume: 659204,
+	},
+	{
+		date: '2022-01-18 13:10:00',
+		open: 169.8039,
+		low: 169.74,
+		high: 170.13,
+		close: 170.07,
+		volume: 741896,
+	},
+	{
+		date: '2022-01-18 13:05:00',
+		open: 169.73,
+		low: 169.6,
+		high: 169.85,
+		close: 169.815,
+		volume: 541942,
+	},
+	{
+		date: '2022-01-18 13:00:00',
+		open: 169.75,
+		low: 169.57,
+		high: 169.815,
+		close: 169.7312,
+		volume: 784417,
+	},
+	{
+		date: '2022-01-18 12:55:00',
+		open: 169.77,
+		low: 169.6,
+		high: 169.87,
+		close: 169.76,
+		volume: 634450,
+	},
+	{
+		date: '2022-01-18 12:50:00',
+		open: 169.87,
+		low: 169.7,
+		high: 169.96,
+		close: 169.765,
+		volume: 540053,
+	},
+	{
+		date: '2022-01-18 12:45:00',
+		open: 169.88,
+		low: 169.71,
+		high: 169.9999,
+		close: 169.86,
+		volume: 637964,
+	},
+	{
+		date: '2022-01-18 12:40:00',
+		open: 169.85,
+		low: 169.84,
+		high: 170.11,
+		close: 169.89,
+		volume: 663153,
+	},
+	{
+		date: '2022-01-18 12:35:00',
+		open: 170.0,
+		low: 169.78,
+		high: 170.0,
+		close: 169.855,
+		volume: 456211,
+	},
+	{
+		date: '2022-01-18 12:30:00',
+		open: 169.995,
+		low: 169.74,
+		high: 170.0406,
+		close: 170.0,
+		volume: 717228,
+	},
+	{
+		date: '2022-01-18 12:25:00',
+		open: 169.9101,
+		low: 169.78,
+		high: 170.01,
+		close: 170.0,
+		volume: 600045,
+	},
+	{
+		date: '2022-01-18 12:20:00',
+		open: 170.16,
+		low: 169.81,
+		high: 170.21,
+		close: 169.91,
+		volume: 1234460,
+	},
+	{
+		date: '2022-01-18 12:15:00',
+		open: 170.33,
+		low: 170.12,
+		high: 170.34,
+		close: 170.155,
+		volume: 690857,
+	},
+	{
+		date: '2022-01-18 12:10:00',
+		open: 170.405,
+		low: 170.14,
+		high: 170.47,
+		close: 170.34,
+		volume: 707865,
+	},
+	{
+		date: '2022-01-18 12:05:00',
+		open: 170.67,
+		low: 170.3803,
+		high: 170.67,
+		close: 170.41,
+		volume: 739727,
+	},
+	{
+		date: '2022-01-18 12:00:00',
+		open: 171.14,
+		low: 170.63,
+		high: 171.28,
+		close: 170.674,
+		volume: 743704,
+	},
+	{
+		date: '2022-01-18 11:55:00',
+		open: 171.12,
+		low: 171.0,
+		high: 171.215,
+		close: 171.1244,
+		volume: 521134,
+	},
+	{
+		date: '2022-01-18 11:50:00',
+		open: 170.97,
+		low: 170.94,
+		high: 171.22,
+		close: 171.13,
+		volume: 732806,
+	},
+	{
+		date: '2022-01-18 11:45:00',
+		open: 170.59,
+		low: 170.58,
+		high: 170.9899,
+		close: 170.97,
+		volume: 664562,
+	},
+	{
+		date: '2022-01-18 11:40:00',
+		open: 170.6286,
+		low: 170.56,
+		high: 170.78,
+		close: 170.6,
+		volume: 576399,
+	},
+	{
+		date: '2022-01-18 11:35:00',
+		open: 170.33,
+		low: 170.27,
+		high: 170.69,
+		close: 170.63,
+		volume: 1102364,
+	},
+	{
+		date: '2022-01-18 11:30:00',
+		open: 170.56,
+		low: 170.2,
+		high: 170.58,
+		close: 170.305,
+		volume: 932364,
+	},
+	{
+		date: '2022-01-18 11:25:00',
+		open: 170.66,
+		low: 170.5,
+		high: 170.77,
+		close: 170.5699,
+		volume: 749958,
+	},
+	{
+		date: '2022-01-18 11:20:00',
+		open: 170.69,
+		low: 170.56,
+		high: 170.86,
+		close: 170.645,
+		volume: 683936,
+	},
+	{
+		date: '2022-01-18 11:15:00',
+		open: 170.79,
+		low: 170.625,
+		high: 170.925,
+		close: 170.6901,
+		volume: 898231,
+	},
+	{
+		date: '2022-01-18 11:10:00',
+		open: 170.8164,
+		low: 170.645,
+		high: 170.96,
+		close: 170.79,
+		volume: 717271,
+	},
+	{
+		date: '2022-01-18 11:05:00',
+		open: 170.95,
+		low: 170.67,
+		high: 171.0872,
+		close: 170.8099,
+		volume: 803145,
+	},
+	{
+		date: '2022-01-18 11:00:00',
+		open: 171.1,
+		low: 170.86,
+		high: 171.1477,
+		close: 170.95,
+		volume: 734247,
+	},
+	{
+		date: '2022-01-18 10:55:00',
+		open: 171.11,
+		low: 170.99,
+		high: 171.34,
+		close: 171.0801,
+		volume: 842535,
+	},
+	{
+		date: '2022-01-18 10:50:00',
+		open: 171.46,
+		low: 171.1,
+		high: 171.46,
+		close: 171.1001,
+		volume: 749605,
+	},
+	{
+		date: '2022-01-18 10:45:00',
+		open: 171.995,
+		low: 171.43,
+		high: 172.11,
+		close: 171.44,
+		volume: 708931,
+	},
+	{
+		date: '2022-01-18 10:40:00',
+		open: 171.45,
+		low: 171.43,
+		high: 172.05,
+		close: 172.01,
+		volume: 791641,
+	},
+	{
+		date: '2022-01-18 10:35:00',
+		open: 171.66,
+		low: 171.4201,
+		high: 171.83,
+		close: 171.4516,
+		volume: 742502,
+	},
+	{
+		date: '2022-01-18 10:30:00',
+		open: 172.32,
+		low: 171.59,
+		high: 172.32,
+		close: 171.65,
+		volume: 816497,
+	},
+	{
+		date: '2022-01-18 10:25:00',
+		open: 172.07,
+		low: 171.94,
+		high: 172.32,
+		close: 172.31,
+		volume: 1084175,
+	},
+	{
+		date: '2022-01-18 10:20:00',
+		open: 171.82,
+		low: 171.65,
+		high: 172.14,
+		close: 172.07,
+		volume: 927999,
+	},
+	{
+		date: '2022-01-18 10:15:00',
+		open: 171.8,
+		low: 171.54,
+		high: 172.05,
+		close: 171.815,
+		volume: 1209835,
+	},
+	{
+		date: '2022-01-18 10:10:00',
+		open: 171.86,
+		low: 171.65,
+		high: 171.97,
+		close: 171.796,
+		volume: 1158991,
+	},
+	{
+		date: '2022-01-18 10:05:00',
+		open: 171.26,
+		low: 171.09,
+		high: 171.95,
+		close: 171.87,
+		volume: 1260297,
+	},
+	{
+		date: '2022-01-18 10:00:00',
+		open: 171.11,
+		low: 170.79,
+		high: 171.3225,
+		close: 171.24,
+		volume: 1215983,
+	},
+	{
+		date: '2022-01-18 09:55:00',
+		open: 170.87,
+		low: 170.59,
+		high: 171.22,
+		close: 171.11,
+		volume: 1163090,
+	},
+	{
+		date: '2022-01-18 09:50:00',
+		open: 170.86,
+		low: 170.64,
+		high: 171.24,
+		close: 170.879,
+		volume: 1396774,
+	},
+	{
+		date: '2022-01-18 09:45:00',
+		open: 172.18,
+		low: 170.75,
+		high: 172.24,
+		close: 170.88,
+		volume: 1942992,
+	},
+	{
+		date: '2022-01-18 09:40:00',
+		open: 171.26,
+		low: 171.12,
+		high: 172.4162,
+		close: 172.19,
+		volume: 2070080,
+	},
+	{
+		date: '2022-01-18 09:35:00',
+		open: 171.665,
+		low: 171.23,
+		high: 171.9,
+		close: 171.25,
+		volume: 1917290,
+	},
+	{
+		date: '2022-01-18 09:30:00',
+		open: 171.51,
+		low: 171.44,
+		high: 172.54,
+		close: 171.67,
+		volume: 4031243,
+	},
 ];
 
 //month / 1day
 export const stockHistory1Month = {
-  symbol: 'AAPL',
-  historical: [
-    {
-      date: '2022-01-14',
-      open: 171.34,
-      high: 173.78,
-      low: 171.09,
-      close: 173.07,
-      adjClose: 173.07,
-      volume: 80357482,
-      unadjustedVolume: 80357482,
-      change: 1.73,
-      changePercent: 1.01,
-      vwap: 172.64667,
-      label: 'January 14, 22',
-      changeOverTime: 0.0101,
-    },
-    {
-      date: '2022-01-13',
-      open: 175.78,
-      high: 176.62,
-      low: 171.79,
-      close: 172.19,
-      adjClose: 172.19,
-      volume: 81576195,
-      unadjustedVolume: 81576195,
-      change: -3.59,
-      changePercent: -2.042,
-      vwap: 173.53333,
-      label: 'January 13, 22',
-      changeOverTime: -0.02042,
-    },
-    {
-      date: '2022-01-12',
-      open: 176.12,
-      high: 177.179,
-      low: 174.821,
-      close: 175.53,
-      adjClose: 175.53,
-      volume: 70583410,
-      unadjustedVolume: 70583410,
-      change: -0.59,
-      changePercent: -0.335,
-      vwap: 175.84333,
-      label: 'January 12, 22',
-      changeOverTime: -0.00335,
-    },
-    {
-      date: '2022-01-11',
-      open: 172.32,
-      high: 175.18,
-      low: 170.82,
-      close: 175.08,
-      adjClose: 175.08,
-      volume: 76021945,
-      unadjustedVolume: 76021945,
-      change: 2.76,
-      changePercent: 1.602,
-      vwap: 173.69333,
-      label: 'January 11, 22',
-      changeOverTime: 0.01602,
-    },
-    {
-      date: '2022-01-10',
-      open: 169.08,
-      high: 172.5,
-      low: 168.17,
-      close: 172.19,
-      adjClose: 172.19,
-      volume: 106616370,
-      unadjustedVolume: 106616370,
-      change: 3.11,
-      changePercent: 1.839,
-      vwap: 170.95333,
-      label: 'January 10, 22',
-      changeOverTime: 0.01839,
-    },
-    {
-      date: '2022-01-07',
-      open: 172.89,
-      high: 174.14,
-      low: 171.03,
-      close: 172.17,
-      adjClose: 172.17,
-      volume: 85121480,
-      unadjustedVolume: 85121480,
-      change: -0.72,
-      changePercent: -0.416,
-      vwap: 172.44667,
-      label: 'January 07, 22',
-      changeOverTime: -0.00416,
-    },
-    {
-      date: '2022-01-06',
-      open: 172.7,
-      high: 175.3,
-      low: 171.64,
-      close: 172.0,
-      adjClose: 172.0,
-      volume: 96758703,
-      unadjustedVolume: 96758703,
-      change: -0.7,
-      changePercent: -0.405,
-      vwap: 172.98,
-      label: 'January 06, 22',
-      changeOverTime: -0.00405,
-    },
-    {
-      date: '2022-01-05',
-      open: 179.61,
-      high: 180.17,
-      low: 174.64,
-      close: 174.92,
-      adjClose: 174.92,
-      volume: 94219974,
-      unadjustedVolume: 94219974,
-      change: -4.69,
-      changePercent: -2.611,
-      vwap: 176.57667,
-      label: 'January 05, 22',
-      changeOverTime: -0.02611,
-    },
-    {
-      date: '2022-01-04',
-      open: 182.63,
-      high: 182.9,
-      low: 179.12,
-      close: 179.7,
-      adjClose: 179.7,
-      volume: 99310666,
-      unadjustedVolume: 99310666,
-      change: -2.93,
-      changePercent: -1.604,
-      vwap: 180.57333,
-      label: 'January 04, 22',
-      changeOverTime: -0.01604,
-    },
-    {
-      date: '2022-01-03',
-      open: 177.83,
-      high: 182.88,
-      low: 177.71,
-      close: 182.01,
-      adjClose: 182.01,
-      volume: 104497653,
-      unadjustedVolume: 104497653,
-      change: 4.18,
-      changePercent: 2.351,
-      vwap: 180.86667,
-      label: 'January 03, 22',
-      changeOverTime: 0.02351,
-    },
-    {
-      date: '2021-12-31',
-      open: 178.085,
-      high: 179.23,
-      low: 177.26,
-      close: 177.57,
-      adjClose: 177.57,
-      volume: 59615961,
-      unadjustedVolume: 59615961,
-      change: -0.515,
-      changePercent: -0.289,
-      vwap: 178.02,
-      label: 'December 31, 21',
-      changeOverTime: -0.00289,
-    },
-    {
-      date: '2021-12-30',
-      open: 179.47,
-      high: 180.57,
-      low: 178.09,
-      close: 178.2,
-      adjClose: 178.2,
-      volume: 59620506,
-      unadjustedVolume: 59620506,
-      change: -1.27,
-      changePercent: -0.708,
-      vwap: 178.95333,
-      label: 'December 30, 21',
-      changeOverTime: -0.00708,
-    },
-    {
-      date: '2021-12-29',
-      open: 179.33,
-      high: 180.63,
-      low: 178.14,
-      close: 179.38,
-      adjClose: 179.38,
-      volume: 62234590,
-      unadjustedVolume: 62234590,
-      change: 0.05,
-      changePercent: 0.028,
-      vwap: 179.38333,
-      label: 'December 29, 21',
-      changeOverTime: 0.00028,
-    },
-    {
-      date: '2021-12-28',
-      open: 180.16,
-      high: 181.33,
-      low: 178.53,
-      close: 179.29,
-      adjClose: 179.29,
-      volume: 76429131,
-      unadjustedVolume: 76429131,
-      change: -0.87,
-      changePercent: -0.483,
-      vwap: 179.71667,
-      label: 'December 28, 21',
-      changeOverTime: -0.00483,
-    },
-    {
-      date: '2021-12-27',
-      open: 177.085,
-      high: 180.42,
-      low: 177.07,
-      close: 180.33,
-      adjClose: 180.33,
-      volume: 73007326,
-      unadjustedVolume: 73007326,
-      change: 3.245,
-      changePercent: 1.832,
-      vwap: 179.27333,
-      label: 'December 27, 21',
-      changeOverTime: 0.01832,
-    },
-    {
-      date: '2021-12-23',
-      open: 175.85,
-      high: 176.85,
-      low: 175.27,
-      close: 176.28,
-      adjClose: 176.28,
-      volume: 68234171,
-      unadjustedVolume: 68234171,
-      change: 0.43,
-      changePercent: 0.245,
-      vwap: 176.13333,
-      label: 'December 23, 21',
-      changeOverTime: 0.00245,
-    },
-    {
-      date: '2021-12-22',
-      open: 173.04,
-      high: 175.86,
-      low: 172.15,
-      close: 175.64,
-      adjClose: 175.64,
-      volume: 92011893,
-      unadjustedVolume: 92011893,
-      change: 2.6,
-      changePercent: 1.503,
-      vwap: 174.55,
-      label: 'December 22, 21',
-      changeOverTime: 0.01503,
-    },
-    {
-      date: '2021-12-21',
-      open: 171.56,
-      high: 173.2,
-      low: 169.12,
-      close: 172.99,
-      adjClose: 172.99,
-      volume: 85851116,
-      unadjustedVolume: 85851116,
-      change: 1.43,
-      changePercent: 0.834,
-      vwap: 171.77,
-      label: 'December 21, 21',
-      changeOverTime: 0.00834,
-    },
-    {
-      date: '2021-12-20',
-      open: 168.28,
-      high: 170.58,
-      low: 167.46,
-      close: 169.75,
-      adjClose: 169.75,
-      volume: 107314653,
-      unadjustedVolume: 107314653,
-      change: 1.47,
-      changePercent: 0.874,
-      vwap: 169.26333,
-      label: 'December 20, 21',
-      changeOverTime: 0.00874,
-    },
-    {
-      date: '2021-12-17',
-      open: 169.93,
-      high: 173.47,
-      low: 169.69,
-      close: 171.14,
-      adjClose: 171.14,
-      volume: 195450953,
-      unadjustedVolume: 195450953,
-      change: 1.21,
-      changePercent: 0.712,
-      vwap: 171.43333,
-      label: 'December 17, 21',
-      changeOverTime: 0.00712,
-    },
-    {
-      date: '2021-12-16',
-      open: 179.28,
-      high: 181.14,
-      low: 170.75,
-      close: 172.26,
-      adjClose: 172.26,
-      volume: 149977388,
-      unadjustedVolume: 149977388,
-      change: -7.02,
-      changePercent: -3.916,
-      vwap: 174.71667,
-      label: 'December 16, 21',
-      changeOverTime: -0.03916,
-    },
-    {
-      date: '2021-12-15',
-      open: 175.11,
-      high: 179.5,
-      low: 172.31,
-      close: 179.3,
-      adjClose: 179.3,
-      volume: 130698555,
-      unadjustedVolume: 130698555,
-      change: 4.19,
-      changePercent: 2.393,
-      vwap: 177.03667,
-      label: 'December 15, 21',
-      changeOverTime: 0.02393,
-    },
-  ],
+	symbol: 'AAPL',
+	historical: [
+		{
+			date: '2022-01-14',
+			open: 171.34,
+			high: 173.78,
+			low: 171.09,
+			close: 173.07,
+			adjClose: 173.07,
+			volume: 80357482,
+			unadjustedVolume: 80357482,
+			change: 1.73,
+			changePercent: 1.01,
+			vwap: 172.64667,
+			label: 'January 14, 22',
+			changeOverTime: 0.0101,
+		},
+		{
+			date: '2022-01-13',
+			open: 175.78,
+			high: 176.62,
+			low: 171.79,
+			close: 172.19,
+			adjClose: 172.19,
+			volume: 81576195,
+			unadjustedVolume: 81576195,
+			change: -3.59,
+			changePercent: -2.042,
+			vwap: 173.53333,
+			label: 'January 13, 22',
+			changeOverTime: -0.02042,
+		},
+		{
+			date: '2022-01-12',
+			open: 176.12,
+			high: 177.179,
+			low: 174.821,
+			close: 175.53,
+			adjClose: 175.53,
+			volume: 70583410,
+			unadjustedVolume: 70583410,
+			change: -0.59,
+			changePercent: -0.335,
+			vwap: 175.84333,
+			label: 'January 12, 22',
+			changeOverTime: -0.00335,
+		},
+		{
+			date: '2022-01-11',
+			open: 172.32,
+			high: 175.18,
+			low: 170.82,
+			close: 175.08,
+			adjClose: 175.08,
+			volume: 76021945,
+			unadjustedVolume: 76021945,
+			change: 2.76,
+			changePercent: 1.602,
+			vwap: 173.69333,
+			label: 'January 11, 22',
+			changeOverTime: 0.01602,
+		},
+		{
+			date: '2022-01-10',
+			open: 169.08,
+			high: 172.5,
+			low: 168.17,
+			close: 172.19,
+			adjClose: 172.19,
+			volume: 106616370,
+			unadjustedVolume: 106616370,
+			change: 3.11,
+			changePercent: 1.839,
+			vwap: 170.95333,
+			label: 'January 10, 22',
+			changeOverTime: 0.01839,
+		},
+		{
+			date: '2022-01-07',
+			open: 172.89,
+			high: 174.14,
+			low: 171.03,
+			close: 172.17,
+			adjClose: 172.17,
+			volume: 85121480,
+			unadjustedVolume: 85121480,
+			change: -0.72,
+			changePercent: -0.416,
+			vwap: 172.44667,
+			label: 'January 07, 22',
+			changeOverTime: -0.00416,
+		},
+		{
+			date: '2022-01-06',
+			open: 172.7,
+			high: 175.3,
+			low: 171.64,
+			close: 172.0,
+			adjClose: 172.0,
+			volume: 96758703,
+			unadjustedVolume: 96758703,
+			change: -0.7,
+			changePercent: -0.405,
+			vwap: 172.98,
+			label: 'January 06, 22',
+			changeOverTime: -0.00405,
+		},
+		{
+			date: '2022-01-05',
+			open: 179.61,
+			high: 180.17,
+			low: 174.64,
+			close: 174.92,
+			adjClose: 174.92,
+			volume: 94219974,
+			unadjustedVolume: 94219974,
+			change: -4.69,
+			changePercent: -2.611,
+			vwap: 176.57667,
+			label: 'January 05, 22',
+			changeOverTime: -0.02611,
+		},
+		{
+			date: '2022-01-04',
+			open: 182.63,
+			high: 182.9,
+			low: 179.12,
+			close: 179.7,
+			adjClose: 179.7,
+			volume: 99310666,
+			unadjustedVolume: 99310666,
+			change: -2.93,
+			changePercent: -1.604,
+			vwap: 180.57333,
+			label: 'January 04, 22',
+			changeOverTime: -0.01604,
+		},
+		{
+			date: '2022-01-03',
+			open: 177.83,
+			high: 182.88,
+			low: 177.71,
+			close: 182.01,
+			adjClose: 182.01,
+			volume: 104497653,
+			unadjustedVolume: 104497653,
+			change: 4.18,
+			changePercent: 2.351,
+			vwap: 180.86667,
+			label: 'January 03, 22',
+			changeOverTime: 0.02351,
+		},
+		{
+			date: '2021-12-31',
+			open: 178.085,
+			high: 179.23,
+			low: 177.26,
+			close: 177.57,
+			adjClose: 177.57,
+			volume: 59615961,
+			unadjustedVolume: 59615961,
+			change: -0.515,
+			changePercent: -0.289,
+			vwap: 178.02,
+			label: 'December 31, 21',
+			changeOverTime: -0.00289,
+		},
+		{
+			date: '2021-12-30',
+			open: 179.47,
+			high: 180.57,
+			low: 178.09,
+			close: 178.2,
+			adjClose: 178.2,
+			volume: 59620506,
+			unadjustedVolume: 59620506,
+			change: -1.27,
+			changePercent: -0.708,
+			vwap: 178.95333,
+			label: 'December 30, 21',
+			changeOverTime: -0.00708,
+		},
+		{
+			date: '2021-12-29',
+			open: 179.33,
+			high: 180.63,
+			low: 178.14,
+			close: 179.38,
+			adjClose: 179.38,
+			volume: 62234590,
+			unadjustedVolume: 62234590,
+			change: 0.05,
+			changePercent: 0.028,
+			vwap: 179.38333,
+			label: 'December 29, 21',
+			changeOverTime: 0.00028,
+		},
+		{
+			date: '2021-12-28',
+			open: 180.16,
+			high: 181.33,
+			low: 178.53,
+			close: 179.29,
+			adjClose: 179.29,
+			volume: 76429131,
+			unadjustedVolume: 76429131,
+			change: -0.87,
+			changePercent: -0.483,
+			vwap: 179.71667,
+			label: 'December 28, 21',
+			changeOverTime: -0.00483,
+		},
+		{
+			date: '2021-12-27',
+			open: 177.085,
+			high: 180.42,
+			low: 177.07,
+			close: 180.33,
+			adjClose: 180.33,
+			volume: 73007326,
+			unadjustedVolume: 73007326,
+			change: 3.245,
+			changePercent: 1.832,
+			vwap: 179.27333,
+			label: 'December 27, 21',
+			changeOverTime: 0.01832,
+		},
+		{
+			date: '2021-12-23',
+			open: 175.85,
+			high: 176.85,
+			low: 175.27,
+			close: 176.28,
+			adjClose: 176.28,
+			volume: 68234171,
+			unadjustedVolume: 68234171,
+			change: 0.43,
+			changePercent: 0.245,
+			vwap: 176.13333,
+			label: 'December 23, 21',
+			changeOverTime: 0.00245,
+		},
+		{
+			date: '2021-12-22',
+			open: 173.04,
+			high: 175.86,
+			low: 172.15,
+			close: 175.64,
+			adjClose: 175.64,
+			volume: 92011893,
+			unadjustedVolume: 92011893,
+			change: 2.6,
+			changePercent: 1.503,
+			vwap: 174.55,
+			label: 'December 22, 21',
+			changeOverTime: 0.01503,
+		},
+		{
+			date: '2021-12-21',
+			open: 171.56,
+			high: 173.2,
+			low: 169.12,
+			close: 172.99,
+			adjClose: 172.99,
+			volume: 85851116,
+			unadjustedVolume: 85851116,
+			change: 1.43,
+			changePercent: 0.834,
+			vwap: 171.77,
+			label: 'December 21, 21',
+			changeOverTime: 0.00834,
+		},
+		{
+			date: '2021-12-20',
+			open: 168.28,
+			high: 170.58,
+			low: 167.46,
+			close: 169.75,
+			adjClose: 169.75,
+			volume: 107314653,
+			unadjustedVolume: 107314653,
+			change: 1.47,
+			changePercent: 0.874,
+			vwap: 169.26333,
+			label: 'December 20, 21',
+			changeOverTime: 0.00874,
+		},
+		{
+			date: '2021-12-17',
+			open: 169.93,
+			high: 173.47,
+			low: 169.69,
+			close: 171.14,
+			adjClose: 171.14,
+			volume: 195450953,
+			unadjustedVolume: 195450953,
+			change: 1.21,
+			changePercent: 0.712,
+			vwap: 171.43333,
+			label: 'December 17, 21',
+			changeOverTime: 0.00712,
+		},
+		{
+			date: '2021-12-16',
+			open: 179.28,
+			high: 181.14,
+			low: 170.75,
+			close: 172.26,
+			adjClose: 172.26,
+			volume: 149977388,
+			unadjustedVolume: 149977388,
+			change: -7.02,
+			changePercent: -3.916,
+			vwap: 174.71667,
+			label: 'December 16, 21',
+			changeOverTime: -0.03916,
+		},
+		{
+			date: '2021-12-15',
+			open: 175.11,
+			high: 179.5,
+			low: 172.31,
+			close: 179.3,
+			adjClose: 179.3,
+			volume: 130698555,
+			unadjustedVolume: 130698555,
+			change: 4.19,
+			changePercent: 2.393,
+			vwap: 177.03667,
+			label: 'December 15, 21',
+			changeOverTime: 0.02393,
+		},
+	],
 };
 
 //year / 1day
 export const stockHistory1Year = {
-  symbol: 'AAPL',
-  historical: [
-    {
-      date: '2019-03-12',
-      open: 45.0,
-      high: 45.6675,
-      low: 44.842499,
-      close: 45.227501,
-      adjClose: 44.094143,
-      volume: 129870400,
-      unadjustedVolume: 129870400,
-      change: 0.2275,
-      changePercent: 0.506,
-      vwap: 45.24583,
-      label: 'March 12, 19',
-      changeOverTime: 0.00506,
-    },
-    {
-      date: '2019-03-11',
-      open: 43.872501,
-      high: 44.779999,
-      low: 43.837502,
-      close: 44.724998,
-      adjClose: 43.604237,
-      volume: 128044000,
-      unadjustedVolume: 128044000,
-      change: 0.8525,
-      changePercent: 1.943,
-      vwap: 44.4475,
-      label: 'March 11, 19',
-      changeOverTime: 0.01943,
-    },
-    {
-      date: '2019-03-08',
-      open: 42.580002,
-      high: 43.267502,
-      low: 42.375,
-      close: 43.227501,
-      adjClose: 42.14426,
-      volume: 95997600,
-      unadjustedVolume: 95997600,
-      change: 0.6475,
-      changePercent: 1.521,
-      vwap: 42.95667,
-      label: 'March 08, 19',
-      changeOverTime: 0.01521,
-    },
-    {
-      date: '2019-03-07',
-      open: 43.467499,
-      high: 43.610001,
-      low: 43.005001,
-      close: 43.125,
-      adjClose: 42.044323,
-      volume: 99185600,
-      unadjustedVolume: 99185600,
-      change: -0.3425,
-      changePercent: -0.788,
-      vwap: 43.24667,
-      label: 'March 07, 19',
-      changeOverTime: -0.00788,
-    },
-    {
-      date: '2019-03-06',
-      open: 43.6675,
-      high: 43.872501,
-      low: 43.485001,
-      close: 43.630001,
-      adjClose: 42.536674,
-      volume: 83241600,
-      unadjustedVolume: 83241600,
-      change: -0.0375,
-      changePercent: -0.086,
-      vwap: 43.6625,
-      label: 'March 06, 19',
-      changeOverTime: -0.00086,
-    },
-    {
-      date: '2019-03-05',
-      open: 43.985001,
-      high: 44.0,
-      low: 43.634998,
-      close: 43.8825,
-      adjClose: 42.782837,
-      volume: 78949600,
-      unadjustedVolume: 78949600,
-      change: -0.1025,
-      changePercent: -0.233,
-      vwap: 43.83917,
-      label: 'March 05, 19',
-      changeOverTime: -0.00233,
-    },
-    {
-      date: '2019-03-04',
-      open: 43.922501,
-      high: 44.4375,
-      low: 43.4925,
-      close: 43.962502,
-      adjClose: 42.86084,
-      volume: 109744800,
-      unadjustedVolume: 109744800,
-      change: 0.04,
-      changePercent: 0.091,
-      vwap: 43.96417,
-      label: 'March 04, 19',
-      changeOverTime: 0.00091,
-    },
-    {
-      date: '2019-03-01',
-      open: 43.57,
-      high: 43.787498,
-      low: 43.2225,
-      close: 43.7425,
-      adjClose: 42.646358,
-      volume: 103544800,
-      unadjustedVolume: 103544800,
-      change: 0.1725,
-      changePercent: 0.396,
-      vwap: 43.58417,
-      label: 'March 01, 19',
-      changeOverTime: 0.00396,
-    },
-    {
-      date: '2019-02-28',
-      open: 43.580002,
-      high: 43.727501,
-      low: 43.23,
-      close: 43.287498,
-      adjClose: 42.202747,
-      volume: 112861600,
-      unadjustedVolume: 112861600,
-      change: -0.2925,
-      changePercent: -0.671,
-      vwap: 43.415,
-      label: 'February 28, 19',
-      changeOverTime: -0.00671,
-    },
-    {
-      date: '2019-02-27',
-      open: 43.302502,
-      high: 43.75,
-      low: 43.182499,
-      close: 43.717499,
-      adjClose: 42.621967,
-      volume: 111341600,
-      unadjustedVolume: 111341600,
-      change: 0.415,
-      changePercent: 0.958,
-      vwap: 43.55,
-      label: 'February 27, 19',
-      changeOverTime: 0.00958,
-    },
-    {
-      date: '2019-02-26',
-      open: 43.427502,
-      high: 43.825001,
-      low: 43.2925,
-      close: 43.5825,
-      adjClose: 42.490368,
-      volume: 68280800,
-      unadjustedVolume: 68280800,
-      change: 0.155,
-      changePercent: 0.357,
-      vwap: 43.56667,
-      label: 'February 26, 19',
-      changeOverTime: 0.00357,
-    },
-    {
-      date: '2019-02-25',
-      open: 43.540001,
-      high: 43.967499,
-      low: 43.487499,
-      close: 43.557499,
-      adjClose: 42.465984,
-      volume: 87493600,
-      unadjustedVolume: 87493600,
-      change: 0.0175,
-      changePercent: 0.04,
-      vwap: 43.67083,
-      label: 'February 25, 19',
-      changeOverTime: 0.0004,
-    },
-    {
-      date: '2019-02-22',
-      open: 42.895,
-      high: 43.25,
-      low: 42.845001,
-      close: 43.2425,
-      adjClose: 42.158882,
-      volume: 75652800,
-      unadjustedVolume: 75652800,
-      change: 0.3475,
-      changePercent: 0.81,
-      vwap: 43.1125,
-      label: 'February 22, 19',
-      changeOverTime: 0.0081,
-    },
-    {
-      date: '2019-02-21',
-      open: 42.950001,
-      high: 43.092499,
-      low: 42.575001,
-      close: 42.764999,
-      adjClose: 41.693352,
-      volume: 68998800,
-      unadjustedVolume: 68998800,
-      change: -0.185,
-      changePercent: -0.431,
-      vwap: 42.81083,
-      label: 'February 21, 19',
-      changeOverTime: -0.00431,
-    },
-    {
-      date: '2019-02-20',
-      open: 42.797501,
-      high: 43.330002,
-      low: 42.747501,
-      close: 43.0075,
-      adjClose: 41.929768,
-      volume: 104457600,
-      unadjustedVolume: 104457600,
-      change: 0.21,
-      changePercent: 0.491,
-      vwap: 43.02833,
-      label: 'February 20, 19',
-      changeOverTime: 0.00491,
-    },
-    {
-      date: '2019-02-19',
-      open: 42.427502,
-      high: 42.860001,
-      low: 42.372501,
-      close: 42.732498,
-      adjClose: 41.661659,
-      volume: 75891200,
-      unadjustedVolume: 75891200,
-      change: 0.305,
-      changePercent: 0.719,
-      vwap: 42.655,
-      label: 'February 19, 19',
-      changeOverTime: 0.00719,
-    },
-    {
-      date: '2019-02-15',
-      open: 42.8125,
-      high: 42.924999,
-      low: 42.4375,
-      close: 42.605,
-      adjClose: 41.537365,
-      volume: 98507200,
-      unadjustedVolume: 98507200,
-      change: -0.2075,
-      changePercent: -0.485,
-      vwap: 42.65583,
-      label: 'February 15, 19',
-      changeOverTime: -0.00485,
-    },
-    {
-      date: '2019-02-14',
-      open: 42.427502,
-      high: 42.814999,
-      low: 42.345001,
-      close: 42.700001,
-      adjClose: 41.629974,
-      volume: 87342800,
-      unadjustedVolume: 87342800,
-      change: 0.2725,
-      changePercent: 0.642,
-      vwap: 42.62,
-      label: 'February 14, 19',
-      changeOverTime: 0.00642,
-    },
-    {
-      date: '2019-02-13',
-      open: 42.8475,
-      high: 43.119999,
-      low: 42.48,
-      close: 42.544998,
-      adjClose: 41.478859,
-      volume: 89960800,
-      unadjustedVolume: 89960800,
-      change: -0.3025,
-      changePercent: -0.706,
-      vwap: 42.715,
-      label: 'February 13, 19',
-      changeOverTime: -0.00706,
-    },
-    {
-      date: '2019-02-12',
-      open: 42.525002,
-      high: 42.75,
-      low: 42.424999,
-      close: 42.7225,
-      adjClose: 41.651917,
-      volume: 89134000,
-      unadjustedVolume: 89134000,
-      change: 0.1975,
-      changePercent: 0.464,
-      vwap: 42.6325,
-      label: 'February 12, 19',
-      changeOverTime: 0.00464,
-    },
-    {
-      date: '2019-02-11',
-      open: 42.762501,
-      high: 42.802502,
-      low: 42.3125,
-      close: 42.357498,
-      adjClose: 41.296066,
-      volume: 83973600,
-      unadjustedVolume: 83973600,
-      change: -0.405,
-      changePercent: -0.947,
-      vwap: 42.49083,
-      label: 'February 11, 19',
-      changeOverTime: -0.00947,
-    },
-    {
-      date: '2019-02-08',
-      open: 42.247501,
-      high: 42.665001,
-      low: 42.105,
-      close: 42.602501,
-      adjClose: 41.53492,
-      volume: 95280000,
-      unadjustedVolume: 95280000,
-      change: 0.355,
-      changePercent: 0.84,
-      vwap: 42.4575,
-      label: 'February 08, 19',
-      changeOverTime: 0.0084,
-    },
-    {
-      date: '2019-02-07',
-      open: 43.099998,
-      high: 43.485001,
-      low: 42.584999,
-      close: 42.735001,
-      adjClose: 41.486176,
-      volume: 126966800,
-      unadjustedVolume: 126966800,
-      change: -0.365,
-      changePercent: -0.847,
-      vwap: 42.935,
-      label: 'February 07, 19',
-      changeOverTime: -0.00847,
-    },
-    {
-      date: '2019-02-06',
-      open: 43.662498,
-      high: 43.892502,
-      low: 43.212502,
-      close: 43.560001,
-      adjClose: 42.28706,
-      volume: 112958400,
-      unadjustedVolume: 112958400,
-      change: -0.1025,
-      changePercent: -0.235,
-      vwap: 43.555,
-      label: 'February 06, 19',
-      changeOverTime: -0.00235,
-    },
-    {
-      date: '2019-02-05',
-      open: 43.215,
-      high: 43.77,
-      low: 43.087502,
-      close: 43.544998,
-      adjClose: 42.272499,
-      volume: 144406400,
-      unadjustedVolume: 144406400,
-      change: 0.33,
-      changePercent: 0.764,
-      vwap: 43.4675,
-      label: 'February 05, 19',
-      changeOverTime: 0.00764,
-    },
-    {
-      date: '2019-02-04',
-      open: 41.852501,
-      high: 42.915001,
-      low: 41.82,
-      close: 42.8125,
-      adjClose: 41.561405,
-      volume: 125982000,
-      unadjustedVolume: 125982000,
-      change: 0.96,
-      changePercent: 2.294,
-      vwap: 42.51583,
-      label: 'February 04, 19',
-      changeOverTime: 0.02294,
-    },
-    {
-      date: '2019-02-01',
-      open: 41.740002,
-      high: 42.244999,
-      low: 41.482498,
-      close: 41.630001,
-      adjClose: 40.413464,
-      volume: 130672400,
-      unadjustedVolume: 130672400,
-      change: -0.11,
-      changePercent: -0.264,
-      vwap: 41.78583,
-      label: 'February 01, 19',
-      changeOverTime: -0.00264,
-    },
-    {
-      date: '2019-01-31',
-      open: 41.5275,
-      high: 42.25,
-      low: 41.139999,
-      close: 41.610001,
-      adjClose: 40.394051,
-      volume: 162958400,
-      unadjustedVolume: 162958400,
-      change: 0.0825,
-      changePercent: 0.199,
-      vwap: 41.66667,
-      label: 'January 31, 19',
-      changeOverTime: 0.00199,
-    },
-    {
-      date: '2019-01-30',
-      open: 40.8125,
-      high: 41.537498,
-      low: 40.057499,
-      close: 41.3125,
-      adjClose: 40.10524,
-      volume: 244439200,
-      unadjustedVolume: 244439200,
-      change: 0.5,
-      changePercent: 1.225,
-      vwap: 40.96917,
-      label: 'January 30, 19',
-      changeOverTime: 0.01225,
-    },
-    {
-      date: '2019-01-29',
-      open: 39.0625,
-      high: 39.532501,
-      low: 38.5275,
-      close: 38.669998,
-      adjClose: 37.539959,
-      volume: 166348800,
-      unadjustedVolume: 166348800,
-      change: -0.3925,
-      changePercent: -1.005,
-      vwap: 38.91,
-      label: 'January 29, 19',
-      changeOverTime: -0.01005,
-    },
-    {
-      date: '2019-01-28',
-      open: 38.947498,
-      high: 39.0825,
-      low: 38.415001,
-      close: 39.075001,
-      adjClose: 37.933128,
-      volume: 104768400,
-      unadjustedVolume: 104768400,
-      change: 0.1275,
-      changePercent: 0.327,
-      vwap: 38.8575,
-      label: 'January 28, 19',
-      changeOverTime: 0.00327,
-    },
-    {
-      date: '2019-01-25',
-      open: 38.869999,
-      high: 39.532501,
-      low: 38.580002,
-      close: 39.439999,
-      adjClose: 38.287464,
-      volume: 134142000,
-      unadjustedVolume: 134142000,
-      change: 0.57,
-      changePercent: 1.466,
-      vwap: 39.18417,
-      label: 'January 25, 19',
-      changeOverTime: 0.01466,
-    },
-    {
-      date: '2019-01-24',
-      open: 38.5275,
-      high: 38.619999,
-      low: 37.935001,
-      close: 38.174999,
-      adjClose: 37.059425,
-      volume: 101766000,
-      unadjustedVolume: 101766000,
-      change: -0.3525,
-      changePercent: -0.915,
-      vwap: 38.24333,
-      label: 'January 24, 19',
-      changeOverTime: -0.00915,
-    },
-    {
-      date: '2019-01-23',
-      open: 38.537498,
-      high: 38.785,
-      low: 37.924999,
-      close: 38.48,
-      adjClose: 37.355515,
-      volume: 92522400,
-      unadjustedVolume: 92522400,
-      change: -0.0575,
-      changePercent: -0.149,
-      vwap: 38.39667,
-      label: 'January 23, 19',
-      changeOverTime: -0.00149,
-    },
-    {
-      date: '2019-01-22',
-      open: 39.102501,
-      high: 39.182499,
-      low: 38.154999,
-      close: 38.325001,
-      adjClose: 37.205048,
-      volume: 121576000,
-      unadjustedVolume: 121576000,
-      change: -0.7775,
-      changePercent: -1.988,
-      vwap: 38.55417,
-      label: 'January 22, 19',
-      changeOverTime: -0.01988,
-    },
-    {
-      date: '2019-01-18',
-      open: 39.375,
-      high: 39.470001,
-      low: 38.994999,
-      close: 39.205002,
-      adjClose: 38.059334,
-      volume: 135004000,
-      unadjustedVolume: 135004000,
-      change: -0.17,
-      changePercent: -0.432,
-      vwap: 39.22333,
-      label: 'January 18, 19',
-      changeOverTime: -0.00432,
-    },
-    {
-      date: '2019-01-17',
-      open: 38.549999,
-      high: 39.415001,
-      low: 38.314999,
-      close: 38.965,
-      adjClose: 37.826347,
-      volume: 119284800,
-      unadjustedVolume: 119284800,
-      change: 0.415,
-      changePercent: 1.077,
-      vwap: 38.89833,
-      label: 'January 17, 19',
-      changeOverTime: 0.01077,
-    },
-    {
-      date: '2019-01-16',
-      open: 38.27,
-      high: 38.970001,
-      low: 38.25,
-      close: 38.735001,
-      adjClose: 37.603065,
-      volume: 122278800,
-      unadjustedVolume: 122278800,
-      change: 0.465,
-      changePercent: 1.215,
-      vwap: 38.65167,
-      label: 'January 16, 19',
-      changeOverTime: 0.01215,
-    },
-    {
-      date: '2019-01-15',
-      open: 37.567501,
-      high: 38.3475,
-      low: 37.512501,
-      close: 38.267502,
-      adjClose: 37.149231,
-      volume: 114843600,
-      unadjustedVolume: 114843600,
-      change: 0.7,
-      changePercent: 1.863,
-      vwap: 38.0425,
-      label: 'January 15, 19',
-      changeOverTime: 0.01863,
-    },
-    {
-      date: '2019-01-14',
-      open: 37.712502,
-      high: 37.817501,
-      low: 37.305,
-      close: 37.5,
-      adjClose: 36.404152,
-      volume: 129756800,
-      unadjustedVolume: 129756800,
-      change: -0.2125,
-      changePercent: -0.563,
-      vwap: 37.54083,
-      label: 'January 14, 19',
-      changeOverTime: -0.00563,
-    },
-    {
-      date: '2019-01-11',
-      open: 38.220001,
-      high: 38.424999,
-      low: 37.877499,
-      close: 38.072498,
-      adjClose: 36.959927,
-      volume: 108092800,
-      unadjustedVolume: 108092800,
-      change: -0.1475,
-      changePercent: -0.386,
-      vwap: 38.125,
-      label: 'January 11, 19',
-      changeOverTime: -0.00386,
-    },
-    {
-      date: '2019-01-10',
-      open: 38.125,
-      high: 38.4925,
-      low: 37.715,
-      close: 38.450001,
-      adjClose: 37.326393,
-      volume: 143122800,
-      unadjustedVolume: 143122800,
-      change: 0.325,
-      changePercent: 0.852,
-      vwap: 38.21917,
-      label: 'January 10, 19',
-      changeOverTime: 0.00852,
-    },
-    {
-      date: '2019-01-09',
-      open: 37.822498,
-      high: 38.6325,
-      low: 37.407501,
-      close: 38.327499,
-      adjClose: 37.20747,
-      volume: 180396400,
-      unadjustedVolume: 180396400,
-      change: 0.505,
-      changePercent: 1.335,
-      vwap: 38.1225,
-      label: 'January 09, 19',
-      changeOverTime: 0.01335,
-    },
-    {
-      date: '2019-01-08',
-      open: 37.389999,
-      high: 37.955002,
-      low: 37.130001,
-      close: 37.6875,
-      adjClose: 36.586174,
-      volume: 164101200,
-      unadjustedVolume: 164101200,
-      change: 0.2975,
-      changePercent: 0.796,
-      vwap: 37.59083,
-      label: 'January 08, 19',
-      changeOverTime: 0.00796,
-    },
-    {
-      date: '2019-01-07',
-      open: 37.174999,
-      high: 37.2075,
-      low: 36.474998,
-      close: 36.982498,
-      adjClose: 35.901775,
-      volume: 219111200,
-      unadjustedVolume: 219111200,
-      change: -0.1925,
-      changePercent: -0.518,
-      vwap: 36.88833,
-      label: 'January 07, 19',
-      changeOverTime: -0.00518,
-    },
-    {
-      date: '2019-01-04',
-      open: 36.1325,
-      high: 37.137501,
-      low: 35.950001,
-      close: 37.064999,
-      adjClose: 35.981861,
-      volume: 234428400,
-      unadjustedVolume: 234428400,
-      change: 0.9325,
-      changePercent: 2.581,
-      vwap: 36.7175,
-      label: 'January 04, 19',
-      changeOverTime: 0.02581,
-    },
-    {
-      date: '2019-01-03',
-      open: 35.994999,
-      high: 36.43,
-      low: 35.5,
-      close: 35.547501,
-      adjClose: 34.508717,
-      volume: 365248800,
-      unadjustedVolume: 365248800,
-      change: -0.4475,
-      changePercent: -1.243,
-      vwap: 35.82583,
-      label: 'January 03, 19',
-      changeOverTime: -0.01243,
-    },
-    {
-      date: '2019-01-02',
-      open: 38.7225,
-      high: 39.712502,
-      low: 38.557499,
-      close: 39.48,
-      adjClose: 38.326283,
-      volume: 148158800,
-      unadjustedVolume: 148158800,
-      change: 0.7575,
-      changePercent: 1.956,
-      vwap: 39.25,
-      label: 'January 02, 19',
-      changeOverTime: 0.01956,
-    },
-    {
-      date: '2018-12-31',
-      open: 39.6325,
-      high: 39.84,
-      low: 39.119999,
-      close: 39.435001,
-      adjClose: 38.282612,
-      volume: 140014000,
-      unadjustedVolume: 140014000,
-      change: -0.1975,
-      changePercent: -0.498,
-      vwap: 39.465,
-      label: 'December 31, 18',
-      changeOverTime: -0.00498,
-    },
-    {
-      date: '2018-12-28',
-      open: 39.375,
-      high: 39.630001,
-      low: 38.637501,
-      close: 39.057499,
-      adjClose: 37.916134,
-      volume: 169165600,
-      unadjustedVolume: 169165600,
-      change: -0.3175,
-      changePercent: -0.806,
-      vwap: 39.10833,
-      label: 'December 28, 18',
-      changeOverTime: -0.00806,
-    },
-    {
-      date: '2018-12-27',
-      open: 38.959999,
-      high: 39.192501,
-      low: 37.517502,
-      close: 39.037498,
-      adjClose: 37.896717,
-      volume: 212468400,
-      unadjustedVolume: 212468400,
-      change: 0.0775,
-      changePercent: 0.199,
-      vwap: 38.5825,
-      label: 'December 27, 18',
-      changeOverTime: 0.00199,
-    },
-    {
-      date: '2018-12-26',
-      open: 37.075001,
-      high: 39.307499,
-      low: 36.68,
-      close: 39.2925,
-      adjClose: 38.144268,
-      volume: 234330000,
-      unadjustedVolume: 234330000,
-      change: 2.2175,
-      changePercent: 5.981,
-      vwap: 38.42667,
-      label: 'December 26, 18',
-      changeOverTime: 0.05981,
-    },
-    {
-      date: '2018-12-24',
-      open: 37.037498,
-      high: 37.887501,
-      low: 36.647499,
-      close: 36.7075,
-      adjClose: 35.634811,
-      volume: 148676800,
-      unadjustedVolume: 148676800,
-      change: -0.33,
-      changePercent: -0.891,
-      vwap: 37.08083,
-      label: 'December 24, 18',
-      changeOverTime: -0.00891,
-    },
-    {
-      date: '2018-12-21',
-      open: 39.215,
-      high: 39.540001,
-      low: 37.407501,
-      close: 37.682499,
-      adjClose: 36.581314,
-      volume: 382978400,
-      unadjustedVolume: 382978400,
-      change: -1.5325,
-      changePercent: -3.908,
-      vwap: 38.21,
-      label: 'December 21, 18',
-      changeOverTime: -0.03908,
-    },
-    {
-      date: '2018-12-20',
-      open: 40.099998,
-      high: 40.5275,
-      low: 38.825001,
-      close: 39.2075,
-      adjClose: 38.061756,
-      volume: 259092000,
-      unadjustedVolume: 259092000,
-      change: -0.8925,
-      changePercent: -2.226,
-      vwap: 39.52,
-      label: 'December 20, 18',
-      changeOverTime: -0.02226,
-    },
-    {
-      date: '2018-12-19',
-      open: 41.5,
-      high: 41.862499,
-      low: 39.772499,
-      close: 40.2225,
-      adjClose: 39.0471,
-      volume: 196189200,
-      unadjustedVolume: 196189200,
-      change: -1.2775,
-      changePercent: -3.078,
-      vwap: 40.61917,
-      label: 'December 19, 18',
-      changeOverTime: -0.03078,
-    },
-    {
-      date: '2018-12-18',
-      open: 41.345001,
-      high: 41.8825,
-      low: 41.0975,
-      close: 41.517502,
-      adjClose: 40.304245,
-      volume: 135366000,
-      unadjustedVolume: 135366000,
-      change: 0.1725,
-      changePercent: 0.417,
-      vwap: 41.49917,
-      label: 'December 18, 18',
-      changeOverTime: 0.00417,
-    },
-    {
-      date: '2018-12-17',
-      open: 41.362499,
-      high: 42.087502,
-      low: 40.682499,
-      close: 40.985001,
-      adjClose: 39.787315,
-      volume: 177151600,
-      unadjustedVolume: 177151600,
-      change: -0.3775,
-      changePercent: -0.913,
-      vwap: 41.25167,
-      label: 'December 17, 18',
-      changeOverTime: -0.00913,
-    },
-    {
-      date: '2018-12-14',
-      open: 42.25,
-      high: 42.27,
-      low: 41.32,
-      close: 41.369999,
-      adjClose: 40.161068,
-      volume: 162814800,
-      unadjustedVolume: 162814800,
-      change: -0.88,
-      changePercent: -2.083,
-      vwap: 41.65333,
-      label: 'December 14, 18',
-      changeOverTime: -0.02083,
-    },
-    {
-      date: '2018-12-13',
-      open: 42.622501,
-      high: 43.142502,
-      low: 42.387501,
-      close: 42.737499,
-      adjClose: 41.488594,
-      volume: 127594400,
-      unadjustedVolume: 127594400,
-      change: 0.115,
-      changePercent: 0.27,
-      vwap: 42.75583,
-      label: 'December 13, 18',
-      changeOverTime: 0.0027,
-    },
-    {
-      date: '2018-12-12',
-      open: 42.599998,
-      high: 42.98,
-      low: 42.255001,
-      close: 42.275002,
-      adjClose: 41.039612,
-      volume: 142510800,
-      unadjustedVolume: 142510800,
-      change: -0.325,
-      changePercent: -0.763,
-      vwap: 42.50333,
-      label: 'December 12, 18',
-      changeOverTime: -0.00763,
-    },
-    {
-      date: '2018-12-11',
-      open: 42.915001,
-      high: 42.947498,
-      low: 41.75,
-      close: 42.157501,
-      adjClose: 40.925549,
-      volume: 189126800,
-      unadjustedVolume: 189126800,
-      change: -0.7575,
-      changePercent: -1.765,
-      vwap: 42.285,
-      label: 'December 11, 18',
-      changeOverTime: -0.01765,
-    },
-    {
-      date: '2018-12-10',
-      open: 41.25,
-      high: 42.522499,
-      low: 40.8325,
-      close: 42.400002,
-      adjClose: 41.160965,
-      volume: 248104000,
-      unadjustedVolume: 248104000,
-      change: 1.15,
-      changePercent: 2.788,
-      vwap: 41.91833,
-      label: 'December 10, 18',
-      changeOverTime: 0.02788,
-    },
-    {
-      date: '2018-12-07',
-      open: 43.372501,
-      high: 43.622501,
-      low: 42.075001,
-      close: 42.122501,
-      adjClose: 40.891575,
-      volume: 169126400,
-      unadjustedVolume: 169126400,
-      change: -1.25,
-      changePercent: -2.882,
-      vwap: 42.60667,
-      label: 'December 07, 18',
-      changeOverTime: -0.02882,
-    },
-    {
-      date: '2018-12-06',
-      open: 42.939999,
-      high: 43.695,
-      low: 42.605,
-      close: 43.68,
-      adjClose: 42.403553,
-      volume: 172393600,
-      unadjustedVolume: 172393600,
-      change: 0.74,
-      changePercent: 1.723,
-      vwap: 43.32667,
-      label: 'December 06, 18',
-      changeOverTime: 0.01723,
-    },
-    {
-      date: '2018-12-04',
-      open: 45.237499,
-      high: 45.5975,
-      low: 44.067501,
-      close: 44.172501,
-      adjClose: 42.881664,
-      volume: 165377200,
-      unadjustedVolume: 165377200,
-      change: -1.065,
-      changePercent: -2.354,
-      vwap: 44.6125,
-      label: 'December 04, 18',
-      changeOverTime: -0.02354,
-    },
-    {
-      date: '2018-12-03',
-      open: 46.115002,
-      high: 46.235001,
-      low: 45.302502,
-      close: 46.205002,
-      adjClose: 44.854771,
-      volume: 163210000,
-      unadjustedVolume: 163210000,
-      change: 0.09,
-      changePercent: 0.195,
-      vwap: 45.91417,
-      label: 'December 03, 18',
-      changeOverTime: 0.00195,
-    },
-    {
-      date: '2018-11-30',
-      open: 45.072498,
-      high: 45.0825,
-      low: 44.2575,
-      close: 44.645,
-      adjClose: 43.340359,
-      volume: 158126000,
-      unadjustedVolume: 158126000,
-      change: -0.4275,
-      changePercent: -0.948,
-      vwap: 44.66167,
-      label: 'November 30, 18',
-      changeOverTime: -0.00948,
-    },
-    {
-      date: '2018-11-29',
-      open: 45.665001,
-      high: 45.700001,
-      low: 44.424999,
-      close: 44.887501,
-      adjClose: 43.575775,
-      volume: 167080000,
-      unadjustedVolume: 167080000,
-      change: -0.7775,
-      changePercent: -1.703,
-      vwap: 45.00417,
-      label: 'November 29, 18',
-      changeOverTime: -0.01703,
-    },
-    {
-      date: '2018-11-28',
-      open: 44.182499,
-      high: 45.322498,
-      low: 43.732498,
-      close: 45.235001,
-      adjClose: 43.913113,
-      volume: 184250000,
-      unadjustedVolume: 184250000,
-      change: 1.0525,
-      changePercent: 2.382,
-      vwap: 44.76333,
-      label: 'November 28, 18',
-      changeOverTime: 0.02382,
-    },
-    {
-      date: '2018-11-27',
-      open: 42.877499,
-      high: 43.692501,
-      low: 42.720001,
-      close: 43.560001,
-      adjClose: 42.28706,
-      volume: 165549600,
-      unadjustedVolume: 165549600,
-      change: 0.6825,
-      changePercent: 1.592,
-      vwap: 43.32417,
-      label: 'November 27, 18',
-      changeOverTime: 0.01592,
-    },
-    {
-      date: '2018-11-26',
-      open: 43.560001,
-      high: 43.737499,
-      low: 42.564999,
-      close: 43.654999,
-      adjClose: 42.37928,
-      volume: 179994000,
-      unadjustedVolume: 179994000,
-      change: 0.095,
-      changePercent: 0.218,
-      vwap: 43.31917,
-      label: 'November 26, 18',
-      changeOverTime: 0.00218,
-    },
-    {
-      date: '2018-11-23',
-      open: 43.735001,
-      high: 44.150002,
-      low: 43.025002,
-      close: 43.072498,
-      adjClose: 41.813808,
-      volume: 94496000,
-      unadjustedVolume: 94496000,
-      change: -0.6625,
-      changePercent: -1.515,
-      vwap: 43.41583,
-      label: 'November 23, 18',
-      changeOverTime: -0.01515,
-    },
-    {
-      date: '2018-11-21',
-      open: 44.932499,
-      high: 45.067501,
-      low: 44.137501,
-      close: 44.195,
-      adjClose: 42.903507,
-      volume: 124496800,
-      unadjustedVolume: 124496800,
-      change: -0.7375,
-      changePercent: -1.641,
-      vwap: 44.46667,
-      label: 'November 21, 18',
-      changeOverTime: -0.01641,
-    },
-    {
-      date: '2018-11-20',
-      open: 44.592499,
-      high: 45.3675,
-      low: 43.877499,
-      close: 44.244999,
-      adjClose: 42.952049,
-      volume: 271300800,
-      unadjustedVolume: 271300800,
-      change: -0.3475,
-      changePercent: -0.779,
-      vwap: 44.49667,
-      label: 'November 20, 18',
-      changeOverTime: -0.00779,
-    },
-    {
-      date: '2018-11-19',
-      open: 47.5,
-      high: 47.674999,
-      low: 46.247501,
-      close: 46.465,
-      adjClose: 45.107174,
-      volume: 167701200,
-      unadjustedVolume: 167701200,
-      change: -1.035,
-      changePercent: -2.179,
-      vwap: 46.79583,
-      label: 'November 19, 18',
-      changeOverTime: -0.02179,
-    },
-    {
-      date: '2018-11-16',
-      open: 47.625,
-      high: 48.7425,
-      low: 47.365002,
-      close: 48.3825,
-      adjClose: 46.968639,
-      volume: 147713200,
-      unadjustedVolume: 147713200,
-      change: 0.7575,
-      changePercent: 1.591,
-      vwap: 48.16333,
-      label: 'November 16, 18',
-      changeOverTime: 0.01591,
-    },
-    {
-      date: '2018-11-15',
-      open: 47.0975,
-      high: 47.9925,
-      low: 46.724998,
-      close: 47.852501,
-      adjClose: 46.454128,
-      volume: 185915200,
-      unadjustedVolume: 185915200,
-      change: 0.755,
-      changePercent: 1.603,
-      vwap: 47.52333,
-      label: 'November 15, 18',
-      changeOverTime: 0.01603,
-    },
-    {
-      date: '2018-11-14',
-      open: 48.474998,
-      high: 48.619999,
-      low: 46.482498,
-      close: 46.700001,
-      adjClose: 45.335304,
-      volume: 243204000,
-      unadjustedVolume: 243204000,
-      change: -1.775,
-      changePercent: -3.662,
-      vwap: 47.2675,
-      label: 'November 14, 18',
-      changeOverTime: -0.03662,
-    },
-    {
-      date: '2018-11-13',
-      open: 47.907501,
-      high: 49.294998,
-      low: 47.862499,
-      close: 48.057499,
-      adjClose: 46.653133,
-      volume: 187531600,
-      unadjustedVolume: 187531600,
-      change: 0.15,
-      changePercent: 0.313,
-      vwap: 48.405,
-      label: 'November 13, 18',
-      changeOverTime: 0.00313,
-    },
-    {
-      date: '2018-11-12',
-      open: 49.75,
-      high: 49.962502,
-      low: 48.447498,
-      close: 48.5425,
-      adjClose: 47.123955,
-      volume: 204542000,
-      unadjustedVolume: 204542000,
-      change: -1.2075,
-      changePercent: -2.427,
-      vwap: 48.98417,
-      label: 'November 12, 18',
-      changeOverTime: -0.02427,
-    },
-    {
-      date: '2018-11-09',
-      open: 51.387501,
-      high: 51.502499,
-      low: 50.5625,
-      close: 51.1175,
-      adjClose: 49.623718,
-      volume: 137463200,
-      unadjustedVolume: 137463200,
-      change: -0.27,
-      changePercent: -0.525,
-      vwap: 51.06083,
-      label: 'November 09, 18',
-      changeOverTime: -0.00525,
-    },
-    {
-      date: '2018-11-08',
-      open: 52.494999,
-      high: 52.529999,
-      low: 51.6875,
-      close: 52.122501,
-      adjClose: 50.59935,
-      volume: 101450400,
-      unadjustedVolume: 101450400,
-      change: -0.3725,
-      changePercent: -0.71,
-      vwap: 52.11333,
-      label: 'November 08, 18',
-      changeOverTime: -0.0071,
-    },
-    {
-      date: '2018-11-07',
-      open: 51.4925,
-      high: 52.514999,
-      low: 51.032501,
-      close: 52.487499,
-      adjClose: 50.776508,
-      volume: 133697600,
-      unadjustedVolume: 133697600,
-      change: 0.995,
-      changePercent: 1.932,
-      vwap: 52.01167,
-      label: 'November 07, 18',
-      changeOverTime: 0.01932,
-    },
-    {
-      date: '2018-11-06',
-      open: 50.48,
-      high: 51.18,
-      low: 50.422501,
-      close: 50.942501,
-      adjClose: 49.281883,
-      volume: 127531600,
-      unadjustedVolume: 127531600,
-      change: 0.4625,
-      changePercent: 0.916,
-      vwap: 50.84833,
-      label: 'November 06, 18',
-      changeOverTime: 0.00916,
-    },
-    {
-      date: '2018-11-05',
-      open: 51.075001,
-      high: 51.0975,
-      low: 49.5425,
-      close: 50.397499,
-      adjClose: 48.754639,
-      volume: 264654800,
-      unadjustedVolume: 264654800,
-      change: -0.6775,
-      changePercent: -1.326,
-      vwap: 50.34583,
-      label: 'November 05, 18',
-      changeOverTime: -0.01326,
-    },
-    {
-      date: '2018-11-02',
-      open: 52.387501,
-      high: 53.412498,
-      low: 51.357498,
-      close: 51.869999,
-      adjClose: 50.179142,
-      volume: 365314800,
-      unadjustedVolume: 365314800,
-      change: -0.5175,
-      changePercent: -0.988,
-      vwap: 52.21333,
-      label: 'November 02, 18',
-      changeOverTime: -0.00988,
-    },
-    {
-      date: '2018-11-01',
-      open: 54.762501,
-      high: 55.59,
-      low: 54.202499,
-      close: 55.555,
-      adjClose: 53.744015,
-      volume: 233292800,
-      unadjustedVolume: 233292800,
-      change: 0.7925,
-      changePercent: 1.447,
-      vwap: 55.11583,
-      label: 'November 01, 18',
-      changeOverTime: 0.01447,
-    },
-    {
-      date: '2018-10-31',
-      open: 54.220001,
-      high: 55.112499,
-      low: 54.154999,
-      close: 54.715,
-      adjClose: 52.931404,
-      volume: 153435600,
-      unadjustedVolume: 153435600,
-      change: 0.495,
-      changePercent: 0.913,
-      vwap: 54.66083,
-      label: 'October 31, 18',
-      changeOverTime: 0.00913,
-    },
-    {
-      date: '2018-10-30',
-      open: 52.787498,
-      high: 53.794998,
-      low: 52.317501,
-      close: 53.325001,
-      adjClose: 51.586708,
-      volume: 146640000,
-      unadjustedVolume: 146640000,
-      change: 0.5375,
-      changePercent: 1.018,
-      vwap: 53.14583,
-      label: 'October 30, 18',
-      changeOverTime: 0.01018,
-    },
-    {
-      date: '2018-10-29',
-      open: 54.797501,
-      high: 54.922501,
-      low: 51.522499,
-      close: 53.060001,
-      adjClose: 51.330357,
-      volume: 183742000,
-      unadjustedVolume: 183742000,
-      change: -1.7375,
-      changePercent: -3.171,
-      vwap: 53.16833,
-      label: 'October 29, 18',
-      changeOverTime: -0.03171,
-    },
-    {
-      date: '2018-10-26',
-      open: 53.974998,
-      high: 55.047501,
-      low: 53.1675,
-      close: 54.075001,
-      adjClose: 52.312263,
-      volume: 189033600,
-      unadjustedVolume: 189033600,
-      change: 0.1,
-      changePercent: 0.185,
-      vwap: 54.09667,
-      label: 'October 26, 18',
-      changeOverTime: 0.00185,
-    },
-    {
-      date: '2018-10-25',
-      open: 54.427502,
-      high: 55.345001,
-      low: 54.1875,
-      close: 54.950001,
-      adjClose: 53.158737,
-      volume: 119423200,
-      unadjustedVolume: 119423200,
-      change: 0.5225,
-      changePercent: 0.96,
-      vwap: 54.8275,
-      label: 'October 25, 18',
-      changeOverTime: 0.0096,
-    },
-    {
-      date: '2018-10-24',
-      open: 55.650002,
-      high: 56.057499,
-      low: 53.634998,
-      close: 53.772499,
-      adjClose: 52.019623,
-      volume: 163702000,
-      unadjustedVolume: 163702000,
-      change: -1.8775,
-      changePercent: -3.374,
-      vwap: 54.48833,
-      label: 'October 24, 18',
-      changeOverTime: -0.03374,
-    },
-    {
-      date: '2018-10-23',
-      open: 53.9575,
-      high: 55.8125,
-      low: 53.674999,
-      close: 55.682499,
-      adjClose: 53.867359,
-      volume: 155071200,
-      unadjustedVolume: 155071200,
-      change: 1.725,
-      changePercent: 3.197,
-      vwap: 55.05667,
-      label: 'October 23, 18',
-      changeOverTime: 0.03197,
-    },
-    {
-      date: '2018-10-22',
-      open: 54.947498,
-      high: 55.84,
-      low: 54.735001,
-      close: 55.162498,
-      adjClose: 53.364311,
-      volume: 115168400,
-      unadjustedVolume: 115168400,
-      change: 0.215,
-      changePercent: 0.391,
-      vwap: 55.24583,
-      label: 'October 22, 18',
-      changeOverTime: 0.00391,
-    },
-    {
-      date: '2018-10-19',
-      open: 54.514999,
-      high: 55.314999,
-      low: 54.357498,
-      close: 54.827499,
-      adjClose: 53.040226,
-      volume: 132314800,
-      unadjustedVolume: 132314800,
-      change: 0.3125,
-      changePercent: 0.573,
-      vwap: 54.83333,
-      label: 'October 19, 18',
-      changeOverTime: 0.00573,
-    },
-    {
-      date: '2018-10-18',
-      open: 54.465,
-      high: 54.935001,
-      low: 53.25,
-      close: 54.005001,
-      adjClose: 52.244549,
-      volume: 130325200,
-      unadjustedVolume: 130325200,
-      change: -0.46,
-      changePercent: -0.845,
-      vwap: 54.06333,
-      label: 'October 18, 18',
-      changeOverTime: -0.00845,
-    },
-    {
-      date: '2018-10-17',
-      open: 55.575001,
-      high: 55.66,
-      low: 54.834999,
-      close: 55.297501,
-      adjClose: 53.494915,
-      volume: 91541600,
-      unadjustedVolume: 91541600,
-      change: -0.2775,
-      changePercent: -0.499,
-      vwap: 55.26417,
-      label: 'October 17, 18',
-      changeOverTime: -0.00499,
-    },
-    {
-      date: '2018-10-16',
-      open: 54.732498,
-      high: 55.747501,
-      low: 54.189999,
-      close: 55.537498,
-      adjClose: 53.727085,
-      volume: 116736000,
-      unadjustedVolume: 116736000,
-      change: 0.805,
-      changePercent: 1.471,
-      vwap: 55.15833,
-      label: 'October 16, 18',
-      changeOverTime: 0.01471,
-    },
-    {
-      date: '2018-10-15',
-      open: 55.290001,
-      high: 55.4575,
-      low: 54.317501,
-      close: 54.34,
-      adjClose: 52.568623,
-      volume: 123164000,
-      unadjustedVolume: 123164000,
-      change: -0.95,
-      changePercent: -1.718,
-      vwap: 54.705,
-      label: 'October 15, 18',
-      changeOverTime: -0.01718,
-    },
-    {
-      date: '2018-10-12',
-      open: 55.105,
-      high: 55.720001,
-      low: 54.209999,
-      close: 55.5275,
-      adjClose: 53.717411,
-      volume: 161351600,
-      unadjustedVolume: 161351600,
-      change: 0.4225,
-      changePercent: 0.767,
-      vwap: 55.1525,
-      label: 'October 12, 18',
-      changeOverTime: 0.00767,
-    },
-    {
-      date: '2018-10-11',
-      open: 53.630001,
-      high: 54.875,
-      low: 53.080002,
-      close: 53.612499,
-      adjClose: 51.864838,
-      volume: 212497600,
-      unadjustedVolume: 212497600,
-      change: -0.0175,
-      changePercent: -0.033,
-      vwap: 53.85583,
-      label: 'October 11, 18',
-      changeOverTime: -0.00033,
-    },
-    {
-      date: '2018-10-10',
-      open: 56.365002,
-      high: 56.587502,
-      low: 54.012501,
-      close: 54.09,
-      adjClose: 52.326775,
-      volume: 167962400,
-      unadjustedVolume: 167962400,
-      change: -2.275,
-      changePercent: -4.036,
-      vwap: 54.89667,
-      label: 'October 10, 18',
-      changeOverTime: -0.04036,
-    },
-    {
-      date: '2018-10-09',
-      open: 55.91,
-      high: 56.817501,
-      low: 55.5625,
-      close: 56.717499,
-      adjClose: 54.868626,
-      volume: 107564000,
-      unadjustedVolume: 107564000,
-      change: 0.8075,
-      changePercent: 1.444,
-      vwap: 56.36583,
-      label: 'October 09, 18',
-      changeOverTime: 0.01444,
-    },
-    {
-      date: '2018-10-08',
-      open: 55.552502,
-      high: 56.200001,
-      low: 55.049999,
-      close: 55.942501,
-      adjClose: 54.118893,
-      volume: 118655600,
-      unadjustedVolume: 118655600,
-      change: 0.39,
-      changePercent: 0.702,
-      vwap: 55.73083,
-      label: 'October 08, 18',
-      changeOverTime: 0.00702,
-    },
-    {
-      date: '2018-10-05',
-      open: 56.990002,
-      high: 57.102501,
-      low: 55.145,
-      close: 56.072498,
-      adjClose: 54.24464,
-      volume: 134322000,
-      unadjustedVolume: 134322000,
-      change: -0.9175,
-      changePercent: -1.61,
-      vwap: 56.10667,
-      label: 'October 05, 18',
-      changeOverTime: -0.0161,
-    },
-    {
-      date: '2018-10-04',
-      open: 57.695,
-      high: 58.087502,
-      low: 56.682499,
-      close: 56.997501,
-      adjClose: 55.139496,
-      volume: 128168000,
-      unadjustedVolume: 128168000,
-      change: -0.6975,
-      changePercent: -1.209,
-      vwap: 57.25583,
-      label: 'October 04, 18',
-      changeOverTime: -0.01209,
-    },
-    {
-      date: '2018-10-03',
-      open: 57.512501,
-      high: 58.3675,
-      low: 57.445,
-      close: 58.017502,
-      adjClose: 56.126244,
-      volume: 114619200,
-      unadjustedVolume: 114619200,
-      change: 0.505,
-      changePercent: 0.878,
-      vwap: 57.94333,
-      label: 'October 03, 18',
-      changeOverTime: 0.00878,
-    },
-    {
-      date: '2018-10-02',
-      open: 56.8125,
-      high: 57.5,
-      low: 56.657501,
-      close: 57.32,
-      adjClose: 55.451481,
-      volume: 99152800,
-      unadjustedVolume: 99152800,
-      change: 0.5075,
-      changePercent: 0.893,
-      vwap: 57.15917,
-      label: 'October 02, 18',
-      changeOverTime: 0.00893,
-    },
-    {
-      date: '2018-10-01',
-      open: 56.987499,
-      high: 57.355,
-      low: 56.587502,
-      close: 56.814999,
-      adjClose: 54.962944,
-      volume: 94403200,
-      unadjustedVolume: 94403200,
-      change: -0.1725,
-      changePercent: -0.303,
-      vwap: 56.91917,
-      label: 'October 01, 18',
-      changeOverTime: -0.00303,
-    },
-    {
-      date: '2018-09-28',
-      open: 56.197498,
-      high: 56.459999,
-      low: 56.005001,
-      close: 56.435001,
-      adjClose: 54.595329,
-      volume: 91717600,
-      unadjustedVolume: 91717600,
-      change: 0.2375,
-      changePercent: 0.423,
-      vwap: 56.3,
-      label: 'September 28, 18',
-      changeOverTime: 0.00423,
-    },
-    {
-      date: '2018-09-27',
-      open: 55.955002,
-      high: 56.610001,
-      low: 55.884998,
-      close: 56.237499,
-      adjClose: 54.404266,
-      volume: 120724800,
-      unadjustedVolume: 120724800,
-      change: 0.2825,
-      changePercent: 0.505,
-      vwap: 56.24417,
-      label: 'September 27, 18',
-      changeOverTime: 0.00505,
-    },
-    {
-      date: '2018-09-26',
-      open: 55.25,
-      high: 55.9375,
-      low: 54.939999,
-      close: 55.105,
-      adjClose: 53.308685,
-      volume: 95938800,
-      unadjustedVolume: 95938800,
-      change: -0.145,
-      changePercent: -0.262,
-      vwap: 55.3275,
-      label: 'September 26, 18',
-      changeOverTime: -0.00262,
-    },
-    {
-      date: '2018-09-25',
-      open: 54.9375,
-      high: 55.705002,
-      low: 54.924999,
-      close: 55.547501,
-      adjClose: 53.736759,
-      volume: 98217600,
-      unadjustedVolume: 98217600,
-      change: 0.61,
-      changePercent: 1.11,
-      vwap: 55.3925,
-      label: 'September 25, 18',
-      changeOverTime: 0.0111,
-    },
-    {
-      date: '2018-09-24',
-      open: 54.205002,
-      high: 55.314999,
-      low: 54.157501,
-      close: 55.197498,
-      adjClose: 53.398167,
-      volume: 110773600,
-      unadjustedVolume: 110773600,
-      change: 0.9925,
-      changePercent: 1.831,
-      vwap: 54.89,
-      label: 'September 24, 18',
-      changeOverTime: 0.01831,
-    },
-    {
-      date: '2018-09-21',
-      open: 55.195,
-      high: 55.34,
-      low: 54.322498,
-      close: 54.415001,
-      adjClose: 52.641186,
-      volume: 384986800,
-      unadjustedVolume: 384986800,
-      change: -0.78,
-      changePercent: -1.413,
-      vwap: 54.6925,
-      label: 'September 21, 18',
-      changeOverTime: -0.01413,
-    },
-    {
-      date: '2018-09-20',
-      open: 55.060001,
-      high: 55.57,
-      low: 54.787498,
-      close: 55.0075,
-      adjClose: 53.214363,
-      volume: 106435200,
-      unadjustedVolume: 106435200,
-      change: -0.0525,
-      changePercent: -0.095,
-      vwap: 55.12167,
-      label: 'September 20, 18',
-      changeOverTime: -0.00095,
-    },
-    {
-      date: '2018-09-19',
-      open: 54.625,
-      high: 54.904999,
-      low: 53.825001,
-      close: 54.592499,
-      adjClose: 52.812893,
-      volume: 108495200,
-      unadjustedVolume: 108495200,
-      change: -0.0325,
-      changePercent: -0.059,
-      vwap: 54.44083,
-      label: 'September 19, 18',
-      changeOverTime: -0.00059,
-    },
-    {
-      date: '2018-09-18',
-      open: 54.447498,
-      high: 55.462502,
-      low: 54.279999,
-      close: 54.560001,
-      adjClose: 52.781452,
-      volume: 126286800,
-      unadjustedVolume: 126286800,
-      change: 0.1125,
-      changePercent: 0.207,
-      vwap: 54.7675,
-      label: 'September 18, 18',
-      changeOverTime: 0.00207,
-    },
-    {
-      date: '2018-09-17',
-      open: 55.537498,
-      high: 55.737499,
-      low: 54.317501,
-      close: 54.470001,
-      adjClose: 52.694386,
-      volume: 148780400,
-      unadjustedVolume: 148780400,
-      change: -1.0675,
-      changePercent: -1.922,
-      vwap: 54.84167,
-      label: 'September 17, 18',
-      changeOverTime: -0.01922,
-    },
-    {
-      date: '2018-09-14',
-      open: 56.4375,
-      high: 56.709999,
-      low: 55.630001,
-      close: 55.959999,
-      adjClose: 54.135818,
-      volume: 127997200,
-      unadjustedVolume: 127997200,
-      change: -0.4775,
-      changePercent: -0.846,
-      vwap: 56.1,
-      label: 'September 14, 18',
-      changeOverTime: -0.00846,
-    },
-    {
-      date: '2018-09-13',
-      open: 55.880001,
-      high: 57.087502,
-      low: 55.642502,
-      close: 56.602501,
-      adjClose: 54.757366,
-      volume: 166825600,
-      unadjustedVolume: 166825600,
-      change: 0.7225,
-      changePercent: 1.293,
-      vwap: 56.44417,
-      label: 'September 13, 18',
-      changeOverTime: 0.01293,
-    },
-    {
-      date: '2018-09-12',
-      open: 56.235001,
-      high: 56.25,
-      low: 54.959999,
-      close: 55.267502,
-      adjClose: 53.465893,
-      volume: 197114800,
-      unadjustedVolume: 197114800,
-      change: -0.9675,
-      changePercent: -1.72,
-      vwap: 55.4925,
-      label: 'September 12, 18',
-      changeOverTime: -0.0172,
-    },
-    {
-      date: '2018-09-11',
-      open: 54.502499,
-      high: 56.075001,
-      low: 54.139999,
-      close: 55.962502,
-      adjClose: 54.138237,
-      volume: 142996000,
-      unadjustedVolume: 142996000,
-      change: 1.46,
-      changePercent: 2.679,
-      vwap: 55.3925,
-      label: 'September 11, 18',
-      changeOverTime: 0.02679,
-    },
-    {
-      date: '2018-09-10',
-      open: 55.237499,
-      high: 55.462502,
-      low: 54.1175,
-      close: 54.5825,
-      adjClose: 52.803215,
-      volume: 158066000,
-      unadjustedVolume: 158066000,
-      change: -0.655,
-      changePercent: -1.186,
-      vwap: 54.72083,
-      label: 'September 10, 18',
-      changeOverTime: -0.01186,
-    },
-    {
-      date: '2018-09-07',
-      open: 55.462502,
-      high: 56.342499,
-      low: 55.177502,
-      close: 55.325001,
-      adjClose: 53.521519,
-      volume: 150479200,
-      unadjustedVolume: 150479200,
-      change: -0.1375,
-      changePercent: -0.248,
-      vwap: 55.615,
-      label: 'September 07, 18',
-      changeOverTime: -0.00248,
-    },
-    {
-      date: '2018-09-06',
-      open: 56.557499,
-      high: 56.837502,
-      low: 55.325001,
-      close: 55.775002,
-      adjClose: 53.956841,
-      volume: 137160000,
-      unadjustedVolume: 137160000,
-      change: -0.7825,
-      changePercent: -1.384,
-      vwap: 55.97917,
-      label: 'September 06, 18',
-      changeOverTime: -0.01384,
-    },
-    {
-      date: '2018-09-05',
-      open: 57.247501,
-      high: 57.4175,
-      low: 56.275002,
-      close: 56.717499,
-      adjClose: 54.868626,
-      volume: 133332000,
-      unadjustedVolume: 133332000,
-      change: -0.53,
-      changePercent: -0.926,
-      vwap: 56.80333,
-      label: 'September 05, 18',
-      changeOverTime: -0.00926,
-    },
-    {
-      date: '2018-09-04',
-      open: 57.102501,
-      high: 57.294998,
-      low: 56.657501,
-      close: 57.09,
-      adjClose: 55.228977,
-      volume: 109560400,
-      unadjustedVolume: 109560400,
-      change: -0.0125,
-      changePercent: -0.022,
-      vwap: 57.01417,
-      label: 'September 04, 18',
-      changeOverTime: -0.00022,
-    },
-    {
-      date: '2018-08-31',
-      open: 56.627499,
-      high: 57.217499,
-      low: 56.5,
-      close: 56.907501,
-      adjClose: 55.052429,
-      volume: 173360400,
-      unadjustedVolume: 173360400,
-      change: 0.28,
-      changePercent: 0.494,
-      vwap: 56.875,
-      label: 'August 31, 18',
-      changeOverTime: 0.00494,
-    },
-    {
-      date: '2018-08-30',
-      open: 55.8125,
-      high: 57.064999,
-      low: 55.599998,
-      close: 56.2575,
-      adjClose: 54.423618,
-      volume: 195175200,
-      unadjustedVolume: 195175200,
-      change: 0.445,
-      changePercent: 0.797,
-      vwap: 56.3075,
-      label: 'August 30, 18',
-      changeOverTime: 0.00797,
-    },
-    {
-      date: '2018-08-29',
-      open: 55.037498,
-      high: 55.872501,
-      low: 54.852501,
-      close: 55.744999,
-      adjClose: 53.927822,
-      volume: 109019200,
-      unadjustedVolume: 109019200,
-      change: 0.7075,
-      changePercent: 1.285,
-      vwap: 55.49,
-      label: 'August 29, 18',
-      changeOverTime: 0.01285,
-    },
-    {
-      date: '2018-08-28',
-      open: 54.752499,
-      high: 55.134998,
-      low: 54.73,
-      close: 54.924999,
-      adjClose: 53.134556,
-      volume: 91107200,
-      unadjustedVolume: 91107200,
-      change: 0.1725,
-      changePercent: 0.315,
-      vwap: 54.93,
-      label: 'August 28, 18',
-      changeOverTime: 0.00315,
-    },
-    {
-      date: '2018-08-27',
-      open: 54.287498,
-      high: 54.685001,
-      low: 54.0825,
-      close: 54.485001,
-      adjClose: 52.708897,
-      volume: 82100400,
-      unadjustedVolume: 82100400,
-      change: 0.1975,
-      changePercent: 0.364,
-      vwap: 54.4175,
-      label: 'August 27, 18',
-      changeOverTime: 0.00364,
-    },
-    {
-      date: '2018-08-24',
-      open: 54.150002,
-      high: 54.224998,
-      low: 53.7775,
-      close: 54.040001,
-      adjClose: 52.278408,
-      volume: 73905600,
-      unadjustedVolume: 73905600,
-      change: -0.11,
-      changePercent: -0.203,
-      vwap: 54.01417,
-      label: 'August 24, 18',
-      changeOverTime: -0.00203,
-    },
-    {
-      date: '2018-08-23',
-      open: 53.662498,
-      high: 54.262501,
-      low: 53.650002,
-      close: 53.872501,
-      adjClose: 52.116367,
-      volume: 75532800,
-      unadjustedVolume: 75532800,
-      change: 0.21,
-      changePercent: 0.391,
-      vwap: 53.92833,
-      label: 'August 23, 18',
-      changeOverTime: 0.00391,
-    },
-    {
-      date: '2018-08-22',
-      open: 53.525002,
-      high: 54.09,
-      low: 53.459999,
-      close: 53.762501,
-      adjClose: 52.009953,
-      volume: 76072400,
-      unadjustedVolume: 76072400,
-      change: 0.2375,
-      changePercent: 0.444,
-      vwap: 53.77083,
-      label: 'August 22, 18',
-      changeOverTime: 0.00444,
-    },
-    {
-      date: '2018-08-21',
-      open: 54.200001,
-      high: 54.297501,
-      low: 53.5075,
-      close: 53.759998,
-      adjClose: 52.00753,
-      volume: 104639200,
-      unadjustedVolume: 104639200,
-      change: -0.44,
-      changePercent: -0.812,
-      vwap: 53.855,
-      label: 'August 21, 18',
-      changeOverTime: -0.00812,
-    },
-    {
-      date: '2018-08-20',
-      open: 54.525002,
-      high: 54.794998,
-      low: 53.7775,
-      close: 53.865002,
-      adjClose: 52.109116,
-      volume: 121150800,
-      unadjustedVolume: 121150800,
-      change: -0.66,
-      changePercent: -1.21,
-      vwap: 54.14583,
-      label: 'August 20, 18',
-      changeOverTime: -0.0121,
-    },
-    {
-      date: '2018-08-17',
-      open: 53.360001,
-      high: 54.487499,
-      low: 53.290001,
-      close: 54.395,
-      adjClose: 52.621826,
-      volume: 141708000,
-      unadjustedVolume: 141708000,
-      change: 1.035,
-      changePercent: 1.94,
-      vwap: 54.0575,
-      label: 'August 17, 18',
-      changeOverTime: 0.0194,
-    },
-    {
-      date: '2018-08-16',
-      open: 52.9375,
-      high: 53.452499,
-      low: 52.8675,
-      close: 53.330002,
-      adjClose: 51.591545,
-      volume: 114001600,
-      unadjustedVolume: 114001600,
-      change: 0.3925,
-      changePercent: 0.741,
-      vwap: 53.21667,
-      label: 'August 16, 18',
-      changeOverTime: 0.00741,
-    },
-    {
-      date: '2018-08-15',
-      open: 52.305,
-      high: 52.685001,
-      low: 52.0825,
-      close: 52.560001,
-      adjClose: 50.846649,
-      volume: 115230400,
-      unadjustedVolume: 115230400,
-      change: 0.255,
-      changePercent: 0.488,
-      vwap: 52.4425,
-      label: 'August 15, 18',
-      changeOverTime: 0.00488,
-    },
-    {
-      date: '2018-08-14',
-      open: 52.540001,
-      high: 52.639999,
-      low: 52.064999,
-      close: 52.4375,
-      adjClose: 50.728138,
-      volume: 82992000,
-      unadjustedVolume: 82992000,
-      change: -0.1025,
-      changePercent: -0.195,
-      vwap: 52.38083,
-      label: 'August 14, 18',
-      changeOverTime: -0.00195,
-    },
-    {
-      date: '2018-08-13',
-      open: 52.327499,
-      high: 52.737499,
-      low: 51.924999,
-      close: 52.217499,
-      adjClose: 50.515312,
-      volume: 103563600,
-      unadjustedVolume: 103563600,
-      change: -0.11,
-      changePercent: -0.21,
-      vwap: 52.29333,
-      label: 'August 13, 18',
-      changeOverTime: -0.0021,
-    },
-    {
-      date: '2018-08-10',
-      open: 51.84,
-      high: 52.275002,
-      low: 51.6675,
-      close: 51.8825,
-      adjClose: 50.191231,
-      volume: 98444800,
-      unadjustedVolume: 98444800,
-      change: 0.0425,
-      changePercent: 0.082,
-      vwap: 51.94167,
-      label: 'August 10, 18',
-      changeOverTime: 0.00082,
-    },
-    {
-      date: '2018-08-09',
-      open: 52.3825,
-      high: 52.445,
-      low: 51.799999,
-      close: 52.220001,
-      adjClose: 50.341187,
-      volume: 93970400,
-      unadjustedVolume: 93970400,
-      change: -0.1625,
-      changePercent: -0.31,
-      vwap: 52.155,
-      label: 'August 09, 18',
-      changeOverTime: -0.0031,
-    },
-    {
-      date: '2018-08-08',
-      open: 51.512501,
-      high: 51.952499,
-      low: 51.130001,
-      close: 51.8125,
-      adjClose: 49.948341,
-      volume: 90102000,
-      unadjustedVolume: 90102000,
-      change: 0.3,
-      changePercent: 0.582,
-      vwap: 51.63167,
-      label: 'August 08, 18',
-      changeOverTime: 0.00582,
-    },
-    {
-      date: '2018-08-07',
-      open: 52.330002,
-      high: 52.375,
-      low: 51.689999,
-      close: 51.7775,
-      adjClose: 49.9146,
-      volume: 102349600,
-      unadjustedVolume: 102349600,
-      change: -0.5525,
-      changePercent: -1.056,
-      vwap: 51.9475,
-      label: 'August 07, 18',
-      changeOverTime: -0.01056,
-    },
-    {
-      date: '2018-08-06',
-      open: 52.0,
-      high: 52.3125,
-      low: 51.767502,
-      close: 52.267502,
-      adjClose: 50.386971,
-      volume: 101701600,
-      unadjustedVolume: 101701600,
-      change: 0.2675,
-      changePercent: 0.514,
-      vwap: 52.11583,
-      label: 'August 06, 18',
-      changeOverTime: 0.00514,
-    },
-    {
-      date: '2018-08-03',
-      open: 51.7575,
-      high: 52.185001,
-      low: 51.369999,
-      close: 51.997501,
-      adjClose: 50.126686,
-      volume: 133789600,
-      unadjustedVolume: 133789600,
-      change: 0.24,
-      changePercent: 0.464,
-      vwap: 51.85083,
-      label: 'August 03, 18',
-      changeOverTime: 0.00464,
-    },
-    {
-      date: '2018-08-02',
-      open: 50.145,
-      high: 52.095001,
-      low: 50.087502,
-      close: 51.8475,
-      adjClose: 49.982079,
-      volume: 249616000,
-      unadjustedVolume: 249616000,
-      change: 1.7025,
-      changePercent: 3.395,
-      vwap: 51.34333,
-      label: 'August 02, 18',
-      changeOverTime: 0.03395,
-    },
-    {
-      date: '2018-08-01',
-      open: 49.782501,
-      high: 50.439999,
-      low: 49.327499,
-      close: 50.375,
-      adjClose: 48.562565,
-      volume: 271742800,
-      unadjustedVolume: 271742800,
-      change: 0.5925,
-      changePercent: 1.19,
-      vwap: 50.0475,
-      label: 'August 01, 18',
-      changeOverTime: 0.0119,
-    },
-    {
-      date: '2018-07-31',
-      open: 47.575001,
-      high: 48.035,
-      low: 47.334999,
-      close: 47.572498,
-      adjClose: 45.860886,
-      volume: 157492000,
-      unadjustedVolume: 157492000,
-      change: -0.0025,
-      changePercent: -0.005,
-      vwap: 47.6475,
-      label: 'July 31, 18',
-      changeOverTime: -0.00005,
-    },
-    {
-      date: '2018-07-30',
-      open: 47.974998,
-      high: 48.049999,
-      low: 47.267502,
-      close: 47.477501,
-      adjClose: 45.769306,
-      volume: 84118000,
-      unadjustedVolume: 84118000,
-      change: -0.4975,
-      changePercent: -1.037,
-      vwap: 47.59833,
-      label: 'July 30, 18',
-      changeOverTime: -0.01037,
-    },
-    {
-      date: '2018-07-27',
-      open: 48.747501,
-      high: 48.797501,
-      low: 47.525002,
-      close: 47.744999,
-      adjClose: 46.027184,
-      volume: 96096000,
-      unadjustedVolume: 96096000,
-      change: -1.0025,
-      changePercent: -2.057,
-      vwap: 48.0225,
-      label: 'July 27, 18',
-      changeOverTime: -0.02057,
-    },
-    {
-      date: '2018-07-26',
-      open: 48.6525,
-      high: 48.990002,
-      low: 48.4025,
-      close: 48.552502,
-      adjClose: 46.805634,
-      volume: 76304000,
-      unadjustedVolume: 76304000,
-      change: -0.1,
-      changePercent: -0.206,
-      vwap: 48.64833,
-      label: 'July 26, 18',
-      changeOverTime: -0.00206,
-    },
-    {
-      date: '2018-07-25',
-      open: 48.264999,
-      high: 48.712502,
-      low: 48.107498,
-      close: 48.705002,
-      adjClose: 46.952641,
-      volume: 66839600,
-      unadjustedVolume: 66839600,
-      change: 0.44,
-      changePercent: 0.912,
-      vwap: 48.50833,
-      label: 'July 25, 18',
-      changeOverTime: 0.00912,
-    },
-    {
-      date: '2018-07-24',
-      open: 48.112499,
-      high: 48.415001,
-      low: 48.012501,
-      close: 48.25,
-      adjClose: 46.514015,
-      volume: 74791600,
-      unadjustedVolume: 74791600,
-      change: 0.1375,
-      changePercent: 0.286,
-      vwap: 48.22583,
-      label: 'July 24, 18',
-      changeOverTime: 0.00286,
-    },
-    {
-      date: '2018-07-23',
-      open: 47.669998,
-      high: 47.990002,
-      low: 47.389999,
-      close: 47.9025,
-      adjClose: 46.179016,
-      volume: 63957600,
-      unadjustedVolume: 63957600,
-      change: 0.2325,
-      changePercent: 0.488,
-      vwap: 47.76083,
-      label: 'July 23, 18',
-      changeOverTime: 0.00488,
-    },
-    {
-      date: '2018-07-20',
-      open: 47.945,
-      high: 48.107498,
-      low: 47.5425,
-      close: 47.860001,
-      adjClose: 46.138042,
-      volume: 82704800,
-      unadjustedVolume: 82704800,
-      change: -0.085,
-      changePercent: -0.177,
-      vwap: 47.83667,
-      label: 'July 20, 18',
-      changeOverTime: -0.00177,
-    },
-    {
-      date: '2018-07-19',
-      open: 47.422501,
-      high: 48.137501,
-      low: 47.422501,
-      close: 47.970001,
-      adjClose: 46.244091,
-      volume: 81147200,
-      unadjustedVolume: 81147200,
-      change: 0.5475,
-      changePercent: 1.155,
-      vwap: 47.84333,
-      label: 'July 19, 18',
-      changeOverTime: 0.01155,
-    },
-    {
-      date: '2018-07-18',
-      open: 47.945,
-      high: 47.950001,
-      low: 47.482498,
-      close: 47.599998,
-      adjClose: 45.887398,
-      volume: 65573600,
-      unadjustedVolume: 65573600,
-      change: -0.345,
-      changePercent: -0.72,
-      vwap: 47.6775,
-      label: 'July 18, 18',
-      changeOverTime: -0.0072,
-    },
-    {
-      date: '2018-07-17',
-      open: 47.4375,
-      high: 47.967499,
-      low: 47.299999,
-      close: 47.862499,
-      adjClose: 46.140453,
-      volume: 62138000,
-      unadjustedVolume: 62138000,
-      change: 0.425,
-      changePercent: 0.896,
-      vwap: 47.71,
-      label: 'July 17, 18',
-      changeOverTime: 0.00896,
-    },
-    {
-      date: '2018-07-16',
-      open: 47.880001,
-      high: 48.162498,
-      low: 47.605,
-      close: 47.727501,
-      adjClose: 46.010319,
-      volume: 60172400,
-      unadjustedVolume: 60172400,
-      change: -0.1525,
-      changePercent: -0.319,
-      vwap: 47.83167,
-      label: 'July 16, 18',
-      changeOverTime: -0.00319,
-    },
-    {
-      date: '2018-07-13',
-      open: 47.77,
-      high: 47.959999,
-      low: 47.724998,
-      close: 47.8325,
-      adjClose: 46.111538,
-      volume: 50055600,
-      unadjustedVolume: 50055600,
-      change: 0.0625,
-      changePercent: 0.131,
-      vwap: 47.83917,
-      label: 'July 13, 18',
-      changeOverTime: 0.00131,
-    },
-    {
-      date: '2018-07-12',
-      open: 47.3825,
-      high: 47.852501,
-      low: 47.327499,
-      close: 47.7575,
-      adjClose: 46.03923,
-      volume: 72164400,
-      unadjustedVolume: 72164400,
-      change: 0.375,
-      changePercent: 0.791,
-      vwap: 47.64583,
-      label: 'July 12, 18',
-      changeOverTime: 0.00791,
-    },
-    {
-      date: '2018-07-11',
-      open: 47.125,
-      high: 47.445,
-      low: 46.9025,
-      close: 46.970001,
-      adjClose: 45.280067,
-      volume: 75326000,
-      unadjustedVolume: 75326000,
-      change: -0.155,
-      changePercent: -0.329,
-      vwap: 47.10583,
-      label: 'July 11, 18',
-      changeOverTime: -0.00329,
-    },
-    {
-      date: '2018-07-10',
-      open: 47.677502,
-      high: 47.82,
-      low: 47.544998,
-      close: 47.587502,
-      adjClose: 45.875359,
-      volume: 63756400,
-      unadjustedVolume: 63756400,
-      change: -0.09,
-      changePercent: -0.189,
-      vwap: 47.65083,
-      label: 'July 10, 18',
-      changeOverTime: -0.00189,
-    },
-    {
-      date: '2018-07-09',
-      open: 47.375,
-      high: 47.669998,
-      low: 47.325001,
-      close: 47.645,
-      adjClose: 45.930782,
-      volume: 79026400,
-      unadjustedVolume: 79026400,
-      change: 0.27,
-      changePercent: 0.57,
-      vwap: 47.54667,
-      label: 'July 09, 18',
-      changeOverTime: 0.0057,
-    },
-    {
-      date: '2018-07-06',
-      open: 46.355,
-      high: 47.107498,
-      low: 46.299999,
-      close: 46.9925,
-      adjClose: 45.301758,
-      volume: 69940800,
-      unadjustedVolume: 69940800,
-      change: 0.6375,
-      changePercent: 1.375,
-      vwap: 46.8,
-      label: 'July 06, 18',
-      changeOverTime: 0.01375,
-    },
-    {
-      date: '2018-07-05',
-      open: 46.314999,
-      high: 46.602501,
-      low: 46.07,
-      close: 46.349998,
-      adjClose: 44.682373,
-      volume: 66416800,
-      unadjustedVolume: 66416800,
-      change: 0.035,
-      changePercent: 0.076,
-      vwap: 46.34083,
-      label: 'July 05, 18',
-      changeOverTime: 0.00076,
-    },
-    {
-      date: '2018-07-03',
-      open: 46.947498,
-      high: 46.987499,
-      low: 45.884998,
-      close: 45.98,
-      adjClose: 44.325691,
-      volume: 55819200,
-      unadjustedVolume: 55819200,
-      change: -0.9675,
-      changePercent: -2.061,
-      vwap: 46.28417,
-      label: 'July 03, 18',
-      changeOverTime: -0.02061,
-    },
-    {
-      date: '2018-07-02',
-      open: 45.955002,
-      high: 46.825001,
-      low: 45.855,
-      close: 46.794998,
-      adjClose: 45.111362,
-      volume: 70925200,
-      unadjustedVolume: 70925200,
-      change: 0.84,
-      changePercent: 1.828,
-      vwap: 46.49167,
-      label: 'July 02, 18',
-      changeOverTime: 0.01828,
-    },
-    {
-      date: '2018-06-29',
-      open: 46.572498,
-      high: 46.797501,
-      low: 45.727501,
-      close: 46.2775,
-      adjClose: 44.61248,
-      volume: 90950800,
-      unadjustedVolume: 90950800,
-      change: -0.295,
-      changePercent: -0.633,
-      vwap: 46.2675,
-      label: 'June 29, 18',
-      changeOverTime: -0.00633,
-    },
-    {
-      date: '2018-06-28',
-      open: 46.025002,
-      high: 46.552502,
-      low: 45.950001,
-      close: 46.375,
-      adjClose: 44.70647,
-      volume: 69460800,
-      unadjustedVolume: 69460800,
-      change: 0.35,
-      changePercent: 0.76,
-      vwap: 46.2925,
-      label: 'June 28, 18',
-      changeOverTime: 0.0076,
-    },
-    {
-      date: '2018-06-27',
-      open: 46.307499,
-      high: 46.82,
-      low: 46.0075,
-      close: 46.040001,
-      adjClose: 44.38353,
-      volume: 101141200,
-      unadjustedVolume: 101141200,
-      change: -0.2675,
-      changePercent: -0.578,
-      vwap: 46.28917,
-      label: 'June 27, 18',
-      changeOverTime: -0.00578,
-    },
-    {
-      date: '2018-06-26',
-      open: 45.747501,
-      high: 46.6325,
-      low: 45.634998,
-      close: 46.107498,
-      adjClose: 44.448593,
-      volume: 98276800,
-      unadjustedVolume: 98276800,
-      change: 0.36,
-      changePercent: 0.787,
-      vwap: 46.125,
-      label: 'June 26, 18',
-      changeOverTime: 0.00787,
-    },
-    {
-      date: '2018-06-25',
-      open: 45.849998,
-      high: 46.23,
-      low: 45.182499,
-      close: 45.5425,
-      adjClose: 43.903927,
-      volume: 126652400,
-      unadjustedVolume: 126652400,
-      change: -0.3075,
-      changePercent: -0.671,
-      vwap: 45.65167,
-      label: 'June 25, 18',
-      changeOverTime: -0.00671,
-    },
-    {
-      date: '2018-06-22',
-      open: 46.529999,
-      high: 46.537498,
-      low: 46.174999,
-      close: 46.23,
-      adjClose: 44.566692,
-      volume: 108801600,
-      unadjustedVolume: 108801600,
-      change: -0.3,
-      changePercent: -0.645,
-      vwap: 46.31417,
-      label: 'June 22, 18',
-      changeOverTime: -0.00645,
-    },
-    {
-      date: '2018-06-21',
-      open: 46.8125,
-      high: 47.087502,
-      low: 46.235001,
-      close: 46.365002,
-      adjClose: 44.696838,
-      volume: 102847600,
-      unadjustedVolume: 102847600,
-      change: -0.4475,
-      changePercent: -0.956,
-      vwap: 46.5625,
-      label: 'June 21, 18',
-      changeOverTime: -0.00956,
-    },
-    {
-      date: '2018-06-20',
-      open: 46.587502,
-      high: 46.799999,
-      low: 46.432499,
-      close: 46.625,
-      adjClose: 44.947487,
-      volume: 82514800,
-      unadjustedVolume: 82514800,
-      change: 0.0375,
-      changePercent: 0.08,
-      vwap: 46.61917,
-      label: 'June 20, 18',
-      changeOverTime: 0.0008,
-    },
-    {
-      date: '2018-06-19',
-      open: 46.285,
-      high: 46.5825,
-      low: 45.862499,
-      close: 46.422501,
-      adjClose: 44.75227,
-      volume: 134314000,
-      unadjustedVolume: 134314000,
-      change: 0.1375,
-      changePercent: 0.297,
-      vwap: 46.28917,
-      label: 'June 19, 18',
-      changeOverTime: 0.00297,
-    },
-    {
-      date: '2018-06-18',
-      open: 46.970001,
-      high: 47.305,
-      low: 46.799999,
-      close: 47.185001,
-      adjClose: 45.487331,
-      volume: 73939600,
-      unadjustedVolume: 73939600,
-      change: 0.215,
-      changePercent: 0.458,
-      vwap: 47.09667,
-      label: 'June 18, 18',
-      changeOverTime: 0.00458,
-    },
-    {
-      date: '2018-06-15',
-      open: 47.5075,
-      high: 47.540001,
-      low: 47.064999,
-      close: 47.209999,
-      adjClose: 45.511429,
-      volume: 246876800,
-      unadjustedVolume: 246876800,
-      change: -0.2975,
-      changePercent: -0.626,
-      vwap: 47.27167,
-      label: 'June 15, 18',
-      changeOverTime: -0.00626,
-    },
-    {
-      date: '2018-06-14',
-      open: 47.887501,
-      high: 47.892502,
-      low: 47.555,
-      close: 47.700001,
-      adjClose: 45.983799,
-      volume: 86440400,
-      unadjustedVolume: 86440400,
-      change: -0.1875,
-      changePercent: -0.392,
-      vwap: 47.71583,
-      label: 'June 14, 18',
-      changeOverTime: -0.00392,
-    },
-    {
-      date: '2018-06-13',
-      open: 48.105,
-      high: 48.220001,
-      low: 47.610001,
-      close: 47.674999,
-      adjClose: 45.959702,
-      volume: 86553600,
-      unadjustedVolume: 86553600,
-      change: -0.43,
-      changePercent: -0.894,
-      vwap: 47.835,
-      label: 'June 13, 18',
-      changeOverTime: -0.00894,
-    },
-    {
-      date: '2018-06-12',
-      open: 47.8475,
-      high: 48.1525,
-      low: 47.787498,
-      close: 48.07,
-      adjClose: 46.3405,
-      volume: 67644400,
-      unadjustedVolume: 67644400,
-      change: 0.2225,
-      changePercent: 0.465,
-      vwap: 48.00333,
-      label: 'June 12, 18',
-      changeOverTime: 0.00465,
-    },
-    {
-      date: '2018-06-11',
-      open: 47.837502,
-      high: 47.9925,
-      low: 47.552502,
-      close: 47.807499,
-      adjClose: 46.087437,
-      volume: 73234000,
-      unadjustedVolume: 73234000,
-      change: -0.03,
-      changePercent: -0.063,
-      vwap: 47.78417,
-      label: 'June 11, 18',
-      changeOverTime: -0.00063,
-    },
-    {
-      date: '2018-06-08',
-      open: 47.7925,
-      high: 48.0,
-      low: 47.442501,
-      close: 47.924999,
-      adjClose: 46.200706,
-      volume: 106627200,
-      unadjustedVolume: 106627200,
-      change: 0.1325,
-      changePercent: 0.277,
-      vwap: 47.78917,
-      label: 'June 08, 18',
-      changeOverTime: 0.00277,
-    },
-    {
-      date: '2018-06-07',
-      open: 48.535,
-      high: 48.549999,
-      low: 48.084999,
-      close: 48.365002,
-      adjClose: 46.624882,
-      volume: 85388800,
-      unadjustedVolume: 85388800,
-      change: -0.17,
-      changePercent: -0.35,
-      vwap: 48.33333,
-      label: 'June 07, 18',
-      changeOverTime: -0.0035,
-    },
-    {
-      date: '2018-06-06',
-      open: 48.407501,
-      high: 48.52,
-      low: 47.98,
-      close: 48.494999,
-      adjClose: 46.750198,
-      volume: 83734400,
-      unadjustedVolume: 83734400,
-      change: 0.0875,
-      changePercent: 0.181,
-      vwap: 48.33167,
-      label: 'June 06, 18',
-      changeOverTime: 0.00181,
-    },
-    {
-      date: '2018-06-05',
-      open: 48.267502,
-      high: 48.485001,
-      low: 48.09,
-      close: 48.327499,
-      adjClose: 46.588726,
-      volume: 86264000,
-      unadjustedVolume: 86264000,
-      change: 0.06,
-      changePercent: 0.124,
-      vwap: 48.30083,
-      label: 'June 05, 18',
-      changeOverTime: 0.00124,
-    },
-    {
-      date: '2018-06-04',
-      open: 47.91,
-      high: 48.355,
-      low: 47.837502,
-      close: 47.9575,
-      adjClose: 46.23204,
-      volume: 105064800,
-      unadjustedVolume: 105064800,
-      change: 0.0475,
-      changePercent: 0.099,
-      vwap: 48.05,
-      label: 'June 04, 18',
-      changeOverTime: 0.00099,
-    },
-    {
-      date: '2018-06-01',
-      open: 46.997501,
-      high: 47.564999,
-      low: 46.9375,
-      close: 47.560001,
-      adjClose: 45.84885,
-      volume: 93770000,
-      unadjustedVolume: 93770000,
-      change: 0.5625,
-      changePercent: 1.197,
-      vwap: 47.35417,
-      label: 'June 01, 18',
-      changeOverTime: 0.01197,
-    },
-    {
-      date: '2018-05-31',
-      open: 46.805,
-      high: 47.057499,
-      low: 46.535,
-      close: 46.717499,
-      adjClose: 45.036652,
-      volume: 109931200,
-      unadjustedVolume: 109931200,
-      change: -0.0875,
-      changePercent: -0.187,
-      vwap: 46.77,
-      label: 'May 31, 18',
-      changeOverTime: -0.00187,
-    },
-    {
-      date: '2018-05-30',
-      open: 46.93,
-      high: 47.0,
-      low: 46.695,
-      close: 46.875,
-      adjClose: 45.188492,
-      volume: 74762000,
-      unadjustedVolume: 74762000,
-      change: -0.055,
-      changePercent: -0.117,
-      vwap: 46.85667,
-      label: 'May 30, 18',
-      changeOverTime: -0.00117,
-    },
-    {
-      date: '2018-05-29',
-      open: 46.900002,
-      high: 47.1875,
-      low: 46.717499,
-      close: 46.974998,
-      adjClose: 45.284893,
-      volume: 90056400,
-      unadjustedVolume: 90056400,
-      change: 0.075,
-      changePercent: 0.16,
-      vwap: 46.96,
-      label: 'May 29, 18',
-      changeOverTime: 0.0016,
-    },
-    {
-      date: '2018-05-25',
-      open: 47.057499,
-      high: 47.412498,
-      low: 46.912498,
-      close: 47.145,
-      adjClose: 45.448772,
-      volume: 69844000,
-      unadjustedVolume: 69844000,
-      change: 0.0875,
-      changePercent: 0.186,
-      vwap: 47.15667,
-      label: 'May 25, 18',
-      changeOverTime: 0.00186,
-    },
-    {
-      date: '2018-05-24',
-      open: 47.192501,
-      high: 47.209999,
-      low: 46.552502,
-      close: 47.037498,
-      adjClose: 45.345131,
-      volume: 92936000,
-      unadjustedVolume: 92936000,
-      change: -0.155,
-      changePercent: -0.328,
-      vwap: 46.93333,
-      label: 'May 24, 18',
-      changeOverTime: -0.00328,
-    },
-    {
-      date: '2018-05-23',
-      open: 46.587502,
-      high: 47.125,
-      low: 46.439999,
-      close: 47.09,
-      adjClose: 45.395748,
-      volume: 80233600,
-      unadjustedVolume: 80233600,
-      change: 0.5025,
-      changePercent: 1.079,
-      vwap: 46.885,
-      label: 'May 23, 18',
-      changeOverTime: 0.01079,
-    },
-    {
-      date: '2018-05-22',
-      open: 47.095001,
-      high: 47.220001,
-      low: 46.695,
-      close: 46.790001,
-      adjClose: 45.106541,
-      volume: 60962800,
-      unadjustedVolume: 60962800,
-      change: -0.305,
-      changePercent: -0.648,
-      vwap: 46.90167,
-      label: 'May 22, 18',
-      changeOverTime: -0.00648,
-    },
-    {
-      date: '2018-05-21',
-      open: 47.0,
-      high: 47.317501,
-      low: 46.727501,
-      close: 46.907501,
-      adjClose: 45.219818,
-      volume: 73603200,
-      unadjustedVolume: 73603200,
-      change: -0.0925,
-      changePercent: -0.197,
-      vwap: 46.98417,
-      label: 'May 21, 18',
-      changeOverTime: -0.00197,
-    },
-    {
-      date: '2018-05-18',
-      open: 46.797501,
-      high: 46.952499,
-      low: 46.532501,
-      close: 46.577499,
-      adjClose: 44.901684,
-      volume: 73190800,
-      unadjustedVolume: 73190800,
-      change: -0.22,
-      changePercent: -0.47,
-      vwap: 46.6875,
-      label: 'May 18, 18',
-      changeOverTime: -0.0047,
-    },
-    {
-      date: '2018-05-17',
-      open: 47.0,
-      high: 47.227501,
-      low: 46.59,
-      close: 46.747501,
-      adjClose: 45.065582,
-      volume: 69176000,
-      unadjustedVolume: 69176000,
-      change: -0.2525,
-      changePercent: -0.537,
-      vwap: 46.855,
-      label: 'May 17, 18',
-      changeOverTime: -0.00537,
-    },
-    {
-      date: '2018-05-16',
-      open: 46.517502,
-      high: 47.115002,
-      low: 46.5,
-      close: 47.044998,
-      adjClose: 45.352375,
-      volume: 76732400,
-      unadjustedVolume: 76732400,
-      change: 0.5275,
-      changePercent: 1.134,
-      vwap: 46.88667,
-      label: 'May 16, 18',
-      changeOverTime: 0.01134,
-    },
-    {
-      date: '2018-05-15',
-      open: 46.695,
-      high: 46.767502,
-      low: 46.275002,
-      close: 46.610001,
-      adjClose: 44.933022,
-      volume: 94780800,
-      unadjustedVolume: 94780800,
-      change: -0.085,
-      changePercent: -0.182,
-      vwap: 46.55084,
-      label: 'May 15, 18',
-      changeOverTime: -0.00182,
-    },
-    {
-      date: '2018-05-14',
-      open: 47.252499,
-      high: 47.3825,
-      low: 46.965,
-      close: 47.037498,
-      adjClose: 45.345131,
-      volume: 83115200,
-      unadjustedVolume: 83115200,
-      change: -0.215,
-      changePercent: -0.455,
-      vwap: 47.12833,
-      label: 'May 14, 18',
-      changeOverTime: -0.00455,
-    },
-    {
-      date: '2018-05-11',
-      open: 47.372501,
-      high: 47.514999,
-      low: 46.862499,
-      close: 47.147499,
-      adjClose: 45.451187,
-      volume: 104848800,
-      unadjustedVolume: 104848800,
-      change: -0.225,
-      changePercent: -0.475,
-      vwap: 47.175,
-      label: 'May 11, 18',
-      changeOverTime: -0.00475,
-    },
-    {
-      date: '2018-05-10',
-      open: 46.935001,
-      high: 47.592499,
-      low: 46.912498,
-      close: 47.509998,
-      adjClose: 45.624702,
-      volume: 111957200,
-      unadjustedVolume: 111957200,
-      change: 0.575,
-      changePercent: 1.225,
-      vwap: 47.33833,
-      label: 'May 10, 18',
-      changeOverTime: 0.01225,
-    },
-    {
-      date: '2018-05-09',
-      open: 46.637501,
-      high: 46.849998,
-      low: 46.305,
-      close: 46.84,
-      adjClose: 44.981293,
-      volume: 92844800,
-      unadjustedVolume: 92844800,
-      change: 0.2025,
-      changePercent: 0.434,
-      vwap: 46.665,
-      label: 'May 09, 18',
-      changeOverTime: 0.00434,
-    },
-    {
-      date: '2018-05-08',
-      open: 46.247501,
-      high: 46.555,
-      low: 45.9175,
-      close: 46.512501,
-      adjClose: 44.666786,
-      volume: 113611200,
-      unadjustedVolume: 113611200,
-      change: 0.265,
-      changePercent: 0.573,
-      vwap: 46.32833,
-      label: 'May 08, 18',
-      changeOverTime: 0.00573,
-    },
-    {
-      date: '2018-05-07',
-      open: 46.294998,
-      high: 46.9175,
-      low: 46.1875,
-      close: 46.290001,
-      adjClose: 44.453125,
-      volume: 169805600,
-      unadjustedVolume: 169805600,
-      change: -0.005,
-      changePercent: -0.011,
-      vwap: 46.465,
-      label: 'May 07, 18',
-      changeOverTime: -0.00011,
-    },
-    {
-      date: '2018-05-04',
-      open: 44.5625,
-      high: 46.0625,
-      low: 44.5425,
-      close: 45.9575,
-      adjClose: 44.133804,
-      volume: 224805200,
-      unadjustedVolume: 224805200,
-      change: 1.395,
-      changePercent: 3.13,
-      vwap: 45.52083,
-      label: 'May 04, 18',
-      changeOverTime: 0.0313,
-    },
-    {
-      date: '2018-05-03',
-      open: 43.970001,
-      high: 44.375,
-      low: 43.610001,
-      close: 44.2225,
-      adjClose: 42.467663,
-      volume: 136272800,
-      unadjustedVolume: 136272800,
-      change: 0.2525,
-      changePercent: 0.574,
-      vwap: 44.06917,
-      label: 'May 03, 18',
-      changeOverTime: 0.00574,
-    },
-    {
-      date: '2018-05-02',
-      open: 43.807499,
-      high: 44.4375,
-      low: 43.450001,
-      close: 44.142502,
-      adjClose: 42.390835,
-      volume: 266157600,
-      unadjustedVolume: 266157600,
-      change: 0.335,
-      changePercent: 0.765,
-      vwap: 44.01,
-      label: 'May 02, 18',
-      changeOverTime: 0.00765,
-    },
-    {
-      date: '2018-05-01',
-      open: 41.602501,
-      high: 42.299999,
-      low: 41.317501,
-      close: 42.275002,
-      adjClose: 40.597446,
-      volume: 214277600,
-      unadjustedVolume: 214277600,
-      change: 0.6725,
-      changePercent: 1.616,
-      vwap: 41.96417,
-      label: 'May 01, 18',
-      changeOverTime: 0.01616,
-    },
-    {
-      date: '2018-04-30',
-      open: 40.532501,
-      high: 41.814999,
-      low: 40.459999,
-      close: 41.314999,
-      adjClose: 39.675537,
-      volume: 169709600,
-      unadjustedVolume: 169709600,
-      change: 0.7825,
-      changePercent: 1.931,
-      vwap: 41.19667,
-      label: 'April 30, 18',
-      changeOverTime: 0.01931,
-    },
-    {
-      date: '2018-04-27',
-      open: 41.0,
-      high: 41.0825,
-      low: 40.157501,
-      close: 40.580002,
-      adjClose: 38.969704,
-      volume: 142623200,
-      unadjustedVolume: 142623200,
-      change: -0.42,
-      changePercent: -1.024,
-      vwap: 40.60667,
-      label: 'April 27, 18',
-      changeOverTime: -0.01024,
-    },
-    {
-      date: '2018-04-26',
-      open: 41.029999,
-      high: 41.432499,
-      low: 40.842499,
-      close: 41.055,
-      adjClose: 39.425854,
-      volume: 111852000,
-      unadjustedVolume: 111852000,
-      change: 0.025,
-      changePercent: 0.061,
-      vwap: 41.11,
-      label: 'April 26, 18',
-      changeOverTime: 0.00061,
-    },
-    {
-      date: '2018-04-25',
-      open: 40.654999,
-      high: 41.355,
-      low: 40.602501,
-      close: 40.912498,
-      adjClose: 39.289009,
-      volume: 113528400,
-      unadjustedVolume: 113528400,
-      change: 0.2575,
-      changePercent: 0.633,
-      vwap: 40.95667,
-      label: 'April 25, 18',
-      changeOverTime: 0.00633,
-    },
-    {
-      date: '2018-04-24',
-      open: 41.4175,
-      high: 41.5825,
-      low: 40.305,
-      close: 40.735001,
-      adjClose: 39.118553,
-      volume: 134768000,
-      unadjustedVolume: 134768000,
-      change: -0.6825,
-      changePercent: -1.648,
-      vwap: 40.87417,
-      label: 'April 24, 18',
-      changeOverTime: -0.01648,
-    },
-    {
-      date: '2018-04-23',
-      open: 41.7075,
-      high: 41.73,
-      low: 41.022499,
-      close: 41.310001,
-      adjClose: 39.670742,
-      volume: 146062000,
-      unadjustedVolume: 146062000,
-      change: -0.3975,
-      changePercent: -0.953,
-      vwap: 41.35417,
-      label: 'April 23, 18',
-      changeOverTime: -0.00953,
-    },
-    {
-      date: '2018-04-20',
-      open: 42.650002,
-      high: 42.805,
-      low: 41.357498,
-      close: 41.43,
-      adjClose: 39.78598,
-      volume: 261964400,
-      unadjustedVolume: 261964400,
-      change: -1.22,
-      changePercent: -2.86,
-      vwap: 41.86417,
-      label: 'April 20, 18',
-      changeOverTime: -0.0286,
-    },
-    {
-      date: '2018-04-19',
-      open: 43.439999,
-      high: 43.8475,
-      low: 43.165001,
-      close: 43.200001,
-      adjClose: 41.485737,
-      volume: 139235200,
-      unadjustedVolume: 139235200,
-      change: -0.24,
-      changePercent: -0.552,
-      vwap: 43.40417,
-      label: 'April 19, 18',
-      changeOverTime: -0.00552,
-    },
-    {
-      date: '2018-04-18',
-      open: 44.452499,
-      high: 44.705002,
-      low: 44.220001,
-      close: 44.459999,
-      adjClose: 42.695728,
-      volume: 83018000,
-      unadjustedVolume: 83018000,
-      change: 0.0075,
-      changePercent: 0.017,
-      vwap: 44.46167,
-      label: 'April 18, 18',
-      changeOverTime: 0.00017,
-    },
-    {
-      date: '2018-04-17',
-      open: 44.122501,
-      high: 44.735001,
-      low: 44.102501,
-      close: 44.560001,
-      adjClose: 42.791782,
-      volume: 106421600,
-      unadjustedVolume: 106421600,
-      change: 0.4375,
-      changePercent: 0.992,
-      vwap: 44.46583,
-      label: 'April 17, 18',
-      changeOverTime: 0.00992,
-    },
-    {
-      date: '2018-04-16',
-      open: 43.7575,
-      high: 44.047501,
-      low: 43.7075,
-      close: 43.955002,
-      adjClose: 42.210781,
-      volume: 86313600,
-      unadjustedVolume: 86313600,
-      change: 0.1975,
-      changePercent: 0.451,
-      vwap: 43.90333,
-      label: 'April 16, 18',
-      changeOverTime: 0.00451,
-    },
-    {
-      date: '2018-04-13',
-      open: 43.695,
-      high: 43.959999,
-      low: 43.462502,
-      close: 43.682499,
-      adjClose: 41.949093,
-      volume: 100497200,
-      unadjustedVolume: 100497200,
-      change: -0.0125,
-      changePercent: -0.029,
-      vwap: 43.70167,
-      label: 'April 13, 18',
-      changeOverTime: -0.00029,
-    },
-    {
-      date: '2018-04-12',
-      open: 43.352501,
-      high: 43.75,
-      low: 43.259998,
-      close: 43.535,
-      adjClose: 41.807442,
-      volume: 91557200,
-      unadjustedVolume: 91557200,
-      change: 0.1825,
-      changePercent: 0.421,
-      vwap: 43.515,
-      label: 'April 12, 18',
-      changeOverTime: 0.00421,
-    },
-    {
-      date: '2018-04-11',
-      open: 43.057499,
-      high: 43.48,
-      low: 42.924999,
-      close: 43.110001,
-      adjClose: 41.399307,
-      volume: 89726400,
-      unadjustedVolume: 89726400,
-      change: 0.0525,
-      changePercent: 0.122,
-      vwap: 43.17167,
-      label: 'April 11, 18',
-      changeOverTime: 0.00122,
-    },
-    {
-      date: '2018-04-10',
-      open: 43.25,
-      high: 43.5,
-      low: 42.8825,
-      close: 43.3125,
-      adjClose: 41.593769,
-      volume: 113634400,
-      unadjustedVolume: 113634400,
-      change: 0.0625,
-      changePercent: 0.145,
-      vwap: 43.23167,
-      label: 'April 10, 18',
-      changeOverTime: 0.00145,
-    },
-    {
-      date: '2018-04-09',
-      open: 42.470001,
-      high: 43.272499,
-      low: 42.462502,
-      close: 42.512501,
-      adjClose: 40.825516,
-      volume: 116070800,
-      unadjustedVolume: 116070800,
-      change: 0.0425,
-      changePercent: 0.1,
-      vwap: 42.74917,
-      label: 'April 09, 18',
-      changeOverTime: 0.001,
-    },
-    {
-      date: '2018-04-06',
-      open: 42.7425,
-      high: 43.119999,
-      low: 42.049999,
-      close: 42.095001,
-      adjClose: 40.424587,
-      volume: 140021200,
-      unadjustedVolume: 140021200,
-      change: -0.6475,
-      changePercent: -1.515,
-      vwap: 42.42167,
-      label: 'April 06, 18',
-      changeOverTime: -0.01515,
-    },
-    {
-      date: '2018-04-05',
-      open: 43.145,
-      high: 43.557499,
-      low: 43.02,
-      close: 43.200001,
-      adjClose: 41.485737,
-      volume: 107732800,
-      unadjustedVolume: 107732800,
-      change: 0.055,
-      changePercent: 0.127,
-      vwap: 43.25917,
-      label: 'April 05, 18',
-      changeOverTime: 0.00127,
-    },
-    {
-      date: '2018-04-04',
-      open: 41.220001,
-      high: 43.002499,
-      low: 41.192501,
-      close: 42.9025,
-      adjClose: 41.200039,
-      volume: 138422000,
-      unadjustedVolume: 138422000,
-      change: 1.6825,
-      changePercent: 4.082,
-      vwap: 42.36583,
-      label: 'April 04, 18',
-      changeOverTime: 0.04082,
-    },
-    {
-      date: '2018-04-03',
-      open: 41.91,
-      high: 42.1875,
-      low: 41.220001,
-      close: 42.0975,
-      adjClose: 40.426987,
-      volume: 121112000,
-      unadjustedVolume: 121112000,
-      change: 0.1875,
-      changePercent: 0.447,
-      vwap: 41.835,
-      label: 'April 03, 18',
-      changeOverTime: 0.00447,
-    },
-    {
-      date: '2018-04-02',
-      open: 41.66,
-      high: 42.235001,
-      low: 41.1175,
-      close: 41.669998,
-      adjClose: 40.016449,
-      volume: 150347200,
-      unadjustedVolume: 150347200,
-      change: 0.01,
-      changePercent: 0.024,
-      vwap: 41.67417,
-      label: 'April 02, 18',
-      changeOverTime: 0.00024,
-    },
-    {
-      date: '2018-03-29',
-      open: 41.952499,
-      high: 42.9375,
-      low: 41.724998,
-      close: 41.945,
-      adjClose: 40.280533,
-      volume: 153594000,
-      unadjustedVolume: 153594000,
-      change: -0.0075,
-      changePercent: -0.018,
-      vwap: 42.2025,
-      label: 'March 29, 18',
-      changeOverTime: -0.00018,
-    },
-    {
-      date: '2018-03-28',
-      open: 41.8125,
-      high: 42.505001,
-      low: 41.297501,
-      close: 41.619999,
-      adjClose: 39.968433,
-      volume: 166674000,
-      unadjustedVolume: 166674000,
-      change: -0.1925,
-      changePercent: -0.46,
-      vwap: 41.8075,
-      label: 'March 28, 18',
-      changeOverTime: -0.0046,
-    },
-    {
-      date: '2018-03-27',
-      open: 43.419998,
-      high: 43.787498,
-      low: 41.73,
-      close: 42.084999,
-      adjClose: 40.414982,
-      volume: 163690400,
-      unadjustedVolume: 163690400,
-      change: -1.335,
-      changePercent: -3.075,
-      vwap: 42.53417,
-      label: 'March 27, 18',
-      changeOverTime: -0.03075,
-    },
-    {
-      date: '2018-03-26',
-      open: 42.017502,
-      high: 43.275002,
-      low: 41.610001,
-      close: 43.192501,
-      adjClose: 41.478535,
-      volume: 150164800,
-      unadjustedVolume: 150164800,
-      change: 1.175,
-      changePercent: 2.796,
-      vwap: 42.6925,
-      label: 'March 26, 18',
-      changeOverTime: 0.02796,
-    },
-    {
-      date: '2018-03-23',
-      open: 42.0975,
-      high: 42.48,
-      low: 41.235001,
-      close: 41.235001,
-      adjClose: 39.598709,
-      volume: 164115200,
-      unadjustedVolume: 164115200,
-      change: -0.8625,
-      changePercent: -2.049,
-      vwap: 41.65,
-      label: 'March 23, 18',
-      changeOverTime: -0.02049,
-    },
-    {
-      date: '2018-03-22',
-      open: 42.5,
-      high: 43.169998,
-      low: 42.150002,
-      close: 42.212502,
-      adjClose: 40.537426,
-      volume: 165963200,
-      unadjustedVolume: 165963200,
-      change: -0.2875,
-      changePercent: -0.676,
-      vwap: 42.51083,
-      label: 'March 22, 18',
-      changeOverTime: -0.00676,
-    },
-    {
-      date: '2018-03-21',
-      open: 43.759998,
-      high: 43.772499,
-      low: 42.814999,
-      close: 42.817501,
-      adjClose: 41.11842,
-      volume: 148219600,
-      unadjustedVolume: 148219600,
-      change: -0.9425,
-      changePercent: -2.154,
-      vwap: 43.135,
-      label: 'March 21, 18',
-      changeOverTime: -0.02154,
-    },
-    {
-      date: '2018-03-20',
-      open: 43.810001,
-      high: 44.200001,
-      low: 43.735001,
-      close: 43.810001,
-      adjClose: 42.071533,
-      volume: 78597600,
-      unadjustedVolume: 78597600,
-      change: 0.0,
-      changePercent: 0.0,
-      vwap: 43.915,
-      label: 'March 20, 18',
-      changeOverTime: 0.0,
-    },
-    {
-      date: '2018-03-19',
-      open: 44.330002,
-      high: 44.3675,
-      low: 43.415001,
-      close: 43.825001,
-      adjClose: 42.085938,
-      volume: 133787200,
-      unadjustedVolume: 133787200,
-      change: -0.505,
-      changePercent: -1.139,
-      vwap: 43.86917,
-      label: 'March 19, 18',
-      changeOverTime: -0.01139,
-    },
-    {
-      date: '2018-03-16',
-      open: 44.662498,
-      high: 44.779999,
-      low: 44.404999,
-      close: 44.505001,
-      adjClose: 42.738956,
-      volume: 157618800,
-      unadjustedVolume: 157618800,
-      change: -0.1575,
-      changePercent: -0.353,
-      vwap: 44.56333,
-      label: 'March 16, 18',
-      changeOverTime: -0.00353,
-    },
-    {
-      date: '2018-03-15',
-      open: 44.625,
-      high: 45.060001,
-      low: 44.517502,
-      close: 44.662498,
-      adjClose: 42.890202,
-      volume: 90975200,
-      unadjustedVolume: 90975200,
-      change: 0.0375,
-      changePercent: 0.084,
-      vwap: 44.74667,
-      label: 'March 15, 18',
-      changeOverTime: 0.00084,
-    },
-    {
-      date: '2018-03-14',
-      open: 45.080002,
-      high: 45.130001,
-      low: 44.452499,
-      close: 44.610001,
-      adjClose: 42.839783,
-      volume: 117473600,
-      unadjustedVolume: 117473600,
-      change: -0.47,
-      changePercent: -1.043,
-      vwap: 44.73083,
-      label: 'March 14, 18',
-      changeOverTime: -0.01043,
-    },
-    {
-      date: '2018-03-13',
-      open: 45.647499,
-      high: 45.875,
-      low: 44.810001,
-      close: 44.9925,
-      adjClose: 43.207115,
-      volume: 126774000,
-      unadjustedVolume: 126774000,
-      change: -0.655,
-      changePercent: -1.435,
-      vwap: 45.22583,
-      label: 'March 13, 18',
-      changeOverTime: -0.01435,
-    },
-    {
-      date: '2018-03-12',
-      open: 45.072498,
-      high: 45.5975,
-      low: 45.052502,
-      close: 45.43,
-      adjClose: 43.627251,
-      volume: 128828400,
-      unadjustedVolume: 128828400,
-      change: 0.3575,
-      changePercent: 0.793,
-      vwap: 45.36,
-      label: 'March 12, 18',
-      changeOverTime: 0.00793,
-    },
-  ],
+	symbol: 'AAPL',
+	historical: [
+		{
+			date: '2019-03-12',
+			open: 45.0,
+			high: 45.6675,
+			low: 44.842499,
+			close: 45.227501,
+			adjClose: 44.094143,
+			volume: 129870400,
+			unadjustedVolume: 129870400,
+			change: 0.2275,
+			changePercent: 0.506,
+			vwap: 45.24583,
+			label: 'March 12, 19',
+			changeOverTime: 0.00506,
+		},
+		{
+			date: '2019-03-11',
+			open: 43.872501,
+			high: 44.779999,
+			low: 43.837502,
+			close: 44.724998,
+			adjClose: 43.604237,
+			volume: 128044000,
+			unadjustedVolume: 128044000,
+			change: 0.8525,
+			changePercent: 1.943,
+			vwap: 44.4475,
+			label: 'March 11, 19',
+			changeOverTime: 0.01943,
+		},
+		{
+			date: '2019-03-08',
+			open: 42.580002,
+			high: 43.267502,
+			low: 42.375,
+			close: 43.227501,
+			adjClose: 42.14426,
+			volume: 95997600,
+			unadjustedVolume: 95997600,
+			change: 0.6475,
+			changePercent: 1.521,
+			vwap: 42.95667,
+			label: 'March 08, 19',
+			changeOverTime: 0.01521,
+		},
+		{
+			date: '2019-03-07',
+			open: 43.467499,
+			high: 43.610001,
+			low: 43.005001,
+			close: 43.125,
+			adjClose: 42.044323,
+			volume: 99185600,
+			unadjustedVolume: 99185600,
+			change: -0.3425,
+			changePercent: -0.788,
+			vwap: 43.24667,
+			label: 'March 07, 19',
+			changeOverTime: -0.00788,
+		},
+		{
+			date: '2019-03-06',
+			open: 43.6675,
+			high: 43.872501,
+			low: 43.485001,
+			close: 43.630001,
+			adjClose: 42.536674,
+			volume: 83241600,
+			unadjustedVolume: 83241600,
+			change: -0.0375,
+			changePercent: -0.086,
+			vwap: 43.6625,
+			label: 'March 06, 19',
+			changeOverTime: -0.00086,
+		},
+		{
+			date: '2019-03-05',
+			open: 43.985001,
+			high: 44.0,
+			low: 43.634998,
+			close: 43.8825,
+			adjClose: 42.782837,
+			volume: 78949600,
+			unadjustedVolume: 78949600,
+			change: -0.1025,
+			changePercent: -0.233,
+			vwap: 43.83917,
+			label: 'March 05, 19',
+			changeOverTime: -0.00233,
+		},
+		{
+			date: '2019-03-04',
+			open: 43.922501,
+			high: 44.4375,
+			low: 43.4925,
+			close: 43.962502,
+			adjClose: 42.86084,
+			volume: 109744800,
+			unadjustedVolume: 109744800,
+			change: 0.04,
+			changePercent: 0.091,
+			vwap: 43.96417,
+			label: 'March 04, 19',
+			changeOverTime: 0.00091,
+		},
+		{
+			date: '2019-03-01',
+			open: 43.57,
+			high: 43.787498,
+			low: 43.2225,
+			close: 43.7425,
+			adjClose: 42.646358,
+			volume: 103544800,
+			unadjustedVolume: 103544800,
+			change: 0.1725,
+			changePercent: 0.396,
+			vwap: 43.58417,
+			label: 'March 01, 19',
+			changeOverTime: 0.00396,
+		},
+		{
+			date: '2019-02-28',
+			open: 43.580002,
+			high: 43.727501,
+			low: 43.23,
+			close: 43.287498,
+			adjClose: 42.202747,
+			volume: 112861600,
+			unadjustedVolume: 112861600,
+			change: -0.2925,
+			changePercent: -0.671,
+			vwap: 43.415,
+			label: 'February 28, 19',
+			changeOverTime: -0.00671,
+		},
+		{
+			date: '2019-02-27',
+			open: 43.302502,
+			high: 43.75,
+			low: 43.182499,
+			close: 43.717499,
+			adjClose: 42.621967,
+			volume: 111341600,
+			unadjustedVolume: 111341600,
+			change: 0.415,
+			changePercent: 0.958,
+			vwap: 43.55,
+			label: 'February 27, 19',
+			changeOverTime: 0.00958,
+		},
+		{
+			date: '2019-02-26',
+			open: 43.427502,
+			high: 43.825001,
+			low: 43.2925,
+			close: 43.5825,
+			adjClose: 42.490368,
+			volume: 68280800,
+			unadjustedVolume: 68280800,
+			change: 0.155,
+			changePercent: 0.357,
+			vwap: 43.56667,
+			label: 'February 26, 19',
+			changeOverTime: 0.00357,
+		},
+		{
+			date: '2019-02-25',
+			open: 43.540001,
+			high: 43.967499,
+			low: 43.487499,
+			close: 43.557499,
+			adjClose: 42.465984,
+			volume: 87493600,
+			unadjustedVolume: 87493600,
+			change: 0.0175,
+			changePercent: 0.04,
+			vwap: 43.67083,
+			label: 'February 25, 19',
+			changeOverTime: 0.0004,
+		},
+		{
+			date: '2019-02-22',
+			open: 42.895,
+			high: 43.25,
+			low: 42.845001,
+			close: 43.2425,
+			adjClose: 42.158882,
+			volume: 75652800,
+			unadjustedVolume: 75652800,
+			change: 0.3475,
+			changePercent: 0.81,
+			vwap: 43.1125,
+			label: 'February 22, 19',
+			changeOverTime: 0.0081,
+		},
+		{
+			date: '2019-02-21',
+			open: 42.950001,
+			high: 43.092499,
+			low: 42.575001,
+			close: 42.764999,
+			adjClose: 41.693352,
+			volume: 68998800,
+			unadjustedVolume: 68998800,
+			change: -0.185,
+			changePercent: -0.431,
+			vwap: 42.81083,
+			label: 'February 21, 19',
+			changeOverTime: -0.00431,
+		},
+		{
+			date: '2019-02-20',
+			open: 42.797501,
+			high: 43.330002,
+			low: 42.747501,
+			close: 43.0075,
+			adjClose: 41.929768,
+			volume: 104457600,
+			unadjustedVolume: 104457600,
+			change: 0.21,
+			changePercent: 0.491,
+			vwap: 43.02833,
+			label: 'February 20, 19',
+			changeOverTime: 0.00491,
+		},
+		{
+			date: '2019-02-19',
+			open: 42.427502,
+			high: 42.860001,
+			low: 42.372501,
+			close: 42.732498,
+			adjClose: 41.661659,
+			volume: 75891200,
+			unadjustedVolume: 75891200,
+			change: 0.305,
+			changePercent: 0.719,
+			vwap: 42.655,
+			label: 'February 19, 19',
+			changeOverTime: 0.00719,
+		},
+		{
+			date: '2019-02-15',
+			open: 42.8125,
+			high: 42.924999,
+			low: 42.4375,
+			close: 42.605,
+			adjClose: 41.537365,
+			volume: 98507200,
+			unadjustedVolume: 98507200,
+			change: -0.2075,
+			changePercent: -0.485,
+			vwap: 42.65583,
+			label: 'February 15, 19',
+			changeOverTime: -0.00485,
+		},
+		{
+			date: '2019-02-14',
+			open: 42.427502,
+			high: 42.814999,
+			low: 42.345001,
+			close: 42.700001,
+			adjClose: 41.629974,
+			volume: 87342800,
+			unadjustedVolume: 87342800,
+			change: 0.2725,
+			changePercent: 0.642,
+			vwap: 42.62,
+			label: 'February 14, 19',
+			changeOverTime: 0.00642,
+		},
+		{
+			date: '2019-02-13',
+			open: 42.8475,
+			high: 43.119999,
+			low: 42.48,
+			close: 42.544998,
+			adjClose: 41.478859,
+			volume: 89960800,
+			unadjustedVolume: 89960800,
+			change: -0.3025,
+			changePercent: -0.706,
+			vwap: 42.715,
+			label: 'February 13, 19',
+			changeOverTime: -0.00706,
+		},
+		{
+			date: '2019-02-12',
+			open: 42.525002,
+			high: 42.75,
+			low: 42.424999,
+			close: 42.7225,
+			adjClose: 41.651917,
+			volume: 89134000,
+			unadjustedVolume: 89134000,
+			change: 0.1975,
+			changePercent: 0.464,
+			vwap: 42.6325,
+			label: 'February 12, 19',
+			changeOverTime: 0.00464,
+		},
+		{
+			date: '2019-02-11',
+			open: 42.762501,
+			high: 42.802502,
+			low: 42.3125,
+			close: 42.357498,
+			adjClose: 41.296066,
+			volume: 83973600,
+			unadjustedVolume: 83973600,
+			change: -0.405,
+			changePercent: -0.947,
+			vwap: 42.49083,
+			label: 'February 11, 19',
+			changeOverTime: -0.00947,
+		},
+		{
+			date: '2019-02-08',
+			open: 42.247501,
+			high: 42.665001,
+			low: 42.105,
+			close: 42.602501,
+			adjClose: 41.53492,
+			volume: 95280000,
+			unadjustedVolume: 95280000,
+			change: 0.355,
+			changePercent: 0.84,
+			vwap: 42.4575,
+			label: 'February 08, 19',
+			changeOverTime: 0.0084,
+		},
+		{
+			date: '2019-02-07',
+			open: 43.099998,
+			high: 43.485001,
+			low: 42.584999,
+			close: 42.735001,
+			adjClose: 41.486176,
+			volume: 126966800,
+			unadjustedVolume: 126966800,
+			change: -0.365,
+			changePercent: -0.847,
+			vwap: 42.935,
+			label: 'February 07, 19',
+			changeOverTime: -0.00847,
+		},
+		{
+			date: '2019-02-06',
+			open: 43.662498,
+			high: 43.892502,
+			low: 43.212502,
+			close: 43.560001,
+			adjClose: 42.28706,
+			volume: 112958400,
+			unadjustedVolume: 112958400,
+			change: -0.1025,
+			changePercent: -0.235,
+			vwap: 43.555,
+			label: 'February 06, 19',
+			changeOverTime: -0.00235,
+		},
+		{
+			date: '2019-02-05',
+			open: 43.215,
+			high: 43.77,
+			low: 43.087502,
+			close: 43.544998,
+			adjClose: 42.272499,
+			volume: 144406400,
+			unadjustedVolume: 144406400,
+			change: 0.33,
+			changePercent: 0.764,
+			vwap: 43.4675,
+			label: 'February 05, 19',
+			changeOverTime: 0.00764,
+		},
+		{
+			date: '2019-02-04',
+			open: 41.852501,
+			high: 42.915001,
+			low: 41.82,
+			close: 42.8125,
+			adjClose: 41.561405,
+			volume: 125982000,
+			unadjustedVolume: 125982000,
+			change: 0.96,
+			changePercent: 2.294,
+			vwap: 42.51583,
+			label: 'February 04, 19',
+			changeOverTime: 0.02294,
+		},
+		{
+			date: '2019-02-01',
+			open: 41.740002,
+			high: 42.244999,
+			low: 41.482498,
+			close: 41.630001,
+			adjClose: 40.413464,
+			volume: 130672400,
+			unadjustedVolume: 130672400,
+			change: -0.11,
+			changePercent: -0.264,
+			vwap: 41.78583,
+			label: 'February 01, 19',
+			changeOverTime: -0.00264,
+		},
+		{
+			date: '2019-01-31',
+			open: 41.5275,
+			high: 42.25,
+			low: 41.139999,
+			close: 41.610001,
+			adjClose: 40.394051,
+			volume: 162958400,
+			unadjustedVolume: 162958400,
+			change: 0.0825,
+			changePercent: 0.199,
+			vwap: 41.66667,
+			label: 'January 31, 19',
+			changeOverTime: 0.00199,
+		},
+		{
+			date: '2019-01-30',
+			open: 40.8125,
+			high: 41.537498,
+			low: 40.057499,
+			close: 41.3125,
+			adjClose: 40.10524,
+			volume: 244439200,
+			unadjustedVolume: 244439200,
+			change: 0.5,
+			changePercent: 1.225,
+			vwap: 40.96917,
+			label: 'January 30, 19',
+			changeOverTime: 0.01225,
+		},
+		{
+			date: '2019-01-29',
+			open: 39.0625,
+			high: 39.532501,
+			low: 38.5275,
+			close: 38.669998,
+			adjClose: 37.539959,
+			volume: 166348800,
+			unadjustedVolume: 166348800,
+			change: -0.3925,
+			changePercent: -1.005,
+			vwap: 38.91,
+			label: 'January 29, 19',
+			changeOverTime: -0.01005,
+		},
+		{
+			date: '2019-01-28',
+			open: 38.947498,
+			high: 39.0825,
+			low: 38.415001,
+			close: 39.075001,
+			adjClose: 37.933128,
+			volume: 104768400,
+			unadjustedVolume: 104768400,
+			change: 0.1275,
+			changePercent: 0.327,
+			vwap: 38.8575,
+			label: 'January 28, 19',
+			changeOverTime: 0.00327,
+		},
+		{
+			date: '2019-01-25',
+			open: 38.869999,
+			high: 39.532501,
+			low: 38.580002,
+			close: 39.439999,
+			adjClose: 38.287464,
+			volume: 134142000,
+			unadjustedVolume: 134142000,
+			change: 0.57,
+			changePercent: 1.466,
+			vwap: 39.18417,
+			label: 'January 25, 19',
+			changeOverTime: 0.01466,
+		},
+		{
+			date: '2019-01-24',
+			open: 38.5275,
+			high: 38.619999,
+			low: 37.935001,
+			close: 38.174999,
+			adjClose: 37.059425,
+			volume: 101766000,
+			unadjustedVolume: 101766000,
+			change: -0.3525,
+			changePercent: -0.915,
+			vwap: 38.24333,
+			label: 'January 24, 19',
+			changeOverTime: -0.00915,
+		},
+		{
+			date: '2019-01-23',
+			open: 38.537498,
+			high: 38.785,
+			low: 37.924999,
+			close: 38.48,
+			adjClose: 37.355515,
+			volume: 92522400,
+			unadjustedVolume: 92522400,
+			change: -0.0575,
+			changePercent: -0.149,
+			vwap: 38.39667,
+			label: 'January 23, 19',
+			changeOverTime: -0.00149,
+		},
+		{
+			date: '2019-01-22',
+			open: 39.102501,
+			high: 39.182499,
+			low: 38.154999,
+			close: 38.325001,
+			adjClose: 37.205048,
+			volume: 121576000,
+			unadjustedVolume: 121576000,
+			change: -0.7775,
+			changePercent: -1.988,
+			vwap: 38.55417,
+			label: 'January 22, 19',
+			changeOverTime: -0.01988,
+		},
+		{
+			date: '2019-01-18',
+			open: 39.375,
+			high: 39.470001,
+			low: 38.994999,
+			close: 39.205002,
+			adjClose: 38.059334,
+			volume: 135004000,
+			unadjustedVolume: 135004000,
+			change: -0.17,
+			changePercent: -0.432,
+			vwap: 39.22333,
+			label: 'January 18, 19',
+			changeOverTime: -0.00432,
+		},
+		{
+			date: '2019-01-17',
+			open: 38.549999,
+			high: 39.415001,
+			low: 38.314999,
+			close: 38.965,
+			adjClose: 37.826347,
+			volume: 119284800,
+			unadjustedVolume: 119284800,
+			change: 0.415,
+			changePercent: 1.077,
+			vwap: 38.89833,
+			label: 'January 17, 19',
+			changeOverTime: 0.01077,
+		},
+		{
+			date: '2019-01-16',
+			open: 38.27,
+			high: 38.970001,
+			low: 38.25,
+			close: 38.735001,
+			adjClose: 37.603065,
+			volume: 122278800,
+			unadjustedVolume: 122278800,
+			change: 0.465,
+			changePercent: 1.215,
+			vwap: 38.65167,
+			label: 'January 16, 19',
+			changeOverTime: 0.01215,
+		},
+		{
+			date: '2019-01-15',
+			open: 37.567501,
+			high: 38.3475,
+			low: 37.512501,
+			close: 38.267502,
+			adjClose: 37.149231,
+			volume: 114843600,
+			unadjustedVolume: 114843600,
+			change: 0.7,
+			changePercent: 1.863,
+			vwap: 38.0425,
+			label: 'January 15, 19',
+			changeOverTime: 0.01863,
+		},
+		{
+			date: '2019-01-14',
+			open: 37.712502,
+			high: 37.817501,
+			low: 37.305,
+			close: 37.5,
+			adjClose: 36.404152,
+			volume: 129756800,
+			unadjustedVolume: 129756800,
+			change: -0.2125,
+			changePercent: -0.563,
+			vwap: 37.54083,
+			label: 'January 14, 19',
+			changeOverTime: -0.00563,
+		},
+		{
+			date: '2019-01-11',
+			open: 38.220001,
+			high: 38.424999,
+			low: 37.877499,
+			close: 38.072498,
+			adjClose: 36.959927,
+			volume: 108092800,
+			unadjustedVolume: 108092800,
+			change: -0.1475,
+			changePercent: -0.386,
+			vwap: 38.125,
+			label: 'January 11, 19',
+			changeOverTime: -0.00386,
+		},
+		{
+			date: '2019-01-10',
+			open: 38.125,
+			high: 38.4925,
+			low: 37.715,
+			close: 38.450001,
+			adjClose: 37.326393,
+			volume: 143122800,
+			unadjustedVolume: 143122800,
+			change: 0.325,
+			changePercent: 0.852,
+			vwap: 38.21917,
+			label: 'January 10, 19',
+			changeOverTime: 0.00852,
+		},
+		{
+			date: '2019-01-09',
+			open: 37.822498,
+			high: 38.6325,
+			low: 37.407501,
+			close: 38.327499,
+			adjClose: 37.20747,
+			volume: 180396400,
+			unadjustedVolume: 180396400,
+			change: 0.505,
+			changePercent: 1.335,
+			vwap: 38.1225,
+			label: 'January 09, 19',
+			changeOverTime: 0.01335,
+		},
+		{
+			date: '2019-01-08',
+			open: 37.389999,
+			high: 37.955002,
+			low: 37.130001,
+			close: 37.6875,
+			adjClose: 36.586174,
+			volume: 164101200,
+			unadjustedVolume: 164101200,
+			change: 0.2975,
+			changePercent: 0.796,
+			vwap: 37.59083,
+			label: 'January 08, 19',
+			changeOverTime: 0.00796,
+		},
+		{
+			date: '2019-01-07',
+			open: 37.174999,
+			high: 37.2075,
+			low: 36.474998,
+			close: 36.982498,
+			adjClose: 35.901775,
+			volume: 219111200,
+			unadjustedVolume: 219111200,
+			change: -0.1925,
+			changePercent: -0.518,
+			vwap: 36.88833,
+			label: 'January 07, 19',
+			changeOverTime: -0.00518,
+		},
+		{
+			date: '2019-01-04',
+			open: 36.1325,
+			high: 37.137501,
+			low: 35.950001,
+			close: 37.064999,
+			adjClose: 35.981861,
+			volume: 234428400,
+			unadjustedVolume: 234428400,
+			change: 0.9325,
+			changePercent: 2.581,
+			vwap: 36.7175,
+			label: 'January 04, 19',
+			changeOverTime: 0.02581,
+		},
+		{
+			date: '2019-01-03',
+			open: 35.994999,
+			high: 36.43,
+			low: 35.5,
+			close: 35.547501,
+			adjClose: 34.508717,
+			volume: 365248800,
+			unadjustedVolume: 365248800,
+			change: -0.4475,
+			changePercent: -1.243,
+			vwap: 35.82583,
+			label: 'January 03, 19',
+			changeOverTime: -0.01243,
+		},
+		{
+			date: '2019-01-02',
+			open: 38.7225,
+			high: 39.712502,
+			low: 38.557499,
+			close: 39.48,
+			adjClose: 38.326283,
+			volume: 148158800,
+			unadjustedVolume: 148158800,
+			change: 0.7575,
+			changePercent: 1.956,
+			vwap: 39.25,
+			label: 'January 02, 19',
+			changeOverTime: 0.01956,
+		},
+		{
+			date: '2018-12-31',
+			open: 39.6325,
+			high: 39.84,
+			low: 39.119999,
+			close: 39.435001,
+			adjClose: 38.282612,
+			volume: 140014000,
+			unadjustedVolume: 140014000,
+			change: -0.1975,
+			changePercent: -0.498,
+			vwap: 39.465,
+			label: 'December 31, 18',
+			changeOverTime: -0.00498,
+		},
+		{
+			date: '2018-12-28',
+			open: 39.375,
+			high: 39.630001,
+			low: 38.637501,
+			close: 39.057499,
+			adjClose: 37.916134,
+			volume: 169165600,
+			unadjustedVolume: 169165600,
+			change: -0.3175,
+			changePercent: -0.806,
+			vwap: 39.10833,
+			label: 'December 28, 18',
+			changeOverTime: -0.00806,
+		},
+		{
+			date: '2018-12-27',
+			open: 38.959999,
+			high: 39.192501,
+			low: 37.517502,
+			close: 39.037498,
+			adjClose: 37.896717,
+			volume: 212468400,
+			unadjustedVolume: 212468400,
+			change: 0.0775,
+			changePercent: 0.199,
+			vwap: 38.5825,
+			label: 'December 27, 18',
+			changeOverTime: 0.00199,
+		},
+		{
+			date: '2018-12-26',
+			open: 37.075001,
+			high: 39.307499,
+			low: 36.68,
+			close: 39.2925,
+			adjClose: 38.144268,
+			volume: 234330000,
+			unadjustedVolume: 234330000,
+			change: 2.2175,
+			changePercent: 5.981,
+			vwap: 38.42667,
+			label: 'December 26, 18',
+			changeOverTime: 0.05981,
+		},
+		{
+			date: '2018-12-24',
+			open: 37.037498,
+			high: 37.887501,
+			low: 36.647499,
+			close: 36.7075,
+			adjClose: 35.634811,
+			volume: 148676800,
+			unadjustedVolume: 148676800,
+			change: -0.33,
+			changePercent: -0.891,
+			vwap: 37.08083,
+			label: 'December 24, 18',
+			changeOverTime: -0.00891,
+		},
+		{
+			date: '2018-12-21',
+			open: 39.215,
+			high: 39.540001,
+			low: 37.407501,
+			close: 37.682499,
+			adjClose: 36.581314,
+			volume: 382978400,
+			unadjustedVolume: 382978400,
+			change: -1.5325,
+			changePercent: -3.908,
+			vwap: 38.21,
+			label: 'December 21, 18',
+			changeOverTime: -0.03908,
+		},
+		{
+			date: '2018-12-20',
+			open: 40.099998,
+			high: 40.5275,
+			low: 38.825001,
+			close: 39.2075,
+			adjClose: 38.061756,
+			volume: 259092000,
+			unadjustedVolume: 259092000,
+			change: -0.8925,
+			changePercent: -2.226,
+			vwap: 39.52,
+			label: 'December 20, 18',
+			changeOverTime: -0.02226,
+		},
+		{
+			date: '2018-12-19',
+			open: 41.5,
+			high: 41.862499,
+			low: 39.772499,
+			close: 40.2225,
+			adjClose: 39.0471,
+			volume: 196189200,
+			unadjustedVolume: 196189200,
+			change: -1.2775,
+			changePercent: -3.078,
+			vwap: 40.61917,
+			label: 'December 19, 18',
+			changeOverTime: -0.03078,
+		},
+		{
+			date: '2018-12-18',
+			open: 41.345001,
+			high: 41.8825,
+			low: 41.0975,
+			close: 41.517502,
+			adjClose: 40.304245,
+			volume: 135366000,
+			unadjustedVolume: 135366000,
+			change: 0.1725,
+			changePercent: 0.417,
+			vwap: 41.49917,
+			label: 'December 18, 18',
+			changeOverTime: 0.00417,
+		},
+		{
+			date: '2018-12-17',
+			open: 41.362499,
+			high: 42.087502,
+			low: 40.682499,
+			close: 40.985001,
+			adjClose: 39.787315,
+			volume: 177151600,
+			unadjustedVolume: 177151600,
+			change: -0.3775,
+			changePercent: -0.913,
+			vwap: 41.25167,
+			label: 'December 17, 18',
+			changeOverTime: -0.00913,
+		},
+		{
+			date: '2018-12-14',
+			open: 42.25,
+			high: 42.27,
+			low: 41.32,
+			close: 41.369999,
+			adjClose: 40.161068,
+			volume: 162814800,
+			unadjustedVolume: 162814800,
+			change: -0.88,
+			changePercent: -2.083,
+			vwap: 41.65333,
+			label: 'December 14, 18',
+			changeOverTime: -0.02083,
+		},
+		{
+			date: '2018-12-13',
+			open: 42.622501,
+			high: 43.142502,
+			low: 42.387501,
+			close: 42.737499,
+			adjClose: 41.488594,
+			volume: 127594400,
+			unadjustedVolume: 127594400,
+			change: 0.115,
+			changePercent: 0.27,
+			vwap: 42.75583,
+			label: 'December 13, 18',
+			changeOverTime: 0.0027,
+		},
+		{
+			date: '2018-12-12',
+			open: 42.599998,
+			high: 42.98,
+			low: 42.255001,
+			close: 42.275002,
+			adjClose: 41.039612,
+			volume: 142510800,
+			unadjustedVolume: 142510800,
+			change: -0.325,
+			changePercent: -0.763,
+			vwap: 42.50333,
+			label: 'December 12, 18',
+			changeOverTime: -0.00763,
+		},
+		{
+			date: '2018-12-11',
+			open: 42.915001,
+			high: 42.947498,
+			low: 41.75,
+			close: 42.157501,
+			adjClose: 40.925549,
+			volume: 189126800,
+			unadjustedVolume: 189126800,
+			change: -0.7575,
+			changePercent: -1.765,
+			vwap: 42.285,
+			label: 'December 11, 18',
+			changeOverTime: -0.01765,
+		},
+		{
+			date: '2018-12-10',
+			open: 41.25,
+			high: 42.522499,
+			low: 40.8325,
+			close: 42.400002,
+			adjClose: 41.160965,
+			volume: 248104000,
+			unadjustedVolume: 248104000,
+			change: 1.15,
+			changePercent: 2.788,
+			vwap: 41.91833,
+			label: 'December 10, 18',
+			changeOverTime: 0.02788,
+		},
+		{
+			date: '2018-12-07',
+			open: 43.372501,
+			high: 43.622501,
+			low: 42.075001,
+			close: 42.122501,
+			adjClose: 40.891575,
+			volume: 169126400,
+			unadjustedVolume: 169126400,
+			change: -1.25,
+			changePercent: -2.882,
+			vwap: 42.60667,
+			label: 'December 07, 18',
+			changeOverTime: -0.02882,
+		},
+		{
+			date: '2018-12-06',
+			open: 42.939999,
+			high: 43.695,
+			low: 42.605,
+			close: 43.68,
+			adjClose: 42.403553,
+			volume: 172393600,
+			unadjustedVolume: 172393600,
+			change: 0.74,
+			changePercent: 1.723,
+			vwap: 43.32667,
+			label: 'December 06, 18',
+			changeOverTime: 0.01723,
+		},
+		{
+			date: '2018-12-04',
+			open: 45.237499,
+			high: 45.5975,
+			low: 44.067501,
+			close: 44.172501,
+			adjClose: 42.881664,
+			volume: 165377200,
+			unadjustedVolume: 165377200,
+			change: -1.065,
+			changePercent: -2.354,
+			vwap: 44.6125,
+			label: 'December 04, 18',
+			changeOverTime: -0.02354,
+		},
+		{
+			date: '2018-12-03',
+			open: 46.115002,
+			high: 46.235001,
+			low: 45.302502,
+			close: 46.205002,
+			adjClose: 44.854771,
+			volume: 163210000,
+			unadjustedVolume: 163210000,
+			change: 0.09,
+			changePercent: 0.195,
+			vwap: 45.91417,
+			label: 'December 03, 18',
+			changeOverTime: 0.00195,
+		},
+		{
+			date: '2018-11-30',
+			open: 45.072498,
+			high: 45.0825,
+			low: 44.2575,
+			close: 44.645,
+			adjClose: 43.340359,
+			volume: 158126000,
+			unadjustedVolume: 158126000,
+			change: -0.4275,
+			changePercent: -0.948,
+			vwap: 44.66167,
+			label: 'November 30, 18',
+			changeOverTime: -0.00948,
+		},
+		{
+			date: '2018-11-29',
+			open: 45.665001,
+			high: 45.700001,
+			low: 44.424999,
+			close: 44.887501,
+			adjClose: 43.575775,
+			volume: 167080000,
+			unadjustedVolume: 167080000,
+			change: -0.7775,
+			changePercent: -1.703,
+			vwap: 45.00417,
+			label: 'November 29, 18',
+			changeOverTime: -0.01703,
+		},
+		{
+			date: '2018-11-28',
+			open: 44.182499,
+			high: 45.322498,
+			low: 43.732498,
+			close: 45.235001,
+			adjClose: 43.913113,
+			volume: 184250000,
+			unadjustedVolume: 184250000,
+			change: 1.0525,
+			changePercent: 2.382,
+			vwap: 44.76333,
+			label: 'November 28, 18',
+			changeOverTime: 0.02382,
+		},
+		{
+			date: '2018-11-27',
+			open: 42.877499,
+			high: 43.692501,
+			low: 42.720001,
+			close: 43.560001,
+			adjClose: 42.28706,
+			volume: 165549600,
+			unadjustedVolume: 165549600,
+			change: 0.6825,
+			changePercent: 1.592,
+			vwap: 43.32417,
+			label: 'November 27, 18',
+			changeOverTime: 0.01592,
+		},
+		{
+			date: '2018-11-26',
+			open: 43.560001,
+			high: 43.737499,
+			low: 42.564999,
+			close: 43.654999,
+			adjClose: 42.37928,
+			volume: 179994000,
+			unadjustedVolume: 179994000,
+			change: 0.095,
+			changePercent: 0.218,
+			vwap: 43.31917,
+			label: 'November 26, 18',
+			changeOverTime: 0.00218,
+		},
+		{
+			date: '2018-11-23',
+			open: 43.735001,
+			high: 44.150002,
+			low: 43.025002,
+			close: 43.072498,
+			adjClose: 41.813808,
+			volume: 94496000,
+			unadjustedVolume: 94496000,
+			change: -0.6625,
+			changePercent: -1.515,
+			vwap: 43.41583,
+			label: 'November 23, 18',
+			changeOverTime: -0.01515,
+		},
+		{
+			date: '2018-11-21',
+			open: 44.932499,
+			high: 45.067501,
+			low: 44.137501,
+			close: 44.195,
+			adjClose: 42.903507,
+			volume: 124496800,
+			unadjustedVolume: 124496800,
+			change: -0.7375,
+			changePercent: -1.641,
+			vwap: 44.46667,
+			label: 'November 21, 18',
+			changeOverTime: -0.01641,
+		},
+		{
+			date: '2018-11-20',
+			open: 44.592499,
+			high: 45.3675,
+			low: 43.877499,
+			close: 44.244999,
+			adjClose: 42.952049,
+			volume: 271300800,
+			unadjustedVolume: 271300800,
+			change: -0.3475,
+			changePercent: -0.779,
+			vwap: 44.49667,
+			label: 'November 20, 18',
+			changeOverTime: -0.00779,
+		},
+		{
+			date: '2018-11-19',
+			open: 47.5,
+			high: 47.674999,
+			low: 46.247501,
+			close: 46.465,
+			adjClose: 45.107174,
+			volume: 167701200,
+			unadjustedVolume: 167701200,
+			change: -1.035,
+			changePercent: -2.179,
+			vwap: 46.79583,
+			label: 'November 19, 18',
+			changeOverTime: -0.02179,
+		},
+		{
+			date: '2018-11-16',
+			open: 47.625,
+			high: 48.7425,
+			low: 47.365002,
+			close: 48.3825,
+			adjClose: 46.968639,
+			volume: 147713200,
+			unadjustedVolume: 147713200,
+			change: 0.7575,
+			changePercent: 1.591,
+			vwap: 48.16333,
+			label: 'November 16, 18',
+			changeOverTime: 0.01591,
+		},
+		{
+			date: '2018-11-15',
+			open: 47.0975,
+			high: 47.9925,
+			low: 46.724998,
+			close: 47.852501,
+			adjClose: 46.454128,
+			volume: 185915200,
+			unadjustedVolume: 185915200,
+			change: 0.755,
+			changePercent: 1.603,
+			vwap: 47.52333,
+			label: 'November 15, 18',
+			changeOverTime: 0.01603,
+		},
+		{
+			date: '2018-11-14',
+			open: 48.474998,
+			high: 48.619999,
+			low: 46.482498,
+			close: 46.700001,
+			adjClose: 45.335304,
+			volume: 243204000,
+			unadjustedVolume: 243204000,
+			change: -1.775,
+			changePercent: -3.662,
+			vwap: 47.2675,
+			label: 'November 14, 18',
+			changeOverTime: -0.03662,
+		},
+		{
+			date: '2018-11-13',
+			open: 47.907501,
+			high: 49.294998,
+			low: 47.862499,
+			close: 48.057499,
+			adjClose: 46.653133,
+			volume: 187531600,
+			unadjustedVolume: 187531600,
+			change: 0.15,
+			changePercent: 0.313,
+			vwap: 48.405,
+			label: 'November 13, 18',
+			changeOverTime: 0.00313,
+		},
+		{
+			date: '2018-11-12',
+			open: 49.75,
+			high: 49.962502,
+			low: 48.447498,
+			close: 48.5425,
+			adjClose: 47.123955,
+			volume: 204542000,
+			unadjustedVolume: 204542000,
+			change: -1.2075,
+			changePercent: -2.427,
+			vwap: 48.98417,
+			label: 'November 12, 18',
+			changeOverTime: -0.02427,
+		},
+		{
+			date: '2018-11-09',
+			open: 51.387501,
+			high: 51.502499,
+			low: 50.5625,
+			close: 51.1175,
+			adjClose: 49.623718,
+			volume: 137463200,
+			unadjustedVolume: 137463200,
+			change: -0.27,
+			changePercent: -0.525,
+			vwap: 51.06083,
+			label: 'November 09, 18',
+			changeOverTime: -0.00525,
+		},
+		{
+			date: '2018-11-08',
+			open: 52.494999,
+			high: 52.529999,
+			low: 51.6875,
+			close: 52.122501,
+			adjClose: 50.59935,
+			volume: 101450400,
+			unadjustedVolume: 101450400,
+			change: -0.3725,
+			changePercent: -0.71,
+			vwap: 52.11333,
+			label: 'November 08, 18',
+			changeOverTime: -0.0071,
+		},
+		{
+			date: '2018-11-07',
+			open: 51.4925,
+			high: 52.514999,
+			low: 51.032501,
+			close: 52.487499,
+			adjClose: 50.776508,
+			volume: 133697600,
+			unadjustedVolume: 133697600,
+			change: 0.995,
+			changePercent: 1.932,
+			vwap: 52.01167,
+			label: 'November 07, 18',
+			changeOverTime: 0.01932,
+		},
+		{
+			date: '2018-11-06',
+			open: 50.48,
+			high: 51.18,
+			low: 50.422501,
+			close: 50.942501,
+			adjClose: 49.281883,
+			volume: 127531600,
+			unadjustedVolume: 127531600,
+			change: 0.4625,
+			changePercent: 0.916,
+			vwap: 50.84833,
+			label: 'November 06, 18',
+			changeOverTime: 0.00916,
+		},
+		{
+			date: '2018-11-05',
+			open: 51.075001,
+			high: 51.0975,
+			low: 49.5425,
+			close: 50.397499,
+			adjClose: 48.754639,
+			volume: 264654800,
+			unadjustedVolume: 264654800,
+			change: -0.6775,
+			changePercent: -1.326,
+			vwap: 50.34583,
+			label: 'November 05, 18',
+			changeOverTime: -0.01326,
+		},
+		{
+			date: '2018-11-02',
+			open: 52.387501,
+			high: 53.412498,
+			low: 51.357498,
+			close: 51.869999,
+			adjClose: 50.179142,
+			volume: 365314800,
+			unadjustedVolume: 365314800,
+			change: -0.5175,
+			changePercent: -0.988,
+			vwap: 52.21333,
+			label: 'November 02, 18',
+			changeOverTime: -0.00988,
+		},
+		{
+			date: '2018-11-01',
+			open: 54.762501,
+			high: 55.59,
+			low: 54.202499,
+			close: 55.555,
+			adjClose: 53.744015,
+			volume: 233292800,
+			unadjustedVolume: 233292800,
+			change: 0.7925,
+			changePercent: 1.447,
+			vwap: 55.11583,
+			label: 'November 01, 18',
+			changeOverTime: 0.01447,
+		},
+		{
+			date: '2018-10-31',
+			open: 54.220001,
+			high: 55.112499,
+			low: 54.154999,
+			close: 54.715,
+			adjClose: 52.931404,
+			volume: 153435600,
+			unadjustedVolume: 153435600,
+			change: 0.495,
+			changePercent: 0.913,
+			vwap: 54.66083,
+			label: 'October 31, 18',
+			changeOverTime: 0.00913,
+		},
+		{
+			date: '2018-10-30',
+			open: 52.787498,
+			high: 53.794998,
+			low: 52.317501,
+			close: 53.325001,
+			adjClose: 51.586708,
+			volume: 146640000,
+			unadjustedVolume: 146640000,
+			change: 0.5375,
+			changePercent: 1.018,
+			vwap: 53.14583,
+			label: 'October 30, 18',
+			changeOverTime: 0.01018,
+		},
+		{
+			date: '2018-10-29',
+			open: 54.797501,
+			high: 54.922501,
+			low: 51.522499,
+			close: 53.060001,
+			adjClose: 51.330357,
+			volume: 183742000,
+			unadjustedVolume: 183742000,
+			change: -1.7375,
+			changePercent: -3.171,
+			vwap: 53.16833,
+			label: 'October 29, 18',
+			changeOverTime: -0.03171,
+		},
+		{
+			date: '2018-10-26',
+			open: 53.974998,
+			high: 55.047501,
+			low: 53.1675,
+			close: 54.075001,
+			adjClose: 52.312263,
+			volume: 189033600,
+			unadjustedVolume: 189033600,
+			change: 0.1,
+			changePercent: 0.185,
+			vwap: 54.09667,
+			label: 'October 26, 18',
+			changeOverTime: 0.00185,
+		},
+		{
+			date: '2018-10-25',
+			open: 54.427502,
+			high: 55.345001,
+			low: 54.1875,
+			close: 54.950001,
+			adjClose: 53.158737,
+			volume: 119423200,
+			unadjustedVolume: 119423200,
+			change: 0.5225,
+			changePercent: 0.96,
+			vwap: 54.8275,
+			label: 'October 25, 18',
+			changeOverTime: 0.0096,
+		},
+		{
+			date: '2018-10-24',
+			open: 55.650002,
+			high: 56.057499,
+			low: 53.634998,
+			close: 53.772499,
+			adjClose: 52.019623,
+			volume: 163702000,
+			unadjustedVolume: 163702000,
+			change: -1.8775,
+			changePercent: -3.374,
+			vwap: 54.48833,
+			label: 'October 24, 18',
+			changeOverTime: -0.03374,
+		},
+		{
+			date: '2018-10-23',
+			open: 53.9575,
+			high: 55.8125,
+			low: 53.674999,
+			close: 55.682499,
+			adjClose: 53.867359,
+			volume: 155071200,
+			unadjustedVolume: 155071200,
+			change: 1.725,
+			changePercent: 3.197,
+			vwap: 55.05667,
+			label: 'October 23, 18',
+			changeOverTime: 0.03197,
+		},
+		{
+			date: '2018-10-22',
+			open: 54.947498,
+			high: 55.84,
+			low: 54.735001,
+			close: 55.162498,
+			adjClose: 53.364311,
+			volume: 115168400,
+			unadjustedVolume: 115168400,
+			change: 0.215,
+			changePercent: 0.391,
+			vwap: 55.24583,
+			label: 'October 22, 18',
+			changeOverTime: 0.00391,
+		},
+		{
+			date: '2018-10-19',
+			open: 54.514999,
+			high: 55.314999,
+			low: 54.357498,
+			close: 54.827499,
+			adjClose: 53.040226,
+			volume: 132314800,
+			unadjustedVolume: 132314800,
+			change: 0.3125,
+			changePercent: 0.573,
+			vwap: 54.83333,
+			label: 'October 19, 18',
+			changeOverTime: 0.00573,
+		},
+		{
+			date: '2018-10-18',
+			open: 54.465,
+			high: 54.935001,
+			low: 53.25,
+			close: 54.005001,
+			adjClose: 52.244549,
+			volume: 130325200,
+			unadjustedVolume: 130325200,
+			change: -0.46,
+			changePercent: -0.845,
+			vwap: 54.06333,
+			label: 'October 18, 18',
+			changeOverTime: -0.00845,
+		},
+		{
+			date: '2018-10-17',
+			open: 55.575001,
+			high: 55.66,
+			low: 54.834999,
+			close: 55.297501,
+			adjClose: 53.494915,
+			volume: 91541600,
+			unadjustedVolume: 91541600,
+			change: -0.2775,
+			changePercent: -0.499,
+			vwap: 55.26417,
+			label: 'October 17, 18',
+			changeOverTime: -0.00499,
+		},
+		{
+			date: '2018-10-16',
+			open: 54.732498,
+			high: 55.747501,
+			low: 54.189999,
+			close: 55.537498,
+			adjClose: 53.727085,
+			volume: 116736000,
+			unadjustedVolume: 116736000,
+			change: 0.805,
+			changePercent: 1.471,
+			vwap: 55.15833,
+			label: 'October 16, 18',
+			changeOverTime: 0.01471,
+		},
+		{
+			date: '2018-10-15',
+			open: 55.290001,
+			high: 55.4575,
+			low: 54.317501,
+			close: 54.34,
+			adjClose: 52.568623,
+			volume: 123164000,
+			unadjustedVolume: 123164000,
+			change: -0.95,
+			changePercent: -1.718,
+			vwap: 54.705,
+			label: 'October 15, 18',
+			changeOverTime: -0.01718,
+		},
+		{
+			date: '2018-10-12',
+			open: 55.105,
+			high: 55.720001,
+			low: 54.209999,
+			close: 55.5275,
+			adjClose: 53.717411,
+			volume: 161351600,
+			unadjustedVolume: 161351600,
+			change: 0.4225,
+			changePercent: 0.767,
+			vwap: 55.1525,
+			label: 'October 12, 18',
+			changeOverTime: 0.00767,
+		},
+		{
+			date: '2018-10-11',
+			open: 53.630001,
+			high: 54.875,
+			low: 53.080002,
+			close: 53.612499,
+			adjClose: 51.864838,
+			volume: 212497600,
+			unadjustedVolume: 212497600,
+			change: -0.0175,
+			changePercent: -0.033,
+			vwap: 53.85583,
+			label: 'October 11, 18',
+			changeOverTime: -0.00033,
+		},
+		{
+			date: '2018-10-10',
+			open: 56.365002,
+			high: 56.587502,
+			low: 54.012501,
+			close: 54.09,
+			adjClose: 52.326775,
+			volume: 167962400,
+			unadjustedVolume: 167962400,
+			change: -2.275,
+			changePercent: -4.036,
+			vwap: 54.89667,
+			label: 'October 10, 18',
+			changeOverTime: -0.04036,
+		},
+		{
+			date: '2018-10-09',
+			open: 55.91,
+			high: 56.817501,
+			low: 55.5625,
+			close: 56.717499,
+			adjClose: 54.868626,
+			volume: 107564000,
+			unadjustedVolume: 107564000,
+			change: 0.8075,
+			changePercent: 1.444,
+			vwap: 56.36583,
+			label: 'October 09, 18',
+			changeOverTime: 0.01444,
+		},
+		{
+			date: '2018-10-08',
+			open: 55.552502,
+			high: 56.200001,
+			low: 55.049999,
+			close: 55.942501,
+			adjClose: 54.118893,
+			volume: 118655600,
+			unadjustedVolume: 118655600,
+			change: 0.39,
+			changePercent: 0.702,
+			vwap: 55.73083,
+			label: 'October 08, 18',
+			changeOverTime: 0.00702,
+		},
+		{
+			date: '2018-10-05',
+			open: 56.990002,
+			high: 57.102501,
+			low: 55.145,
+			close: 56.072498,
+			adjClose: 54.24464,
+			volume: 134322000,
+			unadjustedVolume: 134322000,
+			change: -0.9175,
+			changePercent: -1.61,
+			vwap: 56.10667,
+			label: 'October 05, 18',
+			changeOverTime: -0.0161,
+		},
+		{
+			date: '2018-10-04',
+			open: 57.695,
+			high: 58.087502,
+			low: 56.682499,
+			close: 56.997501,
+			adjClose: 55.139496,
+			volume: 128168000,
+			unadjustedVolume: 128168000,
+			change: -0.6975,
+			changePercent: -1.209,
+			vwap: 57.25583,
+			label: 'October 04, 18',
+			changeOverTime: -0.01209,
+		},
+		{
+			date: '2018-10-03',
+			open: 57.512501,
+			high: 58.3675,
+			low: 57.445,
+			close: 58.017502,
+			adjClose: 56.126244,
+			volume: 114619200,
+			unadjustedVolume: 114619200,
+			change: 0.505,
+			changePercent: 0.878,
+			vwap: 57.94333,
+			label: 'October 03, 18',
+			changeOverTime: 0.00878,
+		},
+		{
+			date: '2018-10-02',
+			open: 56.8125,
+			high: 57.5,
+			low: 56.657501,
+			close: 57.32,
+			adjClose: 55.451481,
+			volume: 99152800,
+			unadjustedVolume: 99152800,
+			change: 0.5075,
+			changePercent: 0.893,
+			vwap: 57.15917,
+			label: 'October 02, 18',
+			changeOverTime: 0.00893,
+		},
+		{
+			date: '2018-10-01',
+			open: 56.987499,
+			high: 57.355,
+			low: 56.587502,
+			close: 56.814999,
+			adjClose: 54.962944,
+			volume: 94403200,
+			unadjustedVolume: 94403200,
+			change: -0.1725,
+			changePercent: -0.303,
+			vwap: 56.91917,
+			label: 'October 01, 18',
+			changeOverTime: -0.00303,
+		},
+		{
+			date: '2018-09-28',
+			open: 56.197498,
+			high: 56.459999,
+			low: 56.005001,
+			close: 56.435001,
+			adjClose: 54.595329,
+			volume: 91717600,
+			unadjustedVolume: 91717600,
+			change: 0.2375,
+			changePercent: 0.423,
+			vwap: 56.3,
+			label: 'September 28, 18',
+			changeOverTime: 0.00423,
+		},
+		{
+			date: '2018-09-27',
+			open: 55.955002,
+			high: 56.610001,
+			low: 55.884998,
+			close: 56.237499,
+			adjClose: 54.404266,
+			volume: 120724800,
+			unadjustedVolume: 120724800,
+			change: 0.2825,
+			changePercent: 0.505,
+			vwap: 56.24417,
+			label: 'September 27, 18',
+			changeOverTime: 0.00505,
+		},
+		{
+			date: '2018-09-26',
+			open: 55.25,
+			high: 55.9375,
+			low: 54.939999,
+			close: 55.105,
+			adjClose: 53.308685,
+			volume: 95938800,
+			unadjustedVolume: 95938800,
+			change: -0.145,
+			changePercent: -0.262,
+			vwap: 55.3275,
+			label: 'September 26, 18',
+			changeOverTime: -0.00262,
+		},
+		{
+			date: '2018-09-25',
+			open: 54.9375,
+			high: 55.705002,
+			low: 54.924999,
+			close: 55.547501,
+			adjClose: 53.736759,
+			volume: 98217600,
+			unadjustedVolume: 98217600,
+			change: 0.61,
+			changePercent: 1.11,
+			vwap: 55.3925,
+			label: 'September 25, 18',
+			changeOverTime: 0.0111,
+		},
+		{
+			date: '2018-09-24',
+			open: 54.205002,
+			high: 55.314999,
+			low: 54.157501,
+			close: 55.197498,
+			adjClose: 53.398167,
+			volume: 110773600,
+			unadjustedVolume: 110773600,
+			change: 0.9925,
+			changePercent: 1.831,
+			vwap: 54.89,
+			label: 'September 24, 18',
+			changeOverTime: 0.01831,
+		},
+		{
+			date: '2018-09-21',
+			open: 55.195,
+			high: 55.34,
+			low: 54.322498,
+			close: 54.415001,
+			adjClose: 52.641186,
+			volume: 384986800,
+			unadjustedVolume: 384986800,
+			change: -0.78,
+			changePercent: -1.413,
+			vwap: 54.6925,
+			label: 'September 21, 18',
+			changeOverTime: -0.01413,
+		},
+		{
+			date: '2018-09-20',
+			open: 55.060001,
+			high: 55.57,
+			low: 54.787498,
+			close: 55.0075,
+			adjClose: 53.214363,
+			volume: 106435200,
+			unadjustedVolume: 106435200,
+			change: -0.0525,
+			changePercent: -0.095,
+			vwap: 55.12167,
+			label: 'September 20, 18',
+			changeOverTime: -0.00095,
+		},
+		{
+			date: '2018-09-19',
+			open: 54.625,
+			high: 54.904999,
+			low: 53.825001,
+			close: 54.592499,
+			adjClose: 52.812893,
+			volume: 108495200,
+			unadjustedVolume: 108495200,
+			change: -0.0325,
+			changePercent: -0.059,
+			vwap: 54.44083,
+			label: 'September 19, 18',
+			changeOverTime: -0.00059,
+		},
+		{
+			date: '2018-09-18',
+			open: 54.447498,
+			high: 55.462502,
+			low: 54.279999,
+			close: 54.560001,
+			adjClose: 52.781452,
+			volume: 126286800,
+			unadjustedVolume: 126286800,
+			change: 0.1125,
+			changePercent: 0.207,
+			vwap: 54.7675,
+			label: 'September 18, 18',
+			changeOverTime: 0.00207,
+		},
+		{
+			date: '2018-09-17',
+			open: 55.537498,
+			high: 55.737499,
+			low: 54.317501,
+			close: 54.470001,
+			adjClose: 52.694386,
+			volume: 148780400,
+			unadjustedVolume: 148780400,
+			change: -1.0675,
+			changePercent: -1.922,
+			vwap: 54.84167,
+			label: 'September 17, 18',
+			changeOverTime: -0.01922,
+		},
+		{
+			date: '2018-09-14',
+			open: 56.4375,
+			high: 56.709999,
+			low: 55.630001,
+			close: 55.959999,
+			adjClose: 54.135818,
+			volume: 127997200,
+			unadjustedVolume: 127997200,
+			change: -0.4775,
+			changePercent: -0.846,
+			vwap: 56.1,
+			label: 'September 14, 18',
+			changeOverTime: -0.00846,
+		},
+		{
+			date: '2018-09-13',
+			open: 55.880001,
+			high: 57.087502,
+			low: 55.642502,
+			close: 56.602501,
+			adjClose: 54.757366,
+			volume: 166825600,
+			unadjustedVolume: 166825600,
+			change: 0.7225,
+			changePercent: 1.293,
+			vwap: 56.44417,
+			label: 'September 13, 18',
+			changeOverTime: 0.01293,
+		},
+		{
+			date: '2018-09-12',
+			open: 56.235001,
+			high: 56.25,
+			low: 54.959999,
+			close: 55.267502,
+			adjClose: 53.465893,
+			volume: 197114800,
+			unadjustedVolume: 197114800,
+			change: -0.9675,
+			changePercent: -1.72,
+			vwap: 55.4925,
+			label: 'September 12, 18',
+			changeOverTime: -0.0172,
+		},
+		{
+			date: '2018-09-11',
+			open: 54.502499,
+			high: 56.075001,
+			low: 54.139999,
+			close: 55.962502,
+			adjClose: 54.138237,
+			volume: 142996000,
+			unadjustedVolume: 142996000,
+			change: 1.46,
+			changePercent: 2.679,
+			vwap: 55.3925,
+			label: 'September 11, 18',
+			changeOverTime: 0.02679,
+		},
+		{
+			date: '2018-09-10',
+			open: 55.237499,
+			high: 55.462502,
+			low: 54.1175,
+			close: 54.5825,
+			adjClose: 52.803215,
+			volume: 158066000,
+			unadjustedVolume: 158066000,
+			change: -0.655,
+			changePercent: -1.186,
+			vwap: 54.72083,
+			label: 'September 10, 18',
+			changeOverTime: -0.01186,
+		},
+		{
+			date: '2018-09-07',
+			open: 55.462502,
+			high: 56.342499,
+			low: 55.177502,
+			close: 55.325001,
+			adjClose: 53.521519,
+			volume: 150479200,
+			unadjustedVolume: 150479200,
+			change: -0.1375,
+			changePercent: -0.248,
+			vwap: 55.615,
+			label: 'September 07, 18',
+			changeOverTime: -0.00248,
+		},
+		{
+			date: '2018-09-06',
+			open: 56.557499,
+			high: 56.837502,
+			low: 55.325001,
+			close: 55.775002,
+			adjClose: 53.956841,
+			volume: 137160000,
+			unadjustedVolume: 137160000,
+			change: -0.7825,
+			changePercent: -1.384,
+			vwap: 55.97917,
+			label: 'September 06, 18',
+			changeOverTime: -0.01384,
+		},
+		{
+			date: '2018-09-05',
+			open: 57.247501,
+			high: 57.4175,
+			low: 56.275002,
+			close: 56.717499,
+			adjClose: 54.868626,
+			volume: 133332000,
+			unadjustedVolume: 133332000,
+			change: -0.53,
+			changePercent: -0.926,
+			vwap: 56.80333,
+			label: 'September 05, 18',
+			changeOverTime: -0.00926,
+		},
+		{
+			date: '2018-09-04',
+			open: 57.102501,
+			high: 57.294998,
+			low: 56.657501,
+			close: 57.09,
+			adjClose: 55.228977,
+			volume: 109560400,
+			unadjustedVolume: 109560400,
+			change: -0.0125,
+			changePercent: -0.022,
+			vwap: 57.01417,
+			label: 'September 04, 18',
+			changeOverTime: -0.00022,
+		},
+		{
+			date: '2018-08-31',
+			open: 56.627499,
+			high: 57.217499,
+			low: 56.5,
+			close: 56.907501,
+			adjClose: 55.052429,
+			volume: 173360400,
+			unadjustedVolume: 173360400,
+			change: 0.28,
+			changePercent: 0.494,
+			vwap: 56.875,
+			label: 'August 31, 18',
+			changeOverTime: 0.00494,
+		},
+		{
+			date: '2018-08-30',
+			open: 55.8125,
+			high: 57.064999,
+			low: 55.599998,
+			close: 56.2575,
+			adjClose: 54.423618,
+			volume: 195175200,
+			unadjustedVolume: 195175200,
+			change: 0.445,
+			changePercent: 0.797,
+			vwap: 56.3075,
+			label: 'August 30, 18',
+			changeOverTime: 0.00797,
+		},
+		{
+			date: '2018-08-29',
+			open: 55.037498,
+			high: 55.872501,
+			low: 54.852501,
+			close: 55.744999,
+			adjClose: 53.927822,
+			volume: 109019200,
+			unadjustedVolume: 109019200,
+			change: 0.7075,
+			changePercent: 1.285,
+			vwap: 55.49,
+			label: 'August 29, 18',
+			changeOverTime: 0.01285,
+		},
+		{
+			date: '2018-08-28',
+			open: 54.752499,
+			high: 55.134998,
+			low: 54.73,
+			close: 54.924999,
+			adjClose: 53.134556,
+			volume: 91107200,
+			unadjustedVolume: 91107200,
+			change: 0.1725,
+			changePercent: 0.315,
+			vwap: 54.93,
+			label: 'August 28, 18',
+			changeOverTime: 0.00315,
+		},
+		{
+			date: '2018-08-27',
+			open: 54.287498,
+			high: 54.685001,
+			low: 54.0825,
+			close: 54.485001,
+			adjClose: 52.708897,
+			volume: 82100400,
+			unadjustedVolume: 82100400,
+			change: 0.1975,
+			changePercent: 0.364,
+			vwap: 54.4175,
+			label: 'August 27, 18',
+			changeOverTime: 0.00364,
+		},
+		{
+			date: '2018-08-24',
+			open: 54.150002,
+			high: 54.224998,
+			low: 53.7775,
+			close: 54.040001,
+			adjClose: 52.278408,
+			volume: 73905600,
+			unadjustedVolume: 73905600,
+			change: -0.11,
+			changePercent: -0.203,
+			vwap: 54.01417,
+			label: 'August 24, 18',
+			changeOverTime: -0.00203,
+		},
+		{
+			date: '2018-08-23',
+			open: 53.662498,
+			high: 54.262501,
+			low: 53.650002,
+			close: 53.872501,
+			adjClose: 52.116367,
+			volume: 75532800,
+			unadjustedVolume: 75532800,
+			change: 0.21,
+			changePercent: 0.391,
+			vwap: 53.92833,
+			label: 'August 23, 18',
+			changeOverTime: 0.00391,
+		},
+		{
+			date: '2018-08-22',
+			open: 53.525002,
+			high: 54.09,
+			low: 53.459999,
+			close: 53.762501,
+			adjClose: 52.009953,
+			volume: 76072400,
+			unadjustedVolume: 76072400,
+			change: 0.2375,
+			changePercent: 0.444,
+			vwap: 53.77083,
+			label: 'August 22, 18',
+			changeOverTime: 0.00444,
+		},
+		{
+			date: '2018-08-21',
+			open: 54.200001,
+			high: 54.297501,
+			low: 53.5075,
+			close: 53.759998,
+			adjClose: 52.00753,
+			volume: 104639200,
+			unadjustedVolume: 104639200,
+			change: -0.44,
+			changePercent: -0.812,
+			vwap: 53.855,
+			label: 'August 21, 18',
+			changeOverTime: -0.00812,
+		},
+		{
+			date: '2018-08-20',
+			open: 54.525002,
+			high: 54.794998,
+			low: 53.7775,
+			close: 53.865002,
+			adjClose: 52.109116,
+			volume: 121150800,
+			unadjustedVolume: 121150800,
+			change: -0.66,
+			changePercent: -1.21,
+			vwap: 54.14583,
+			label: 'August 20, 18',
+			changeOverTime: -0.0121,
+		},
+		{
+			date: '2018-08-17',
+			open: 53.360001,
+			high: 54.487499,
+			low: 53.290001,
+			close: 54.395,
+			adjClose: 52.621826,
+			volume: 141708000,
+			unadjustedVolume: 141708000,
+			change: 1.035,
+			changePercent: 1.94,
+			vwap: 54.0575,
+			label: 'August 17, 18',
+			changeOverTime: 0.0194,
+		},
+		{
+			date: '2018-08-16',
+			open: 52.9375,
+			high: 53.452499,
+			low: 52.8675,
+			close: 53.330002,
+			adjClose: 51.591545,
+			volume: 114001600,
+			unadjustedVolume: 114001600,
+			change: 0.3925,
+			changePercent: 0.741,
+			vwap: 53.21667,
+			label: 'August 16, 18',
+			changeOverTime: 0.00741,
+		},
+		{
+			date: '2018-08-15',
+			open: 52.305,
+			high: 52.685001,
+			low: 52.0825,
+			close: 52.560001,
+			adjClose: 50.846649,
+			volume: 115230400,
+			unadjustedVolume: 115230400,
+			change: 0.255,
+			changePercent: 0.488,
+			vwap: 52.4425,
+			label: 'August 15, 18',
+			changeOverTime: 0.00488,
+		},
+		{
+			date: '2018-08-14',
+			open: 52.540001,
+			high: 52.639999,
+			low: 52.064999,
+			close: 52.4375,
+			adjClose: 50.728138,
+			volume: 82992000,
+			unadjustedVolume: 82992000,
+			change: -0.1025,
+			changePercent: -0.195,
+			vwap: 52.38083,
+			label: 'August 14, 18',
+			changeOverTime: -0.00195,
+		},
+		{
+			date: '2018-08-13',
+			open: 52.327499,
+			high: 52.737499,
+			low: 51.924999,
+			close: 52.217499,
+			adjClose: 50.515312,
+			volume: 103563600,
+			unadjustedVolume: 103563600,
+			change: -0.11,
+			changePercent: -0.21,
+			vwap: 52.29333,
+			label: 'August 13, 18',
+			changeOverTime: -0.0021,
+		},
+		{
+			date: '2018-08-10',
+			open: 51.84,
+			high: 52.275002,
+			low: 51.6675,
+			close: 51.8825,
+			adjClose: 50.191231,
+			volume: 98444800,
+			unadjustedVolume: 98444800,
+			change: 0.0425,
+			changePercent: 0.082,
+			vwap: 51.94167,
+			label: 'August 10, 18',
+			changeOverTime: 0.00082,
+		},
+		{
+			date: '2018-08-09',
+			open: 52.3825,
+			high: 52.445,
+			low: 51.799999,
+			close: 52.220001,
+			adjClose: 50.341187,
+			volume: 93970400,
+			unadjustedVolume: 93970400,
+			change: -0.1625,
+			changePercent: -0.31,
+			vwap: 52.155,
+			label: 'August 09, 18',
+			changeOverTime: -0.0031,
+		},
+		{
+			date: '2018-08-08',
+			open: 51.512501,
+			high: 51.952499,
+			low: 51.130001,
+			close: 51.8125,
+			adjClose: 49.948341,
+			volume: 90102000,
+			unadjustedVolume: 90102000,
+			change: 0.3,
+			changePercent: 0.582,
+			vwap: 51.63167,
+			label: 'August 08, 18',
+			changeOverTime: 0.00582,
+		},
+		{
+			date: '2018-08-07',
+			open: 52.330002,
+			high: 52.375,
+			low: 51.689999,
+			close: 51.7775,
+			adjClose: 49.9146,
+			volume: 102349600,
+			unadjustedVolume: 102349600,
+			change: -0.5525,
+			changePercent: -1.056,
+			vwap: 51.9475,
+			label: 'August 07, 18',
+			changeOverTime: -0.01056,
+		},
+		{
+			date: '2018-08-06',
+			open: 52.0,
+			high: 52.3125,
+			low: 51.767502,
+			close: 52.267502,
+			adjClose: 50.386971,
+			volume: 101701600,
+			unadjustedVolume: 101701600,
+			change: 0.2675,
+			changePercent: 0.514,
+			vwap: 52.11583,
+			label: 'August 06, 18',
+			changeOverTime: 0.00514,
+		},
+		{
+			date: '2018-08-03',
+			open: 51.7575,
+			high: 52.185001,
+			low: 51.369999,
+			close: 51.997501,
+			adjClose: 50.126686,
+			volume: 133789600,
+			unadjustedVolume: 133789600,
+			change: 0.24,
+			changePercent: 0.464,
+			vwap: 51.85083,
+			label: 'August 03, 18',
+			changeOverTime: 0.00464,
+		},
+		{
+			date: '2018-08-02',
+			open: 50.145,
+			high: 52.095001,
+			low: 50.087502,
+			close: 51.8475,
+			adjClose: 49.982079,
+			volume: 249616000,
+			unadjustedVolume: 249616000,
+			change: 1.7025,
+			changePercent: 3.395,
+			vwap: 51.34333,
+			label: 'August 02, 18',
+			changeOverTime: 0.03395,
+		},
+		{
+			date: '2018-08-01',
+			open: 49.782501,
+			high: 50.439999,
+			low: 49.327499,
+			close: 50.375,
+			adjClose: 48.562565,
+			volume: 271742800,
+			unadjustedVolume: 271742800,
+			change: 0.5925,
+			changePercent: 1.19,
+			vwap: 50.0475,
+			label: 'August 01, 18',
+			changeOverTime: 0.0119,
+		},
+		{
+			date: '2018-07-31',
+			open: 47.575001,
+			high: 48.035,
+			low: 47.334999,
+			close: 47.572498,
+			adjClose: 45.860886,
+			volume: 157492000,
+			unadjustedVolume: 157492000,
+			change: -0.0025,
+			changePercent: -0.005,
+			vwap: 47.6475,
+			label: 'July 31, 18',
+			changeOverTime: -0.00005,
+		},
+		{
+			date: '2018-07-30',
+			open: 47.974998,
+			high: 48.049999,
+			low: 47.267502,
+			close: 47.477501,
+			adjClose: 45.769306,
+			volume: 84118000,
+			unadjustedVolume: 84118000,
+			change: -0.4975,
+			changePercent: -1.037,
+			vwap: 47.59833,
+			label: 'July 30, 18',
+			changeOverTime: -0.01037,
+		},
+		{
+			date: '2018-07-27',
+			open: 48.747501,
+			high: 48.797501,
+			low: 47.525002,
+			close: 47.744999,
+			adjClose: 46.027184,
+			volume: 96096000,
+			unadjustedVolume: 96096000,
+			change: -1.0025,
+			changePercent: -2.057,
+			vwap: 48.0225,
+			label: 'July 27, 18',
+			changeOverTime: -0.02057,
+		},
+		{
+			date: '2018-07-26',
+			open: 48.6525,
+			high: 48.990002,
+			low: 48.4025,
+			close: 48.552502,
+			adjClose: 46.805634,
+			volume: 76304000,
+			unadjustedVolume: 76304000,
+			change: -0.1,
+			changePercent: -0.206,
+			vwap: 48.64833,
+			label: 'July 26, 18',
+			changeOverTime: -0.00206,
+		},
+		{
+			date: '2018-07-25',
+			open: 48.264999,
+			high: 48.712502,
+			low: 48.107498,
+			close: 48.705002,
+			adjClose: 46.952641,
+			volume: 66839600,
+			unadjustedVolume: 66839600,
+			change: 0.44,
+			changePercent: 0.912,
+			vwap: 48.50833,
+			label: 'July 25, 18',
+			changeOverTime: 0.00912,
+		},
+		{
+			date: '2018-07-24',
+			open: 48.112499,
+			high: 48.415001,
+			low: 48.012501,
+			close: 48.25,
+			adjClose: 46.514015,
+			volume: 74791600,
+			unadjustedVolume: 74791600,
+			change: 0.1375,
+			changePercent: 0.286,
+			vwap: 48.22583,
+			label: 'July 24, 18',
+			changeOverTime: 0.00286,
+		},
+		{
+			date: '2018-07-23',
+			open: 47.669998,
+			high: 47.990002,
+			low: 47.389999,
+			close: 47.9025,
+			adjClose: 46.179016,
+			volume: 63957600,
+			unadjustedVolume: 63957600,
+			change: 0.2325,
+			changePercent: 0.488,
+			vwap: 47.76083,
+			label: 'July 23, 18',
+			changeOverTime: 0.00488,
+		},
+		{
+			date: '2018-07-20',
+			open: 47.945,
+			high: 48.107498,
+			low: 47.5425,
+			close: 47.860001,
+			adjClose: 46.138042,
+			volume: 82704800,
+			unadjustedVolume: 82704800,
+			change: -0.085,
+			changePercent: -0.177,
+			vwap: 47.83667,
+			label: 'July 20, 18',
+			changeOverTime: -0.00177,
+		},
+		{
+			date: '2018-07-19',
+			open: 47.422501,
+			high: 48.137501,
+			low: 47.422501,
+			close: 47.970001,
+			adjClose: 46.244091,
+			volume: 81147200,
+			unadjustedVolume: 81147200,
+			change: 0.5475,
+			changePercent: 1.155,
+			vwap: 47.84333,
+			label: 'July 19, 18',
+			changeOverTime: 0.01155,
+		},
+		{
+			date: '2018-07-18',
+			open: 47.945,
+			high: 47.950001,
+			low: 47.482498,
+			close: 47.599998,
+			adjClose: 45.887398,
+			volume: 65573600,
+			unadjustedVolume: 65573600,
+			change: -0.345,
+			changePercent: -0.72,
+			vwap: 47.6775,
+			label: 'July 18, 18',
+			changeOverTime: -0.0072,
+		},
+		{
+			date: '2018-07-17',
+			open: 47.4375,
+			high: 47.967499,
+			low: 47.299999,
+			close: 47.862499,
+			adjClose: 46.140453,
+			volume: 62138000,
+			unadjustedVolume: 62138000,
+			change: 0.425,
+			changePercent: 0.896,
+			vwap: 47.71,
+			label: 'July 17, 18',
+			changeOverTime: 0.00896,
+		},
+		{
+			date: '2018-07-16',
+			open: 47.880001,
+			high: 48.162498,
+			low: 47.605,
+			close: 47.727501,
+			adjClose: 46.010319,
+			volume: 60172400,
+			unadjustedVolume: 60172400,
+			change: -0.1525,
+			changePercent: -0.319,
+			vwap: 47.83167,
+			label: 'July 16, 18',
+			changeOverTime: -0.00319,
+		},
+		{
+			date: '2018-07-13',
+			open: 47.77,
+			high: 47.959999,
+			low: 47.724998,
+			close: 47.8325,
+			adjClose: 46.111538,
+			volume: 50055600,
+			unadjustedVolume: 50055600,
+			change: 0.0625,
+			changePercent: 0.131,
+			vwap: 47.83917,
+			label: 'July 13, 18',
+			changeOverTime: 0.00131,
+		},
+		{
+			date: '2018-07-12',
+			open: 47.3825,
+			high: 47.852501,
+			low: 47.327499,
+			close: 47.7575,
+			adjClose: 46.03923,
+			volume: 72164400,
+			unadjustedVolume: 72164400,
+			change: 0.375,
+			changePercent: 0.791,
+			vwap: 47.64583,
+			label: 'July 12, 18',
+			changeOverTime: 0.00791,
+		},
+		{
+			date: '2018-07-11',
+			open: 47.125,
+			high: 47.445,
+			low: 46.9025,
+			close: 46.970001,
+			adjClose: 45.280067,
+			volume: 75326000,
+			unadjustedVolume: 75326000,
+			change: -0.155,
+			changePercent: -0.329,
+			vwap: 47.10583,
+			label: 'July 11, 18',
+			changeOverTime: -0.00329,
+		},
+		{
+			date: '2018-07-10',
+			open: 47.677502,
+			high: 47.82,
+			low: 47.544998,
+			close: 47.587502,
+			adjClose: 45.875359,
+			volume: 63756400,
+			unadjustedVolume: 63756400,
+			change: -0.09,
+			changePercent: -0.189,
+			vwap: 47.65083,
+			label: 'July 10, 18',
+			changeOverTime: -0.00189,
+		},
+		{
+			date: '2018-07-09',
+			open: 47.375,
+			high: 47.669998,
+			low: 47.325001,
+			close: 47.645,
+			adjClose: 45.930782,
+			volume: 79026400,
+			unadjustedVolume: 79026400,
+			change: 0.27,
+			changePercent: 0.57,
+			vwap: 47.54667,
+			label: 'July 09, 18',
+			changeOverTime: 0.0057,
+		},
+		{
+			date: '2018-07-06',
+			open: 46.355,
+			high: 47.107498,
+			low: 46.299999,
+			close: 46.9925,
+			adjClose: 45.301758,
+			volume: 69940800,
+			unadjustedVolume: 69940800,
+			change: 0.6375,
+			changePercent: 1.375,
+			vwap: 46.8,
+			label: 'July 06, 18',
+			changeOverTime: 0.01375,
+		},
+		{
+			date: '2018-07-05',
+			open: 46.314999,
+			high: 46.602501,
+			low: 46.07,
+			close: 46.349998,
+			adjClose: 44.682373,
+			volume: 66416800,
+			unadjustedVolume: 66416800,
+			change: 0.035,
+			changePercent: 0.076,
+			vwap: 46.34083,
+			label: 'July 05, 18',
+			changeOverTime: 0.00076,
+		},
+		{
+			date: '2018-07-03',
+			open: 46.947498,
+			high: 46.987499,
+			low: 45.884998,
+			close: 45.98,
+			adjClose: 44.325691,
+			volume: 55819200,
+			unadjustedVolume: 55819200,
+			change: -0.9675,
+			changePercent: -2.061,
+			vwap: 46.28417,
+			label: 'July 03, 18',
+			changeOverTime: -0.02061,
+		},
+		{
+			date: '2018-07-02',
+			open: 45.955002,
+			high: 46.825001,
+			low: 45.855,
+			close: 46.794998,
+			adjClose: 45.111362,
+			volume: 70925200,
+			unadjustedVolume: 70925200,
+			change: 0.84,
+			changePercent: 1.828,
+			vwap: 46.49167,
+			label: 'July 02, 18',
+			changeOverTime: 0.01828,
+		},
+		{
+			date: '2018-06-29',
+			open: 46.572498,
+			high: 46.797501,
+			low: 45.727501,
+			close: 46.2775,
+			adjClose: 44.61248,
+			volume: 90950800,
+			unadjustedVolume: 90950800,
+			change: -0.295,
+			changePercent: -0.633,
+			vwap: 46.2675,
+			label: 'June 29, 18',
+			changeOverTime: -0.00633,
+		},
+		{
+			date: '2018-06-28',
+			open: 46.025002,
+			high: 46.552502,
+			low: 45.950001,
+			close: 46.375,
+			adjClose: 44.70647,
+			volume: 69460800,
+			unadjustedVolume: 69460800,
+			change: 0.35,
+			changePercent: 0.76,
+			vwap: 46.2925,
+			label: 'June 28, 18',
+			changeOverTime: 0.0076,
+		},
+		{
+			date: '2018-06-27',
+			open: 46.307499,
+			high: 46.82,
+			low: 46.0075,
+			close: 46.040001,
+			adjClose: 44.38353,
+			volume: 101141200,
+			unadjustedVolume: 101141200,
+			change: -0.2675,
+			changePercent: -0.578,
+			vwap: 46.28917,
+			label: 'June 27, 18',
+			changeOverTime: -0.00578,
+		},
+		{
+			date: '2018-06-26',
+			open: 45.747501,
+			high: 46.6325,
+			low: 45.634998,
+			close: 46.107498,
+			adjClose: 44.448593,
+			volume: 98276800,
+			unadjustedVolume: 98276800,
+			change: 0.36,
+			changePercent: 0.787,
+			vwap: 46.125,
+			label: 'June 26, 18',
+			changeOverTime: 0.00787,
+		},
+		{
+			date: '2018-06-25',
+			open: 45.849998,
+			high: 46.23,
+			low: 45.182499,
+			close: 45.5425,
+			adjClose: 43.903927,
+			volume: 126652400,
+			unadjustedVolume: 126652400,
+			change: -0.3075,
+			changePercent: -0.671,
+			vwap: 45.65167,
+			label: 'June 25, 18',
+			changeOverTime: -0.00671,
+		},
+		{
+			date: '2018-06-22',
+			open: 46.529999,
+			high: 46.537498,
+			low: 46.174999,
+			close: 46.23,
+			adjClose: 44.566692,
+			volume: 108801600,
+			unadjustedVolume: 108801600,
+			change: -0.3,
+			changePercent: -0.645,
+			vwap: 46.31417,
+			label: 'June 22, 18',
+			changeOverTime: -0.00645,
+		},
+		{
+			date: '2018-06-21',
+			open: 46.8125,
+			high: 47.087502,
+			low: 46.235001,
+			close: 46.365002,
+			adjClose: 44.696838,
+			volume: 102847600,
+			unadjustedVolume: 102847600,
+			change: -0.4475,
+			changePercent: -0.956,
+			vwap: 46.5625,
+			label: 'June 21, 18',
+			changeOverTime: -0.00956,
+		},
+		{
+			date: '2018-06-20',
+			open: 46.587502,
+			high: 46.799999,
+			low: 46.432499,
+			close: 46.625,
+			adjClose: 44.947487,
+			volume: 82514800,
+			unadjustedVolume: 82514800,
+			change: 0.0375,
+			changePercent: 0.08,
+			vwap: 46.61917,
+			label: 'June 20, 18',
+			changeOverTime: 0.0008,
+		},
+		{
+			date: '2018-06-19',
+			open: 46.285,
+			high: 46.5825,
+			low: 45.862499,
+			close: 46.422501,
+			adjClose: 44.75227,
+			volume: 134314000,
+			unadjustedVolume: 134314000,
+			change: 0.1375,
+			changePercent: 0.297,
+			vwap: 46.28917,
+			label: 'June 19, 18',
+			changeOverTime: 0.00297,
+		},
+		{
+			date: '2018-06-18',
+			open: 46.970001,
+			high: 47.305,
+			low: 46.799999,
+			close: 47.185001,
+			adjClose: 45.487331,
+			volume: 73939600,
+			unadjustedVolume: 73939600,
+			change: 0.215,
+			changePercent: 0.458,
+			vwap: 47.09667,
+			label: 'June 18, 18',
+			changeOverTime: 0.00458,
+		},
+		{
+			date: '2018-06-15',
+			open: 47.5075,
+			high: 47.540001,
+			low: 47.064999,
+			close: 47.209999,
+			adjClose: 45.511429,
+			volume: 246876800,
+			unadjustedVolume: 246876800,
+			change: -0.2975,
+			changePercent: -0.626,
+			vwap: 47.27167,
+			label: 'June 15, 18',
+			changeOverTime: -0.00626,
+		},
+		{
+			date: '2018-06-14',
+			open: 47.887501,
+			high: 47.892502,
+			low: 47.555,
+			close: 47.700001,
+			adjClose: 45.983799,
+			volume: 86440400,
+			unadjustedVolume: 86440400,
+			change: -0.1875,
+			changePercent: -0.392,
+			vwap: 47.71583,
+			label: 'June 14, 18',
+			changeOverTime: -0.00392,
+		},
+		{
+			date: '2018-06-13',
+			open: 48.105,
+			high: 48.220001,
+			low: 47.610001,
+			close: 47.674999,
+			adjClose: 45.959702,
+			volume: 86553600,
+			unadjustedVolume: 86553600,
+			change: -0.43,
+			changePercent: -0.894,
+			vwap: 47.835,
+			label: 'June 13, 18',
+			changeOverTime: -0.00894,
+		},
+		{
+			date: '2018-06-12',
+			open: 47.8475,
+			high: 48.1525,
+			low: 47.787498,
+			close: 48.07,
+			adjClose: 46.3405,
+			volume: 67644400,
+			unadjustedVolume: 67644400,
+			change: 0.2225,
+			changePercent: 0.465,
+			vwap: 48.00333,
+			label: 'June 12, 18',
+			changeOverTime: 0.00465,
+		},
+		{
+			date: '2018-06-11',
+			open: 47.837502,
+			high: 47.9925,
+			low: 47.552502,
+			close: 47.807499,
+			adjClose: 46.087437,
+			volume: 73234000,
+			unadjustedVolume: 73234000,
+			change: -0.03,
+			changePercent: -0.063,
+			vwap: 47.78417,
+			label: 'June 11, 18',
+			changeOverTime: -0.00063,
+		},
+		{
+			date: '2018-06-08',
+			open: 47.7925,
+			high: 48.0,
+			low: 47.442501,
+			close: 47.924999,
+			adjClose: 46.200706,
+			volume: 106627200,
+			unadjustedVolume: 106627200,
+			change: 0.1325,
+			changePercent: 0.277,
+			vwap: 47.78917,
+			label: 'June 08, 18',
+			changeOverTime: 0.00277,
+		},
+		{
+			date: '2018-06-07',
+			open: 48.535,
+			high: 48.549999,
+			low: 48.084999,
+			close: 48.365002,
+			adjClose: 46.624882,
+			volume: 85388800,
+			unadjustedVolume: 85388800,
+			change: -0.17,
+			changePercent: -0.35,
+			vwap: 48.33333,
+			label: 'June 07, 18',
+			changeOverTime: -0.0035,
+		},
+		{
+			date: '2018-06-06',
+			open: 48.407501,
+			high: 48.52,
+			low: 47.98,
+			close: 48.494999,
+			adjClose: 46.750198,
+			volume: 83734400,
+			unadjustedVolume: 83734400,
+			change: 0.0875,
+			changePercent: 0.181,
+			vwap: 48.33167,
+			label: 'June 06, 18',
+			changeOverTime: 0.00181,
+		},
+		{
+			date: '2018-06-05',
+			open: 48.267502,
+			high: 48.485001,
+			low: 48.09,
+			close: 48.327499,
+			adjClose: 46.588726,
+			volume: 86264000,
+			unadjustedVolume: 86264000,
+			change: 0.06,
+			changePercent: 0.124,
+			vwap: 48.30083,
+			label: 'June 05, 18',
+			changeOverTime: 0.00124,
+		},
+		{
+			date: '2018-06-04',
+			open: 47.91,
+			high: 48.355,
+			low: 47.837502,
+			close: 47.9575,
+			adjClose: 46.23204,
+			volume: 105064800,
+			unadjustedVolume: 105064800,
+			change: 0.0475,
+			changePercent: 0.099,
+			vwap: 48.05,
+			label: 'June 04, 18',
+			changeOverTime: 0.00099,
+		},
+		{
+			date: '2018-06-01',
+			open: 46.997501,
+			high: 47.564999,
+			low: 46.9375,
+			close: 47.560001,
+			adjClose: 45.84885,
+			volume: 93770000,
+			unadjustedVolume: 93770000,
+			change: 0.5625,
+			changePercent: 1.197,
+			vwap: 47.35417,
+			label: 'June 01, 18',
+			changeOverTime: 0.01197,
+		},
+		{
+			date: '2018-05-31',
+			open: 46.805,
+			high: 47.057499,
+			low: 46.535,
+			close: 46.717499,
+			adjClose: 45.036652,
+			volume: 109931200,
+			unadjustedVolume: 109931200,
+			change: -0.0875,
+			changePercent: -0.187,
+			vwap: 46.77,
+			label: 'May 31, 18',
+			changeOverTime: -0.00187,
+		},
+		{
+			date: '2018-05-30',
+			open: 46.93,
+			high: 47.0,
+			low: 46.695,
+			close: 46.875,
+			adjClose: 45.188492,
+			volume: 74762000,
+			unadjustedVolume: 74762000,
+			change: -0.055,
+			changePercent: -0.117,
+			vwap: 46.85667,
+			label: 'May 30, 18',
+			changeOverTime: -0.00117,
+		},
+		{
+			date: '2018-05-29',
+			open: 46.900002,
+			high: 47.1875,
+			low: 46.717499,
+			close: 46.974998,
+			adjClose: 45.284893,
+			volume: 90056400,
+			unadjustedVolume: 90056400,
+			change: 0.075,
+			changePercent: 0.16,
+			vwap: 46.96,
+			label: 'May 29, 18',
+			changeOverTime: 0.0016,
+		},
+		{
+			date: '2018-05-25',
+			open: 47.057499,
+			high: 47.412498,
+			low: 46.912498,
+			close: 47.145,
+			adjClose: 45.448772,
+			volume: 69844000,
+			unadjustedVolume: 69844000,
+			change: 0.0875,
+			changePercent: 0.186,
+			vwap: 47.15667,
+			label: 'May 25, 18',
+			changeOverTime: 0.00186,
+		},
+		{
+			date: '2018-05-24',
+			open: 47.192501,
+			high: 47.209999,
+			low: 46.552502,
+			close: 47.037498,
+			adjClose: 45.345131,
+			volume: 92936000,
+			unadjustedVolume: 92936000,
+			change: -0.155,
+			changePercent: -0.328,
+			vwap: 46.93333,
+			label: 'May 24, 18',
+			changeOverTime: -0.00328,
+		},
+		{
+			date: '2018-05-23',
+			open: 46.587502,
+			high: 47.125,
+			low: 46.439999,
+			close: 47.09,
+			adjClose: 45.395748,
+			volume: 80233600,
+			unadjustedVolume: 80233600,
+			change: 0.5025,
+			changePercent: 1.079,
+			vwap: 46.885,
+			label: 'May 23, 18',
+			changeOverTime: 0.01079,
+		},
+		{
+			date: '2018-05-22',
+			open: 47.095001,
+			high: 47.220001,
+			low: 46.695,
+			close: 46.790001,
+			adjClose: 45.106541,
+			volume: 60962800,
+			unadjustedVolume: 60962800,
+			change: -0.305,
+			changePercent: -0.648,
+			vwap: 46.90167,
+			label: 'May 22, 18',
+			changeOverTime: -0.00648,
+		},
+		{
+			date: '2018-05-21',
+			open: 47.0,
+			high: 47.317501,
+			low: 46.727501,
+			close: 46.907501,
+			adjClose: 45.219818,
+			volume: 73603200,
+			unadjustedVolume: 73603200,
+			change: -0.0925,
+			changePercent: -0.197,
+			vwap: 46.98417,
+			label: 'May 21, 18',
+			changeOverTime: -0.00197,
+		},
+		{
+			date: '2018-05-18',
+			open: 46.797501,
+			high: 46.952499,
+			low: 46.532501,
+			close: 46.577499,
+			adjClose: 44.901684,
+			volume: 73190800,
+			unadjustedVolume: 73190800,
+			change: -0.22,
+			changePercent: -0.47,
+			vwap: 46.6875,
+			label: 'May 18, 18',
+			changeOverTime: -0.0047,
+		},
+		{
+			date: '2018-05-17',
+			open: 47.0,
+			high: 47.227501,
+			low: 46.59,
+			close: 46.747501,
+			adjClose: 45.065582,
+			volume: 69176000,
+			unadjustedVolume: 69176000,
+			change: -0.2525,
+			changePercent: -0.537,
+			vwap: 46.855,
+			label: 'May 17, 18',
+			changeOverTime: -0.00537,
+		},
+		{
+			date: '2018-05-16',
+			open: 46.517502,
+			high: 47.115002,
+			low: 46.5,
+			close: 47.044998,
+			adjClose: 45.352375,
+			volume: 76732400,
+			unadjustedVolume: 76732400,
+			change: 0.5275,
+			changePercent: 1.134,
+			vwap: 46.88667,
+			label: 'May 16, 18',
+			changeOverTime: 0.01134,
+		},
+		{
+			date: '2018-05-15',
+			open: 46.695,
+			high: 46.767502,
+			low: 46.275002,
+			close: 46.610001,
+			adjClose: 44.933022,
+			volume: 94780800,
+			unadjustedVolume: 94780800,
+			change: -0.085,
+			changePercent: -0.182,
+			vwap: 46.55084,
+			label: 'May 15, 18',
+			changeOverTime: -0.00182,
+		},
+		{
+			date: '2018-05-14',
+			open: 47.252499,
+			high: 47.3825,
+			low: 46.965,
+			close: 47.037498,
+			adjClose: 45.345131,
+			volume: 83115200,
+			unadjustedVolume: 83115200,
+			change: -0.215,
+			changePercent: -0.455,
+			vwap: 47.12833,
+			label: 'May 14, 18',
+			changeOverTime: -0.00455,
+		},
+		{
+			date: '2018-05-11',
+			open: 47.372501,
+			high: 47.514999,
+			low: 46.862499,
+			close: 47.147499,
+			adjClose: 45.451187,
+			volume: 104848800,
+			unadjustedVolume: 104848800,
+			change: -0.225,
+			changePercent: -0.475,
+			vwap: 47.175,
+			label: 'May 11, 18',
+			changeOverTime: -0.00475,
+		},
+		{
+			date: '2018-05-10',
+			open: 46.935001,
+			high: 47.592499,
+			low: 46.912498,
+			close: 47.509998,
+			adjClose: 45.624702,
+			volume: 111957200,
+			unadjustedVolume: 111957200,
+			change: 0.575,
+			changePercent: 1.225,
+			vwap: 47.33833,
+			label: 'May 10, 18',
+			changeOverTime: 0.01225,
+		},
+		{
+			date: '2018-05-09',
+			open: 46.637501,
+			high: 46.849998,
+			low: 46.305,
+			close: 46.84,
+			adjClose: 44.981293,
+			volume: 92844800,
+			unadjustedVolume: 92844800,
+			change: 0.2025,
+			changePercent: 0.434,
+			vwap: 46.665,
+			label: 'May 09, 18',
+			changeOverTime: 0.00434,
+		},
+		{
+			date: '2018-05-08',
+			open: 46.247501,
+			high: 46.555,
+			low: 45.9175,
+			close: 46.512501,
+			adjClose: 44.666786,
+			volume: 113611200,
+			unadjustedVolume: 113611200,
+			change: 0.265,
+			changePercent: 0.573,
+			vwap: 46.32833,
+			label: 'May 08, 18',
+			changeOverTime: 0.00573,
+		},
+		{
+			date: '2018-05-07',
+			open: 46.294998,
+			high: 46.9175,
+			low: 46.1875,
+			close: 46.290001,
+			adjClose: 44.453125,
+			volume: 169805600,
+			unadjustedVolume: 169805600,
+			change: -0.005,
+			changePercent: -0.011,
+			vwap: 46.465,
+			label: 'May 07, 18',
+			changeOverTime: -0.00011,
+		},
+		{
+			date: '2018-05-04',
+			open: 44.5625,
+			high: 46.0625,
+			low: 44.5425,
+			close: 45.9575,
+			adjClose: 44.133804,
+			volume: 224805200,
+			unadjustedVolume: 224805200,
+			change: 1.395,
+			changePercent: 3.13,
+			vwap: 45.52083,
+			label: 'May 04, 18',
+			changeOverTime: 0.0313,
+		},
+		{
+			date: '2018-05-03',
+			open: 43.970001,
+			high: 44.375,
+			low: 43.610001,
+			close: 44.2225,
+			adjClose: 42.467663,
+			volume: 136272800,
+			unadjustedVolume: 136272800,
+			change: 0.2525,
+			changePercent: 0.574,
+			vwap: 44.06917,
+			label: 'May 03, 18',
+			changeOverTime: 0.00574,
+		},
+		{
+			date: '2018-05-02',
+			open: 43.807499,
+			high: 44.4375,
+			low: 43.450001,
+			close: 44.142502,
+			adjClose: 42.390835,
+			volume: 266157600,
+			unadjustedVolume: 266157600,
+			change: 0.335,
+			changePercent: 0.765,
+			vwap: 44.01,
+			label: 'May 02, 18',
+			changeOverTime: 0.00765,
+		},
+		{
+			date: '2018-05-01',
+			open: 41.602501,
+			high: 42.299999,
+			low: 41.317501,
+			close: 42.275002,
+			adjClose: 40.597446,
+			volume: 214277600,
+			unadjustedVolume: 214277600,
+			change: 0.6725,
+			changePercent: 1.616,
+			vwap: 41.96417,
+			label: 'May 01, 18',
+			changeOverTime: 0.01616,
+		},
+		{
+			date: '2018-04-30',
+			open: 40.532501,
+			high: 41.814999,
+			low: 40.459999,
+			close: 41.314999,
+			adjClose: 39.675537,
+			volume: 169709600,
+			unadjustedVolume: 169709600,
+			change: 0.7825,
+			changePercent: 1.931,
+			vwap: 41.19667,
+			label: 'April 30, 18',
+			changeOverTime: 0.01931,
+		},
+		{
+			date: '2018-04-27',
+			open: 41.0,
+			high: 41.0825,
+			low: 40.157501,
+			close: 40.580002,
+			adjClose: 38.969704,
+			volume: 142623200,
+			unadjustedVolume: 142623200,
+			change: -0.42,
+			changePercent: -1.024,
+			vwap: 40.60667,
+			label: 'April 27, 18',
+			changeOverTime: -0.01024,
+		},
+		{
+			date: '2018-04-26',
+			open: 41.029999,
+			high: 41.432499,
+			low: 40.842499,
+			close: 41.055,
+			adjClose: 39.425854,
+			volume: 111852000,
+			unadjustedVolume: 111852000,
+			change: 0.025,
+			changePercent: 0.061,
+			vwap: 41.11,
+			label: 'April 26, 18',
+			changeOverTime: 0.00061,
+		},
+		{
+			date: '2018-04-25',
+			open: 40.654999,
+			high: 41.355,
+			low: 40.602501,
+			close: 40.912498,
+			adjClose: 39.289009,
+			volume: 113528400,
+			unadjustedVolume: 113528400,
+			change: 0.2575,
+			changePercent: 0.633,
+			vwap: 40.95667,
+			label: 'April 25, 18',
+			changeOverTime: 0.00633,
+		},
+		{
+			date: '2018-04-24',
+			open: 41.4175,
+			high: 41.5825,
+			low: 40.305,
+			close: 40.735001,
+			adjClose: 39.118553,
+			volume: 134768000,
+			unadjustedVolume: 134768000,
+			change: -0.6825,
+			changePercent: -1.648,
+			vwap: 40.87417,
+			label: 'April 24, 18',
+			changeOverTime: -0.01648,
+		},
+		{
+			date: '2018-04-23',
+			open: 41.7075,
+			high: 41.73,
+			low: 41.022499,
+			close: 41.310001,
+			adjClose: 39.670742,
+			volume: 146062000,
+			unadjustedVolume: 146062000,
+			change: -0.3975,
+			changePercent: -0.953,
+			vwap: 41.35417,
+			label: 'April 23, 18',
+			changeOverTime: -0.00953,
+		},
+		{
+			date: '2018-04-20',
+			open: 42.650002,
+			high: 42.805,
+			low: 41.357498,
+			close: 41.43,
+			adjClose: 39.78598,
+			volume: 261964400,
+			unadjustedVolume: 261964400,
+			change: -1.22,
+			changePercent: -2.86,
+			vwap: 41.86417,
+			label: 'April 20, 18',
+			changeOverTime: -0.0286,
+		},
+		{
+			date: '2018-04-19',
+			open: 43.439999,
+			high: 43.8475,
+			low: 43.165001,
+			close: 43.200001,
+			adjClose: 41.485737,
+			volume: 139235200,
+			unadjustedVolume: 139235200,
+			change: -0.24,
+			changePercent: -0.552,
+			vwap: 43.40417,
+			label: 'April 19, 18',
+			changeOverTime: -0.00552,
+		},
+		{
+			date: '2018-04-18',
+			open: 44.452499,
+			high: 44.705002,
+			low: 44.220001,
+			close: 44.459999,
+			adjClose: 42.695728,
+			volume: 83018000,
+			unadjustedVolume: 83018000,
+			change: 0.0075,
+			changePercent: 0.017,
+			vwap: 44.46167,
+			label: 'April 18, 18',
+			changeOverTime: 0.00017,
+		},
+		{
+			date: '2018-04-17',
+			open: 44.122501,
+			high: 44.735001,
+			low: 44.102501,
+			close: 44.560001,
+			adjClose: 42.791782,
+			volume: 106421600,
+			unadjustedVolume: 106421600,
+			change: 0.4375,
+			changePercent: 0.992,
+			vwap: 44.46583,
+			label: 'April 17, 18',
+			changeOverTime: 0.00992,
+		},
+		{
+			date: '2018-04-16',
+			open: 43.7575,
+			high: 44.047501,
+			low: 43.7075,
+			close: 43.955002,
+			adjClose: 42.210781,
+			volume: 86313600,
+			unadjustedVolume: 86313600,
+			change: 0.1975,
+			changePercent: 0.451,
+			vwap: 43.90333,
+			label: 'April 16, 18',
+			changeOverTime: 0.00451,
+		},
+		{
+			date: '2018-04-13',
+			open: 43.695,
+			high: 43.959999,
+			low: 43.462502,
+			close: 43.682499,
+			adjClose: 41.949093,
+			volume: 100497200,
+			unadjustedVolume: 100497200,
+			change: -0.0125,
+			changePercent: -0.029,
+			vwap: 43.70167,
+			label: 'April 13, 18',
+			changeOverTime: -0.00029,
+		},
+		{
+			date: '2018-04-12',
+			open: 43.352501,
+			high: 43.75,
+			low: 43.259998,
+			close: 43.535,
+			adjClose: 41.807442,
+			volume: 91557200,
+			unadjustedVolume: 91557200,
+			change: 0.1825,
+			changePercent: 0.421,
+			vwap: 43.515,
+			label: 'April 12, 18',
+			changeOverTime: 0.00421,
+		},
+		{
+			date: '2018-04-11',
+			open: 43.057499,
+			high: 43.48,
+			low: 42.924999,
+			close: 43.110001,
+			adjClose: 41.399307,
+			volume: 89726400,
+			unadjustedVolume: 89726400,
+			change: 0.0525,
+			changePercent: 0.122,
+			vwap: 43.17167,
+			label: 'April 11, 18',
+			changeOverTime: 0.00122,
+		},
+		{
+			date: '2018-04-10',
+			open: 43.25,
+			high: 43.5,
+			low: 42.8825,
+			close: 43.3125,
+			adjClose: 41.593769,
+			volume: 113634400,
+			unadjustedVolume: 113634400,
+			change: 0.0625,
+			changePercent: 0.145,
+			vwap: 43.23167,
+			label: 'April 10, 18',
+			changeOverTime: 0.00145,
+		},
+		{
+			date: '2018-04-09',
+			open: 42.470001,
+			high: 43.272499,
+			low: 42.462502,
+			close: 42.512501,
+			adjClose: 40.825516,
+			volume: 116070800,
+			unadjustedVolume: 116070800,
+			change: 0.0425,
+			changePercent: 0.1,
+			vwap: 42.74917,
+			label: 'April 09, 18',
+			changeOverTime: 0.001,
+		},
+		{
+			date: '2018-04-06',
+			open: 42.7425,
+			high: 43.119999,
+			low: 42.049999,
+			close: 42.095001,
+			adjClose: 40.424587,
+			volume: 140021200,
+			unadjustedVolume: 140021200,
+			change: -0.6475,
+			changePercent: -1.515,
+			vwap: 42.42167,
+			label: 'April 06, 18',
+			changeOverTime: -0.01515,
+		},
+		{
+			date: '2018-04-05',
+			open: 43.145,
+			high: 43.557499,
+			low: 43.02,
+			close: 43.200001,
+			adjClose: 41.485737,
+			volume: 107732800,
+			unadjustedVolume: 107732800,
+			change: 0.055,
+			changePercent: 0.127,
+			vwap: 43.25917,
+			label: 'April 05, 18',
+			changeOverTime: 0.00127,
+		},
+		{
+			date: '2018-04-04',
+			open: 41.220001,
+			high: 43.002499,
+			low: 41.192501,
+			close: 42.9025,
+			adjClose: 41.200039,
+			volume: 138422000,
+			unadjustedVolume: 138422000,
+			change: 1.6825,
+			changePercent: 4.082,
+			vwap: 42.36583,
+			label: 'April 04, 18',
+			changeOverTime: 0.04082,
+		},
+		{
+			date: '2018-04-03',
+			open: 41.91,
+			high: 42.1875,
+			low: 41.220001,
+			close: 42.0975,
+			adjClose: 40.426987,
+			volume: 121112000,
+			unadjustedVolume: 121112000,
+			change: 0.1875,
+			changePercent: 0.447,
+			vwap: 41.835,
+			label: 'April 03, 18',
+			changeOverTime: 0.00447,
+		},
+		{
+			date: '2018-04-02',
+			open: 41.66,
+			high: 42.235001,
+			low: 41.1175,
+			close: 41.669998,
+			adjClose: 40.016449,
+			volume: 150347200,
+			unadjustedVolume: 150347200,
+			change: 0.01,
+			changePercent: 0.024,
+			vwap: 41.67417,
+			label: 'April 02, 18',
+			changeOverTime: 0.00024,
+		},
+		{
+			date: '2018-03-29',
+			open: 41.952499,
+			high: 42.9375,
+			low: 41.724998,
+			close: 41.945,
+			adjClose: 40.280533,
+			volume: 153594000,
+			unadjustedVolume: 153594000,
+			change: -0.0075,
+			changePercent: -0.018,
+			vwap: 42.2025,
+			label: 'March 29, 18',
+			changeOverTime: -0.00018,
+		},
+		{
+			date: '2018-03-28',
+			open: 41.8125,
+			high: 42.505001,
+			low: 41.297501,
+			close: 41.619999,
+			adjClose: 39.968433,
+			volume: 166674000,
+			unadjustedVolume: 166674000,
+			change: -0.1925,
+			changePercent: -0.46,
+			vwap: 41.8075,
+			label: 'March 28, 18',
+			changeOverTime: -0.0046,
+		},
+		{
+			date: '2018-03-27',
+			open: 43.419998,
+			high: 43.787498,
+			low: 41.73,
+			close: 42.084999,
+			adjClose: 40.414982,
+			volume: 163690400,
+			unadjustedVolume: 163690400,
+			change: -1.335,
+			changePercent: -3.075,
+			vwap: 42.53417,
+			label: 'March 27, 18',
+			changeOverTime: -0.03075,
+		},
+		{
+			date: '2018-03-26',
+			open: 42.017502,
+			high: 43.275002,
+			low: 41.610001,
+			close: 43.192501,
+			adjClose: 41.478535,
+			volume: 150164800,
+			unadjustedVolume: 150164800,
+			change: 1.175,
+			changePercent: 2.796,
+			vwap: 42.6925,
+			label: 'March 26, 18',
+			changeOverTime: 0.02796,
+		},
+		{
+			date: '2018-03-23',
+			open: 42.0975,
+			high: 42.48,
+			low: 41.235001,
+			close: 41.235001,
+			adjClose: 39.598709,
+			volume: 164115200,
+			unadjustedVolume: 164115200,
+			change: -0.8625,
+			changePercent: -2.049,
+			vwap: 41.65,
+			label: 'March 23, 18',
+			changeOverTime: -0.02049,
+		},
+		{
+			date: '2018-03-22',
+			open: 42.5,
+			high: 43.169998,
+			low: 42.150002,
+			close: 42.212502,
+			adjClose: 40.537426,
+			volume: 165963200,
+			unadjustedVolume: 165963200,
+			change: -0.2875,
+			changePercent: -0.676,
+			vwap: 42.51083,
+			label: 'March 22, 18',
+			changeOverTime: -0.00676,
+		},
+		{
+			date: '2018-03-21',
+			open: 43.759998,
+			high: 43.772499,
+			low: 42.814999,
+			close: 42.817501,
+			adjClose: 41.11842,
+			volume: 148219600,
+			unadjustedVolume: 148219600,
+			change: -0.9425,
+			changePercent: -2.154,
+			vwap: 43.135,
+			label: 'March 21, 18',
+			changeOverTime: -0.02154,
+		},
+		{
+			date: '2018-03-20',
+			open: 43.810001,
+			high: 44.200001,
+			low: 43.735001,
+			close: 43.810001,
+			adjClose: 42.071533,
+			volume: 78597600,
+			unadjustedVolume: 78597600,
+			change: 0.0,
+			changePercent: 0.0,
+			vwap: 43.915,
+			label: 'March 20, 18',
+			changeOverTime: 0.0,
+		},
+		{
+			date: '2018-03-19',
+			open: 44.330002,
+			high: 44.3675,
+			low: 43.415001,
+			close: 43.825001,
+			adjClose: 42.085938,
+			volume: 133787200,
+			unadjustedVolume: 133787200,
+			change: -0.505,
+			changePercent: -1.139,
+			vwap: 43.86917,
+			label: 'March 19, 18',
+			changeOverTime: -0.01139,
+		},
+		{
+			date: '2018-03-16',
+			open: 44.662498,
+			high: 44.779999,
+			low: 44.404999,
+			close: 44.505001,
+			adjClose: 42.738956,
+			volume: 157618800,
+			unadjustedVolume: 157618800,
+			change: -0.1575,
+			changePercent: -0.353,
+			vwap: 44.56333,
+			label: 'March 16, 18',
+			changeOverTime: -0.00353,
+		},
+		{
+			date: '2018-03-15',
+			open: 44.625,
+			high: 45.060001,
+			low: 44.517502,
+			close: 44.662498,
+			adjClose: 42.890202,
+			volume: 90975200,
+			unadjustedVolume: 90975200,
+			change: 0.0375,
+			changePercent: 0.084,
+			vwap: 44.74667,
+			label: 'March 15, 18',
+			changeOverTime: 0.00084,
+		},
+		{
+			date: '2018-03-14',
+			open: 45.080002,
+			high: 45.130001,
+			low: 44.452499,
+			close: 44.610001,
+			adjClose: 42.839783,
+			volume: 117473600,
+			unadjustedVolume: 117473600,
+			change: -0.47,
+			changePercent: -1.043,
+			vwap: 44.73083,
+			label: 'March 14, 18',
+			changeOverTime: -0.01043,
+		},
+		{
+			date: '2018-03-13',
+			open: 45.647499,
+			high: 45.875,
+			low: 44.810001,
+			close: 44.9925,
+			adjClose: 43.207115,
+			volume: 126774000,
+			unadjustedVolume: 126774000,
+			change: -0.655,
+			changePercent: -1.435,
+			vwap: 45.22583,
+			label: 'March 13, 18',
+			changeOverTime: -0.01435,
+		},
+		{
+			date: '2018-03-12',
+			open: 45.072498,
+			high: 45.5975,
+			low: 45.052502,
+			close: 45.43,
+			adjClose: 43.627251,
+			volume: 128828400,
+			unadjustedVolume: 128828400,
+			change: 0.3575,
+			changePercent: 0.793,
+			vwap: 45.36,
+			label: 'March 12, 18',
+			changeOverTime: 0.00793,
+		},
+	],
 };
